@@ -1,101 +1,97 @@
 <?php
 return [
-    '1'   => 'Loading a frame(set)', //ok
-    '2'   => 'Viewing dashboard', //ok
-    '3'   => 'Viewing data for resource',  //ok
-    '4'   => 'Creating a resource', //ok == 27, 72, 85, 4
-    '5'   => 'Saving resource', //ok
-    '6'   => 'Deleting resource', //ok
-    '7'   => 'Waiting while Evolution CMS cleans up', //ok
-    '8'   => 'Logged out', //ok
-    '9'   => 'Viewing help', //ok
-    '10'  => 'Viewing/ composing messages', //ok
-    '11'  => 'Creating a user', //ok. == 12
-    '12'  => 'Editing user', //ok. == 11
-    '13'  => 'Viewing logging', //ok
-    '14'  => 'Editing a parser', //not found
-    '15'  => 'Saving a parser', //not found
-    '16'  => 'Editing template', //ok
-    '17'  => 'Editing settings', //ok
-    '18'  => 'Viewing Credits :)', //ok
-    '19'  => 'Creating a new template', //ok
-    '20'  => 'Saving template', //ok
-    '21'  => 'Deleting template', //ok
-    '22'  => 'Editing Snippet', //ok == 23
-    '23'  => 'Creating a new Snippet', //ok == 22
-    '24'  => 'Saving Snippet', //ok
-    '25'  => 'Deleting Snippet', //ok
-    '26'  => 'Refreshing site', //ok
-    '27'  => 'Editing resource', //ok = 27, 72, 85, 4
-    '28'  => 'Changing password', //ok
-    '29'  => 'Error', //ok
-    '30'  => 'Saving settings', //ok
-    '31'  => 'Using file manager', //ok
-    '32'  => 'Saving user', //ok
-    '33'  => 'Deleting user',//ok
-    '34'  => 'Saving new password', //ok
-    '35'  => 'Editing role',//ok == 38
-    '36'  => 'Saving role', //ok
-    '37'  => 'Deleting role', //ok
-    '38'  => 'Creating new role', //ok == 35
-    '40'  => 'Editing Access Permissions', //ok
-    '41'  => 'Editing Access Permissions', //ok access groups processor
-    '42'  => 'Editing Access Permissions', //not found
-    '43'  => 'Editing Access Permissions', //not found
-    '44'  => 'Editing Access Permissions', //not found
-    '45'  => 'Idle', //not found
-    '46'  => 'Editing Access Permissions', //not found
-    '47'  => 'Editing Access Permissions', //not found
-    '48'  => 'Editing Access Permissions', //not found
-    '49'  => 'Editing Access Permissions', //not found
-    '50'  => 'Editing Access Permissions', //not found
-    '51'  => 'Moving resource', //ok
-    '52'  => 'Moved resource', //ok
-    '53'  => 'Viewing system info', //ok
-    '54'  => 'Optimizing a table', //ok
-    '55'  => 'Empty logs', //55
-    '56'  => 'Refresh resource tree', //ok
-    '57'  => 'Refresh menu', //not found
-    '58'  => 'Logged in', //not found
-    '59'  => 'About Evolution CMS', //ok
-    '60'  => 'Emptying Recycle Bin', //not found
-    '61'  => 'Publishing a resource', //ok
-    '62'  => 'Un-publishing a resource', //ok
-    '63'  => 'Un-deleting a resource', //ok
-    '64'  => 'Removing deleted content', //ok
-    '65'  => 'Deleting a message', //ok
-    '66'  => 'Sending a message', //ok
-    '67'  => 'Removing locks', //ok
-    '68'  => 'Viewing site logging', //not found
-    '69'  => 'Viewing online visitors',//not found
-    '70'  => 'Viewing site schedule', //ok
-    '71'  => 'Searching', //ok
-    '72'  => 'Adding a weblink', //ok  27, 72, 85, 4
-    '73'  => 'Editing a weblink', //not found
-    //case "74" : return "Changing personal preferences"; break;
-    '75'  => 'User/ role management', //ok
-    '76'  => 'Element management', //ok
-    '77'  => 'Creating a new Chunk (HTML Snippet)', //ok 77, 78
-    '78'  => 'Editing Chunk (HTML Snippet)', //ok 77, 78
-    '79'  => 'Saving Chunk (HTML Snippet)', //ok
-    '80'  => 'Deleting Chunk (HTML Snippet)', //ok
-    '83'  => 'Exporting a resource to HTML', //ok
-    '84'  => 'Load Element Selector', //ok
-    '85'  => 'Create Folder', //ok 27, 72, 85, 4
-    '86'  => 'Role management', //ok
-    '87'  => 'Create new web user', //ok 87, 88
-    '88'  => 'Editing web user',//ok 87, 88
-    '89'  => 'Saving web user', //ok
-    '90'  => 'Deleting web user', //ok
-    '91'  => 'Editing Web Access Permissions', //ok
-    '92'  => 'Editing Access Permissions',  // ok access groups processor
-    '93'  => 'Backup Manager', //ok
-    '94'  => 'Duplicate resource', //ok
-    '95'  => 'Importing resources from HTML', //ok
-    '96'  => 'Duplicate Template', //ok
-    '97'  => 'Duplicate Chunk (HTML Snippet)',//ok
-    '98'  => 'Duplicate Snippet', //ok
-    '99'  => 'Manage Web Users', //ok
+    '1' => 'Loading a frame(set)', //ok
+    '2' => 'Viewing dashboard', //ok
+    '3' => 'Viewing data for resource', //ok
+    '4' => 'Creating a resource', //ok == 27, 72, 85, 4
+    '5' => 'Saving resource', //ok
+    '6' => 'Deleting resource', //ok
+    '7' => 'Waiting while Evolution CMS cleans up', //ok
+    '8' => 'Logged out', //ok
+    '9' => 'Viewing help', //ok
+    '10' => 'Viewing/ composing messages', //ok
+    '13' => 'Viewing logging', //ok
+    '14' => 'Editing a parser', //not found
+    '15' => 'Saving a parser', //not found
+    '16' => 'Editing template', //ok
+    '17' => 'Editing settings', //ok
+    '18' => 'Viewing Credits :)', //ok
+    '19' => 'Creating a new template', //ok
+    '20' => 'Saving template', //ok
+    '21' => 'Deleting template', //ok
+    '22' => 'Editing Snippet', //ok == 23
+    '23' => 'Creating a new Snippet', //ok == 22
+    '24' => 'Saving Snippet', //ok
+    '25' => 'Deleting Snippet', //ok
+    '26' => 'Refreshing site', //ok
+    '27' => 'Editing resource', //ok = 27, 72, 85, 4
+    '28' => 'Changing password', //ok
+    '29' => 'Error', //ok
+    '30' => 'Saving settings', //ok
+    '31' => 'Using file manager', //ok
+    '32' => 'Saving user', //ok
+    '33' => 'Deleting user', //ok
+    '34' => 'Saving new password', //ok
+    '35' => 'Editing role', //ok == 38
+    '36' => 'Saving role', //ok
+    '37' => 'Deleting role', //ok
+    '38' => 'Creating new role', //ok == 35
+    '40' => 'Editing Access Permissions', //ok
+    '41' => 'Editing Access Permissions', //ok access groups processor
+    '42' => 'Editing Access Permissions', //not found
+    '43' => 'Editing Access Permissions', //not found
+    '44' => 'Editing Access Permissions', //not found
+    '45' => 'Idle', //not found
+    '46' => 'Editing Access Permissions', //not found
+    '47' => 'Editing Access Permissions', //not found
+    '48' => 'Editing Access Permissions', //not found
+    '49' => 'Editing Access Permissions', //not found
+    '50' => 'Editing Access Permissions', //not found
+    '51' => 'Moving resource', //ok
+    '52' => 'Moved resource', //ok
+    '53' => 'Viewing system info', //ok
+    '54' => 'Optimizing a table', //ok
+    '55' => 'Empty logs', //55
+    '56' => 'Refresh resource tree', //ok
+    '57' => 'Refresh menu', //not found
+    '58' => 'Logged in', //not found
+    '59' => 'About Evolution CMS', //ok
+    '60' => 'Emptying Recycle Bin', //not found
+    '61' => 'Publishing a resource', //ok
+    '62' => 'Un-publishing a resource', //ok
+    '63' => 'Un-deleting a resource', //ok
+    '64' => 'Removing deleted content', //ok
+    '65' => 'Deleting a message', //ok
+    '66' => 'Sending a message', //ok
+    '67' => 'Removing locks', //ok
+    '68' => 'Viewing site logging', //not found
+    '69' => 'Viewing online visitors', //not found
+    '70' => 'Viewing site schedule', //ok
+    '71' => 'Searching', //ok
+    '72' => 'Adding a weblink', //ok  27, 72, 85, 4
+    '73' => 'Editing a weblink', //not found
+    '76' => 'Element management', //ok
+    '77' => 'Creating a new Chunk (HTML Snippet)', //ok 77, 78
+    '78' => 'Editing Chunk (HTML Snippet)', //ok 77, 78
+    '79' => 'Saving Chunk (HTML Snippet)', //ok
+    '80' => 'Deleting Chunk (HTML Snippet)', //ok
+    '83' => 'Exporting a resource to HTML', //ok
+    '84' => 'Load Element Selector', //ok
+    '85' => 'Create Folder', //ok 27, 72, 85, 4
+    '86' => 'Role management', //ok
+    '87' => 'Create new user', //ok 87, 88
+    '88' => 'Editing user', //ok 87, 88
+    '89' => 'Saving user', //ok
+    '90' => 'Deleting user', //ok
+    '91' => 'Editing Web Access Permissions', //ok
+    '92' => 'Editing Access Permissions', // ok access groups processor
+    '93' => 'Backup Manager', //ok
+    '94' => 'Duplicate resource', //ok
+    '95' => 'Importing resources from HTML', //ok
+    '96' => 'Duplicate Template', //ok
+    '97' => 'Duplicate Chunk (HTML Snippet)', //ok
+    '98' => 'Duplicate Snippet', //ok
+    '99' => 'Manage Users', //ok
     '100' => 'Previewing resource', //ok change the plugin priority
     '101' => 'Create new plugin', //ok 101, 102
     '102' => 'Edit plugin', //ok 101, 102
@@ -109,19 +105,19 @@ return [
     '110' => 'Delete module', //ok
     '111' => 'Duplicate module', //ok
     '112' => 'Execute module', //ok
-    '113' => 'Manage module dependencies',//ok
+    '113' => 'Manage module dependencies', //ok
     '114' => 'View event log', //ok
     '115' => 'View event log details', //ok
     '116' => 'Delete event log', //ok
-    '117' => 'Editing tv rank',  //ok
+    '117' => 'Editing tv rank', //ok
     '118' => 'Call settings ajax include', //ok
-    '119' => 'Login Fail (Temporary Block)',//ok  get the purge processor
+    '119' => 'Login Fail (Temporary Block)', //ok  get the purge processor
 
     '300' => 'Create Template Variable', //ok 300,301
-    '301' => 'Edit Template Variable',//ok 300,301
-    '302' => 'Save Template Variable',//ok
-    '303' => 'Delete Template Variable',//ok
-    '304' => 'Duplicate Template Variable',//ok
+    '301' => 'Edit Template Variable', //ok 300,301
+    '302' => 'Save Template Variable', //ok
+    '303' => 'Delete Template Variable', //ok
+    '304' => 'Duplicate Template Variable', //ok
 
     '200' => 'Viewing phpInfo()', //ok
     '501' => 'Delete category', //ok
@@ -132,3 +128,5 @@ return [
 //120  get the edit category page
 //121  for ajax-requests
 //305  get the tv-rank action
+
+//case "74" : return "Changing personal preferences"; break;
