@@ -7,7 +7,6 @@ return [
     'site_unavailable_page' => '',
     'top_howmany' => 10,
     'custom_contenttype' => 'application/rss+xml,application/pdf,application/vnd.ms-word,application/vnd.ms-excel,text/html,text/css,text/xml,text/javascript,text/plain,application/json',
-    'docid_incrmnt_method' => 0,
     'valid_hostnames' => '',
     'enable_filter' => 0,
     'enable_at_syntax' => 0,
@@ -54,8 +53,8 @@ return [
     'use_editor' => 1,
     'editor_css_path' => '',
     'filemanager_path' => '[(base_path)]',
-    'upload_files' => 'bmp,ico,gif,jpeg,jpg,png,psd,tif,tiff,fla,flv,swf,aac,au,avi,css,cache,doc,docx,gz,gzip,htaccess,htm,html,js,mp3,mp4,mpeg,mpg,ods,odp,odt,pdf,ppt,pptx,rar,tar,tgz,txt,wav,wmv,xls,xlsx,xml,z,zip,JPG,JPEG,PNG,GIF,svg,tpl',
-    'upload_images' => 'bmp,ico,gif,jpeg,jpg,png,psd,tif,tiff,svg',
+    'upload_files' => 'bmp,ico,gif,jpeg,jpg,png,psd,tif,tiff,fla,flv,swf,aac,au,avi,css,cache,doc,docx,gz,gzip,htaccess,htm,html,js,mp3,mp4,mpeg,mpg,ods,odp,odt,pdf,ppt,pptx,rar,tar,tgz,txt,wav,wmv,xls,xlsx,xml,z,zip,JPG,JPEG,PNG,GIF,svg,webp,tpl',
+    'upload_images' => 'bmp,ico,gif,jpeg,jpg,png,psd,tif,tiff,svg,webp',
     'upload_media' => 'au,avi,mp3,mp4,mpeg,mpg,wav,wmv',
     'upload_maxsize' => '5000000',
     'new_file_permissions' => '0644',
