@@ -1,7 +1,7 @@
 <?php
 return [
-    'version' => '3.1.15', // Current version number
-    'release_date' => 'February 2, 2023', // Date of release
+    'version' => '3.1.16', // Current version number
+    'release_date' => 'April 16, 2023', // Date of release
     'branch' => 'Evolution CE', // Codebase name
-    'full_appname' => 'Evolution CE 3.1.15 (February 2, 2023)'
+    'full_appname' => 'Evolution CE 3.1.16 (April 16, 2023)'
 ];
