@@ -37,6 +37,7 @@ $_lang["access_permissions_off"] = '<span class="warning">Дазволы дос�
 $_lang["access_permissions_resource_groups"] = 'Групы рэсурсаў';
 $_lang["access_permissions_resources_in_group"] = '<b>Рэсурсы ў групе:</b>';
 $_lang["access_permissions_resources_tab"] = 'Паглядзіце, якія групы рэсурсаў былі створаны. Таксама стварайце новыя групы, пераназывайце групы, выдаляйце групы і глядзіце, якія рэсурсы знаходзяцца ў розных групах (навядзіце курсор на ідэнтыфікатар рэсурсу, каб убачыць яго назву). Каб дадаць рэсурс у групу або выдаліць рэсурс з групы, адрэдагуйце рэсурс непасрэдна.';
+$_lang["access_permissions_user_toggle"] = 'Toggle access permissions';
 $_lang["access_permissions_user_groups"] = 'Групы карыстальнікаў';
 $_lang["access_permissions_user_message"] = 'Выберыце, да якіх груп карыстальнікаў належыць гэты карыстальнік:';
 $_lang["access_permissions_users_in_group"] = 'Карыстальнікі ў групе:';
@@ -84,10 +85,11 @@ $_lang["captcha_title"] = 'Выкарыстоўвайце коды CAPTCHA';
 $_lang["captcha_words_default"] = 'EVO,Access,Better,BitCode,Chunk,Cache,Desc,Design,Excell,Enjoy,URLs,TechView,Gerald,Griff,Humphrey,Holiday,Intel,Integration,Joystick,Join(),Oscope,Genetic,Light,Likeness, Marit,Maaike,Niche,Netherlands,Ordinance,Oscillo,Parser,Phusion,Query,Question,Regalia,Righteous,Snippet,Sentinel,Template,Thespian,Unity,Enterprise,Verily,Tattoo,Veri,Website,WideWeb,Yap,Yellow, Зебра, зігота';
 $_lang["captcha_words_message"] = 'Увядзіце спіс слоў CAPTCHA для выкарыстання, калі CAPTCHA ўключана. Раздзяліце словы коскамі. Гэта поле ўводу абмежавана 255 сімваламі.';
 $_lang["captcha_words_title"] = 'CAPTCHA Словы';
+$_lang["category"] = 'Category';
+$_lang["categories"] = 'Categories';
 $_lang["category_heading"] = 'Катэгорыя';
 $_lang["category_manager"] = 'Менеджэр катэгорый';
 $_lang["category_management"] = 'Кіраванне катэгорыямі';
-$_lang["manage_categories"] = 'Кіраванне катэгорыямі';
 $_lang["category_msg"] = 'Прагляд і рэдагаванне ўсіх элементаў, згрупаваных па катэгорыях.';
 $_lang["cfg_base_path"] = 'MODX_BASE_PATH';
 $_lang["cfg_base_url"] = 'MODX_BASE_URL';
@@ -322,6 +324,7 @@ $_lang["file_unzip"] = 'Распакаванне прайшло паспяхов
 $_lang["file_unzip_fail"] = 'Не атрымалася распакаваць!';
 $_lang["filemanager_path_message"] = 'IIS часта не запаўняе належным чынам параметр document_root, які выкарыстоўваецца файлавым мэнэджарам для вызначэння таго, што вы бачыце. Калі ў вас узніклі праблемы з выкарыстаннем дыспетчара файлаў, пераканайцеся, што гэты шлях паказвае на корань вашай устаноўкі Evolution CMS.';
 $_lang["filemanager_path_title"] = 'Шлях дыспетчара файлаў';
+$_lang["files"] = 'Files';
 $_lang["files_access_denied"] = 'Доступ забаронены!';
 $_lang["files_data"] = 'даныя';
 $_lang["files_dir_listing"] = 'Спіс каталога для:';
@@ -474,16 +477,17 @@ $_lang["login_message"] = 'Калі ласка, увядзіце ўліковы�
 $_lang["logo_slogan"] = 'Evolution CMS - \nСтварайце і рабіце больш з меншымі выдаткамі';
 $_lang["logout"] = 'Выйсці з сістэмы';
 $_lang["long_title"] = 'Доўгі загаловак';
+$_lang["manage_categories"] = 'Кіраванне катэгорыямі';
 $_lang["manage_depends"] = 'Кіраванне залежнасцямі';
 $_lang["manage_files"] = 'Кіраванне файламі';
-$_lang["manage_htmlsnippets"] = 'Кавалачкі';
+$_lang["manage_htmlsnippets"] = 'Manage Chunks';
 $_lang["manage_metatags"] = 'Кіраванне мета-тэгамі і ключавымі словамі';
 $_lang["manage_modules"] = 'Кіраванне модулямі';
-$_lang["manage_plugins"] = 'Убудовы';
-$_lang["manage_snippets"] = 'Фрагменты';
-$_lang["manage_templates"] = 'Шаблоны';
-$_lang["manage_documents"] = 'Дакументы';
-$_lang["manage_permission"] = 'Дазволы';
+$_lang["manage_plugins"] = 'Manage Plugins';
+$_lang["manage_snippets"] = 'Manage Snippets';
+$_lang["manage_templates"] = 'Manage Templates';
+$_lang["manage_documents"] = 'Manage Documents';
+$_lang["manage_permission"] = 'Manage Permissions';
 $_lang["manager"] = 'Менеджэр';
 $_lang["manager_lockout_message"] = 'Зараз вы ўвайшлі ў Менеджэр кантэнту. Калі вы жадаеце закрыць сеанс ўваходу, націсніце кнопку «Выхад». <br>Каб перайсці на стартавую або галоўную старонку, націсніце кнопку «Галоўная».';
 $_lang["manager_permissions"] = 'Дазволы менеджэра';
@@ -1027,6 +1031,7 @@ $_lang["template_assignedtv_tab"] = 'Прызначаныя зменныя ша�
 $_lang["template_code"] = 'Код шаблону (html)';
 $_lang["template_desc"] = 'Апісанне';
 $_lang["template_edit_tab"] = 'Рэдагаваць шаблон';
+$_lang["template_inuse"] = 'This template is in use. Please set the documents using the template to another template. Documents using this template:';
 $_lang["template_management_msg"] = 'Выберыце, які шаблон вы хочаце змяніць.';
 $_lang["template_msg"] = 'Стварэнне і рэдагаванне шаблонаў. Змененыя або новыя шаблоны не будуць бачныя на кэшаваных старонках вашага сайта, пакуль кэш не будзе ачышчаны, аднак вы можаце выкарыстоўваць функцыю папярэдняга прагляду на старонцы, каб убачыць шаблон у дзеянні.';
 $_lang["template_name"] = 'Імя шаблона';
@@ -1279,10 +1284,6 @@ $_lang["setting_resource_tree_node_name_desc_add"] = 'Заўвага: пачын
 $_lang["resource_opt_alvisibled"] = 'Выкарыстоўваць бягучы псеўданім у шляху да псеўданіма';
 $_lang["resource_opt_alvisibled_help"] = 'Псеўданім гэтага рэсурсу ўстаўлены ў дружалюбны шлях да псеўданіма URL';
 $_lang['resource_opt_is_published'] = 'Апублікавана';
-$_lang["docid_incrmnt_method_title"] = 'Метад інкрэментавання ідэнтыфікатара рэсурсу';
-$_lang["docid_incrmnt_method_0"] = 'Аўтаматычны прырост БД';
-$_lang["docid_incrmnt_method_1"] = 'Мінімальны прапушчаны ID';
-$_lang["docid_incrmnt_method_2"] = 'Максімальны ID + 1';
 
 $_lang["enable_cache_title"] = 'Кэшаванне дакументаў';
 $_lang["disable_chunk_cache_title"] = 'Адключыць кэшаванне сніпетаў';
