@@ -37,6 +37,7 @@ $_lang["access_permissions_off"] = '<span class="warning">Правата за Д
 $_lang["access_permissions_resource_groups"] = 'Групи Документи';
 $_lang["access_permissions_resources_in_group"] = '<b>Документи в група:</b> ';
 $_lang["access_permissions_resources_tab"] = 'Тук можете да видите коя Група Документи е настроена. Също така можете да създадете Нова Група, преименувате Групи, изтриете Групи и да видите кои Документи са в различните Групи (сложете мишката върху id на Документа, за да видите неговото име). За да добавите или махнете Документ от групата, редактирайте директно.';
+$_lang["access_permissions_user_toggle"] = 'Toggle access permissions';
 $_lang["access_permissions_user_groups"] = 'Потребителски групи';
 $_lang["access_permissions_user_message"] = 'Тук можете да изберете към кои Потребителски Групи принадлежи този Потребител:';
 $_lang["access_permissions_users_in_group"] = '<b>Потребители в група:</b> ';
@@ -84,10 +85,11 @@ $_lang["captcha_title"] = 'Използване на CAPTCHA код:';
 $_lang["captcha_words_default"] = 'MODX,Access,Better,BitCode,Chunk,Cache,Desc,Design,Excell,Enjoy,URLs,TechView,Gerald,Griff,Humphrey,Holiday,Intel,Integration,Joystick,Join(),Oscope,Genetic,Light,Likeness,Marit,Maaike,Niche,Netherlands,Ordinance,Oscillo,Parser,Phusion,Query,Question,Regalia,Righteous,Snippet,Sentinel,Template,Thespian,Unity,Enterprise,Verily,Tattoo,Veri,Website,WideWeb,Yap,Yellow,Zebra,Zygote';
 $_lang["captcha_words_message"] = 'Тук можете да въведете списък от CAPTCHA думи, които ще се използват, ако опцията за използване на CAPTCHA кода е включена. Разделете думите със запетаи. Полето е с лимит до 255 символа.';
 $_lang["captcha_words_title"] = 'CAPTCHA Думи';
+$_lang["category"] = 'Category';
+$_lang["categories"] = 'Categories';
 $_lang["category_heading"] = 'Категория';
 $_lang["category_manager"] = 'Category Manager';
 $_lang["category_management"] = 'Category management';
-$_lang["manage_categories"] = 'Manage Categories';
 $_lang["category_msg"] = 'Тук можете да разгледате и редактирате всички ресурси групирани по категория.';
 $_lang["cfg_base_path"] = 'MODX_BASE_PATH';
 $_lang["cfg_base_url"] = 'MODX_BASE_URL';
@@ -139,7 +141,7 @@ $_lang["configcheck_notok"] = 'Конфигурацията съдържа гр�
 $_lang["configcheck_ok"] = 'Проверката мина успешно.';
 $_lang["configcheck_php_gdzip"] = 'GD and/or Zip PHP extensions not found';
 $_lang["configcheck_php_gdzip_msg"] = 'Evolution CMS needs the GD and Zip extension enabled for PHP. While Evolution CMS will work without them, you will not be able to take full advantage of the built-in File Manager, Image Editor or Captcha for logins.';
-$_lang["configcheck_rb_base_dir"] = 'The currently set <a href="index.php?a=17&tab=5">File Browser path</a> seems incorrect.';
+$_lang["configcheck_rb_base_dir"] = 'The currently set <a href="index.php?a=17&tab=5">File base path</a> seems incorrect.';
 $_lang["configcheck_rb_base_dir_msg"] = 'This can happen for example by moving your installation to a different directory or server. Please check and update your Evolution CMS configuration.';
 $_lang["configcheck_register_globals"] = 'register_globals е установено на ON във Вашия php.ini конфигурационен файл';
 $_lang["configcheck_register_globals_msg"] = 'Тази конфигурация прави сайта ви податлив към Cross Site Scripting (XSS) атаки. Трябва да се обърнете към Вашия хостинг какво да направите, за да забраните тази настройка.';
@@ -322,6 +324,7 @@ $_lang["file_unzip"] = 'Разархивирането е успешно!';
 $_lang["file_unzip_fail"] = 'Грешка при разархивирането!';
 $_lang["filemanager_path_message"] = 'IIS много често не интерпретира правилно настройките на document_root , който се използва от файловия мениджър за определяне какво да виждате. В случай, че имате проблеми при използването на файловия мениджър, уверете се, че пътя сочи към root на EVO инсталацията.';
 $_lang["filemanager_path_title"] = 'Път на Файловия Мениджър:';
+$_lang["files"] = 'Files';
 $_lang["files_access_denied"] = 'Достъпът забранен!';
 $_lang["files_data"] = 'Данни';
 $_lang["files_dir_listing"] = 'Списък на файловете в директорията за:';
@@ -476,16 +479,17 @@ $_lang["login_message"] = 'Моля, въведете Потребителско
 $_lang["logo_slogan"] = 'EVO Content Manager - \nCreate and do more with less';
 $_lang["logout"] = 'Изход';
 $_lang["long_title"] = 'Дълго заглавие';
+$_lang["manage_categories"] = 'Manage Categories';
 $_lang["manage_depends"] = 'Управление на взаимовръзките';
 $_lang["manage_files"] = 'Управление на файловете';
-$_lang["manage_htmlsnippets"] = 'Чънкове';
+$_lang["manage_htmlsnippets"] = 'Manage Chunks';
 $_lang["manage_metatags"] = 'Управление на META тагове и Ключови думи';
 $_lang["manage_modules"] = 'Управление на модули';
-$_lang["manage_plugins"] = 'Плъгини';
-$_lang["manage_snippets"] = 'Снипети';
-$_lang["manage_templates"] = 'Шаблони';
-$_lang["manage_documents"] = 'Documents';
-$_lang["manage_permission"] = 'Permissions';
+$_lang["manage_plugins"] = 'Manage Plugins';
+$_lang["manage_snippets"] = 'Manage Snippets';
+$_lang["manage_templates"] = 'Manage Templates';
+$_lang["manage_documents"] = 'Manage Documents';
+$_lang["manage_permission"] = 'Manage Permissions';
 $_lang["manager"] = 'Мениджър';
 $_lang["manager_lockout_message"] = 'В момента сте влезли в Мениджъра. Ако желаете да прекратите сесията си, моля щракнете върху бутона "Изход" . <p />За да отидете във Вашата начална страница щракнете върху бутона "Начало" .';
 $_lang["manager_permissions"] = 'Права за достъп до Мениджъра';
@@ -1028,6 +1032,7 @@ $_lang["template_assignedtv_tab"] = 'Присвоени Шаблонни Про�
 $_lang["template_code"] = 'Код на шаблона (html)';
 $_lang["template_desc"] = 'Описание';
 $_lang["template_edit_tab"] = 'Редактиране на Шаблон';
+$_lang["template_inuse"] = 'This template is in use. Please set the documents using the template to another template. Documents using this template:';
 $_lang["template_management_msg"] = 'Тук можете да изберете кой шаблон желаете да редактирате.';
 $_lang["template_msg"] = 'Създаване и редактиране на шаблони. Променените или новите шаблони няма да бъдат видими на кешираните страници на сайта докато кеша не се изчисти. Обаче можете да използвате функцията за предварителен преглед на страницата, за да видите шаблона в действие.';
 $_lang["template_name"] = 'Име на Шаблона';
@@ -1280,10 +1285,6 @@ $_lang["setting_resource_tree_node_name_desc_add"] = 'Note: Since Evolution CMS 
 $_lang["resource_opt_alvisibled"] = 'Use current alias in alias path';
 $_lang["resource_opt_alvisibled_help"] = 'The alias of this Resource is inserted in Friendly URL alias path';
 $_lang['resource_opt_is_published'] = 'Published';
-$_lang["docid_incrmnt_method_title"] = 'Increment Resource ID method';
-$_lang["docid_incrmnt_method_0"] = 'DB auto increment';
-$_lang["docid_incrmnt_method_1"] = 'Minimum missed ID';
-$_lang["docid_incrmnt_method_2"] = 'Maximal ID + 1';
 
 $_lang["enable_cache_title"] = 'Document caching';
 $_lang["disable_chunk_cache_title"] = 'Disable chunk caching';
