@@ -37,6 +37,7 @@ $_lang["access_permissions_off"] = '<span class="warning">Přístupová práva n
 $_lang["access_permissions_resource_groups"] = 'Skupiny dokumentů';
 $_lang["access_permissions_resources_in_group"] = '<b>Dokumenty ve skupině:</b> ';
 $_lang["access_permissions_resources_tab"] = 'Tady můžete vidět, které skupiny dokumentů jsou nastavené. Můžete vytvářet nové skupiny, přejmenovávat je, mazat je a procházet, které dokumenty se nacházejí, v kterých skupinách (po přesunu kurzoru myši na ID dokumentu se zobrazí jeho název). Přidat dokument do skupiny nebo ho odstranit se da jen prímou změnou v samotném dokumentu.';
+$_lang["access_permissions_user_toggle"] = 'Toggle access permissions';
 $_lang["access_permissions_user_groups"] = 'Skupiny návštěvníků';
 $_lang["access_permissions_user_message"] = 'Tady můžete zvolit, který uživatel patří, do které skupiny:';
 $_lang["access_permissions_users_in_group"] = '<b>Uživatelé ve skupině:</b> ';
@@ -84,10 +85,11 @@ $_lang["captcha_title"] = 'Použít generovaný obrázkový<br /> (CAPTCHA) kód
 $_lang["captcha_words_default"] = 'EVO,Access,Better,BitCode,Chunk,Cache,Desc,Design,Excell,Enjoy,URLs,TechView,Gerald,Griff,Humphrey,Holiday,Intel,Integration,Joystick,Join(),Oscope,Genetic,Light,Likeness,Marit,Maaike,Niche,Netherlands,Ordinance,Oscillo,Parser,Phusion,Query,Question,Regalia,Righteous,Snippet,Sentinel,Template,Thespian,Unity,Enterprise,Verily,Tattoo,Veri,Website,WideWeb,Yap,Yellow,Zebra,Zygote';
 $_lang["captcha_words_message"] = 'Zde můžete zadat seznam CAPTCHA slov, které se používají při povolené CAPTCHA. Oddělte slova čárkami. Toto pole je limitované max. počtem 255 znaků.';
 $_lang["captcha_words_title"] = 'CAPTCHA slova';
+$_lang["category"] = 'Category';
+$_lang["categories"] = 'Categories';
 $_lang["category_heading"] = 'Kategorie';
 $_lang["category_manager"] = 'Category Manager';
 $_lang["category_management"] = 'Category management';
-$_lang["manage_categories"] = 'Manage Categories';
 $_lang["category_msg"] = 'Tady můžete zobrazit a upravit všechny zdroje seskupené do kategorií.';
 $_lang["cfg_base_path"] = 'MODX_BASE_PATH';
 $_lang["cfg_base_url"] = 'MODX_BASE_URL';
@@ -139,7 +141,7 @@ $_lang["configcheck_notok"] = 'Byly nalezeny následující problémy: ';
 $_lang["configcheck_ok"] = 'Kontola proběhla úspěšně - žádná varovavání nejsou třeba.';
 $_lang["configcheck_php_gdzip"] = 'GD a/nebo Zip PHP extenze nenalezena';
 $_lang["configcheck_php_gdzip_msg"] = 'Evolution CMS potřebuje aktivní extenzi GD a Zip v PHP. Dokud bude Evolution CMS bežet bez nich, nebude možno používat všech výhod vestavěného správce souborů, editoru obrázků nebo captcha.';
-$_lang["configcheck_rb_base_dir"] = 'The currently set <a href="index.php?a=17&tab=5">File Browser path</a> seems incorrect.';
+$_lang["configcheck_rb_base_dir"] = 'The currently set <a href="index.php?a=17&tab=5">File base path</a> seems incorrect.';
 $_lang["configcheck_rb_base_dir_msg"] = 'This can happen for example by moving your installation to a different directory or server. Please check and update your Evolution CMS configuration.';
 $_lang["configcheck_register_globals"] = 'register_globals je nastaven na zapnuto (ON) v konfiguračním souboru php.ini';
 $_lang["configcheck_register_globals_msg"] = 'Tato konfigurace dělá Vaše stránky více citlivé na útoky typu Cross Site Scripting (XSS). ';
@@ -322,6 +324,7 @@ $_lang["file_unzip"] = 'Rozbalení bylo úspěšné!';
 $_lang["file_unzip_fail"] = 'Rozbalení selhalo!';
 $_lang["filemanager_path_message"] = 'IIS občas nedokáže správně pracovat s nastaveným koreňovým adresářem dokumentů, který používá souborový správce. Jestliže máte problémy s jeho používaním, nadefinujte tuto cestu do kořenového adresáře ve vaší instalaci Evolution CMS.';
 $_lang["filemanager_path_title"] = 'Umístění souborového správce:';
+$_lang["files"] = 'Files';
 $_lang["files_access_denied"] = 'Přístup zamítnutý!';
 $_lang["files_data"] = 'Data';
 $_lang["files_dir_listing"] = 'Výpis adresáře:';
@@ -476,16 +479,17 @@ $_lang["login_message"] = 'Vložte prosím své přihlasovací jméno a heslo pr
 $_lang["logo_slogan"] = 'EVO Content Manager - \nCreate and do more with less';
 $_lang["logout"] = 'Odhlásit';
 $_lang["long_title"] = 'Dlouhý název';
+$_lang["manage_categories"] = 'Manage Categories';
 $_lang["manage_depends"] = 'Správa závislostí';
 $_lang["manage_files"] = 'Správa souborů';
-$_lang["manage_htmlsnippets"] = 'Chunky';
+$_lang["manage_htmlsnippets"] = 'Manage Chunks';
 $_lang["manage_metatags"] = 'Správa META tagů a klíčových slov';
 $_lang["manage_modules"] = 'Správa modulů';
-$_lang["manage_plugins"] = 'Pluginy';
-$_lang["manage_snippets"] = 'Snippety';
-$_lang["manage_templates"] = 'Šablony';
-$_lang["manage_documents"] = 'Documents';
-$_lang["manage_permission"] = 'Permissions';
+$_lang["manage_plugins"] = 'Manage Plugins';
+$_lang["manage_snippets"] = 'Manage Snippets';
+$_lang["manage_templates"] = 'Manage Templates';
+$_lang["manage_documents"] = 'Manage Documents';
+$_lang["manage_permission"] = 'Manage Permissions';
 $_lang["manager"] = 'Správce';
 $_lang["manager_lockout_message"] = 'Jste přihlášen do Správce obsahu. Pokud chcete opustit Evolution CMS, klikněte na tlačítko "Ohlásit". <p />Pro spuštění nebo přechod na domovskou stránku klikněte na tlačítko "Domů".';
 $_lang["manager_permissions"] = 'Správa přístupových práv';
@@ -1032,6 +1036,7 @@ $_lang["template_assignedtv_tab"] = 'Přiřazené Template Variables';
 $_lang["template_code"] = 'Kód šablony (html)';
 $_lang["template_desc"] = 'Popis';
 $_lang["template_edit_tab"] = 'Upravit šablonu';
+$_lang["template_inuse"] = 'This template is in use. Please set the documents using the template to another template. Documents using this template:';
 $_lang["template_management_msg"] = 'Zde můžete zvolit šablonu, kterou chcete upravit.';
 $_lang["template_msg"] = 'Zde můžete vytvářet/upravovat šablony. Nové anebo upravené šablony nebudou viditelné ve stránkach uložených v zásobníku až do jeho vyprázdnění. Můžete použít náhled stránek pro prohlédnutí šablony v činnosti.';
 $_lang["template_name"] = 'Název šablony';
@@ -1284,10 +1289,6 @@ $_lang["setting_resource_tree_node_name_desc_add"] = 'Note: Since Evolution CMS 
 $_lang["resource_opt_alvisibled"] = 'Use current alias in alias path';
 $_lang["resource_opt_alvisibled_help"] = 'The alias of this Resource is inserted in Friendly URL alias path';
 $_lang['resource_opt_is_published'] = 'Published';
-$_lang["docid_incrmnt_method_title"] = 'Increment Resource ID method';
-$_lang["docid_incrmnt_method_0"] = 'DB auto increment';
-$_lang["docid_incrmnt_method_1"] = 'Minimum missed ID';
-$_lang["docid_incrmnt_method_2"] = 'Maximal ID + 1';
 
 $_lang["enable_cache_title"] = 'Document caching';
 $_lang["disable_chunk_cache_title"] = 'Disable chunk caching';
