@@ -37,6 +37,7 @@ $_lang["access_permissions_off"] = '<span class="warning">Adgangsstyringen er ik
 $_lang["access_permissions_resource_groups"] = 'Ressource grupper';
 $_lang["access_permissions_resources_in_group"] = '<b>Ressourcer i gruppen:</b> ';
 $_lang["access_permissions_resources_tab"] = 'Se hvilke ressource grupper der er oprettet. Du kan også oprette nye grupper, omdøbe grupper, slette grupper, samt se hvilke ressourcer der er i de respektive grupper. (Hold musen over ressourcens ID for at se dets navn). For at tilføje eller fjerne en ressource til en gruppe, skal du redigere selve ressourcen.';
+$_lang["access_permissions_user_toggle"] = 'Toggle access permissions';
 $_lang["access_permissions_user_groups"] = 'Bruger grupper';
 $_lang["access_permissions_user_message"] = 'Vælg bruger gruppen som denne bruger skal tilhøre:';
 $_lang["access_permissions_users_in_group"] = 'Brugere i gruppen:';
@@ -84,10 +85,11 @@ $_lang["captcha_title"] = 'Brug følgende CAPTCHA koder:';
 $_lang["captcha_words_default"] = 'MODX,Access,Better,BitCode,Chunk,Cache,Desc,Design,Excell,Enjoy,URLs,TechView,Gerald,Griff,Humphrey,Holiday,Intel,Integration,Joystick,Join(),Oscope,Genetic,Light,Likeness,Marit,Maaike,Niche,Netherlands,Ordinance,Oscillo,Parser,Phusion,Query,Question,Regalia,Righteous,Snippet,Sentinel,Template,Thespian,Unity,Enterprise,Verily,Tattoo,Veri,Website,WideWeb,Yap,Yellow,Zebra,Zygote';
 $_lang["captcha_words_message"] = 'Indtast en liste af CAPTCHA ord som bliver brugt såfremt CAPTCHA er aktiveret. Adskil de enkelte ord med komma. Feltet kan maksimalt indeholde 255 karakterer.';
 $_lang["captcha_words_title"] = 'CAPTCHA ord';
+$_lang["category"] = 'Category';
+$_lang["categories"] = 'Categories';
 $_lang["category_heading"] = 'Kategori';
 $_lang["category_manager"] = 'Category Manager';
 $_lang["category_management"] = 'Category management';
-$_lang["manage_categories"] = 'Manage Categories';
 $_lang["category_msg"] = 'Se og rediger alle elementer, grupperet efter kategori.';
 $_lang["cfg_base_path"] = 'MODX_BASE_PATH';
 $_lang["cfg_base_url"] = 'MODX_BASE_URL';
@@ -139,7 +141,7 @@ $_lang["configcheck_notok"] = 'Der er en eller flere fejl i konfigurationen: ';
 $_lang["configcheck_ok"] = 'Checket er gennemført korrekt - uden advarsler.';
 $_lang["configcheck_php_gdzip"] = 'GD og/eller Zip udvidelserne i PHP blev ikke fundet';
 $_lang["configcheck_php_gdzip_msg"] = 'EVO behøver GD og Zip udvidelserne i PHP. Selvom EVO vil kunne bruges uden disse, vil du ikke få det fulde udbytte af den indbyggede filbrowser, billede editeringen eller Captcha til brug ved login.';
-$_lang["configcheck_rb_base_dir"] = 'The currently set <a href="index.php?a=17&tab=5">File Browser path</a> seems incorrect.';
+$_lang["configcheck_rb_base_dir"] = 'The currently set <a href="index.php?a=17&tab=5">File base path</a> seems incorrect.';
 $_lang["configcheck_rb_base_dir_msg"] = 'This can happen for example by moving your installation to a different directory or server. Please check and update your Evolution CMS configuration.';
 $_lang["configcheck_register_globals"] = 'register_globals er sat til ON i din php.ini fil';
 $_lang["configcheck_register_globals_msg"] = 'Det betyder, at websitet kan blive mere udsat for "Cross Site Scripting" (XSS) angreb. Du bør kontakte system administratoren eller hosting firmaet for at deaktivere dette, såfremt du ikke selv kan gøre det.';
@@ -322,6 +324,7 @@ $_lang["file_unzip"] = 'Filen blev udpakket!';
 $_lang["file_unzip_fail"] = 'Udpakningen mislykkedes!';
 $_lang["filemanager_path_message"] = 'Ofte udfylder IIS ikke \'document_root\' indstillingen korrekt. Denne bliver brugt af fil browseren til at angive, hvad man kan se. Hvis der er problemer med fil browseren, så kontroller om denne sti henviser til roden af EVO installationen.';
 $_lang["filemanager_path_title"] = 'Fil browser sti:';
+$_lang["files"] = 'Files';
 $_lang["files_access_denied"] = 'Adgang nægtet!';
 $_lang["files_data"] = 'Data';
 $_lang["files_dir_listing"] = 'Sti:';
@@ -476,16 +479,17 @@ $_lang["login_message"] = 'Indtast brugernavn og kodeord for at få adgang til C
 $_lang["logo_slogan"] = 'EVO Content Manager - \nSkab og gør mere med mindre';
 $_lang["logout"] = 'Log ud';
 $_lang["long_title"] = 'Lang titel';
+$_lang["manage_categories"] = 'Manage Categories';
 $_lang["manage_depends"] = 'Administrer afhængigheder';
 $_lang["manage_files"] = 'Filhåndtering';
-$_lang["manage_htmlsnippets"] = 'Chunks';
+$_lang["manage_htmlsnippets"] = 'Manage Chunks';
 $_lang["manage_metatags"] = 'Administrer META tags og keywords';
 $_lang["manage_modules"] = 'Administrer Moduler';
-$_lang["manage_plugins"] = 'Plugins';
-$_lang["manage_snippets"] = 'Snippets';
-$_lang["manage_templates"] = 'Skabeloner';
-$_lang["manage_documents"] = 'Documents';
-$_lang["manage_permission"] = 'Permissions';
+$_lang["manage_plugins"] = 'Manage Plugins';
+$_lang["manage_snippets"] = 'Manage Snippets';
+$_lang["manage_templates"] = 'Manage Templates';
+$_lang["manage_documents"] = 'Manage Documents';
+$_lang["manage_permission"] = 'Manage Permissions';
 $_lang["manager"] = 'Administration';
 $_lang["manager_lockout_message"] = 'Du er i øjeblikket logget ind i CMS\'et. Hvis du vil afslutte dit log ind, skal du klikke på  "Log ud" knappen. <p />For at komme til startsiden, skal du klikke på "Startside" knappen.';
 $_lang["manager_permissions"] = 'Adgangstilladelser for CMS\'ets brugere';
@@ -1030,6 +1034,7 @@ $_lang["template_assignedtv_tab"] = 'Tildelte skabelon variabler';
 $_lang["template_code"] = 'Skabelon kode (HTML)';
 $_lang["template_desc"] = 'Beskrivelse';
 $_lang["template_edit_tab"] = 'Rediger skabelon';
+$_lang["template_inuse"] = 'This template is in use. Please set the documents using the template to another template. Documents using this template:';
 $_lang["template_management_msg"] = 'Vælg skabelonen der skal redigeres.';
 $_lang["template_msg"] = 'Opret og rediger skabeloner. Nye eller ændrede skabeloner vil ikke være synlige, før websites cache er nulstillet. Man kan dog bruge preview funktionen på en ressource for at se resultatet af skabelonen.';
 $_lang["template_name"] = 'Skabelon navn';
@@ -1282,10 +1287,6 @@ $_lang["setting_resource_tree_node_name_desc_add"] = 'Note: Since Evolution CMS 
 $_lang["resource_opt_alvisibled"] = 'Use current alias in alias path';
 $_lang["resource_opt_alvisibled_help"] = 'The alias of this Resource is inserted in Friendly URL alias path';
 $_lang['resource_opt_is_published'] = 'Published';
-$_lang["docid_incrmnt_method_title"] = 'Increment Resource ID method';
-$_lang["docid_incrmnt_method_0"] = 'DB auto increment';
-$_lang["docid_incrmnt_method_1"] = 'Minimum missed ID';
-$_lang["docid_incrmnt_method_2"] = 'Maximal ID + 1';
 
 $_lang["enable_cache_title"] = 'Document caching';
 $_lang["disable_chunk_cache_title"] = 'Disable chunk caching';
