@@ -37,6 +37,7 @@ $_lang["access_permissions_off"] = '<span class="warning">Zugriffsberechtigungen
 $_lang["access_permissions_resource_groups"] = 'Ressourcen-Gruppen';
 $_lang["access_permissions_resources_in_group"] = '<b>Ressourcen in Gruppe:</b> ';
 $_lang["access_permissions_resources_tab"] = 'Hier sehen Sie die angelegten Ressourcen-Gruppen. Sie können auch neue Gruppen anlegen, Gruppen umbenennen, Gruppen löschen und sehen welche Ressourcen in den unterschiedlichen Gruppen enthalten sind (den Namen der Ressource sehen Sie, in dem Sie mit der Maus über die ID fahren). Um eine Ressource zu einer Gruppe hinzuzufügen oder von einer Gruppe zu entfernen, ändern Sie die entsprechende Ressource direkt.';
+$_lang["access_permissions_user_toggle"] = 'Toggle access permissions';
 $_lang["access_permissions_user_groups"] = 'Benutzergruppen';
 $_lang["access_permissions_user_message"] = 'Hier können Sie auswählen welcher Benutzergruppe dieser Benutzer zugeordnet ist:';
 $_lang["access_permissions_users_in_group"] = 'Benutzer in Gruppe:';
@@ -84,10 +85,11 @@ $_lang["captcha_title"] = 'Captcha aktivieren';
 $_lang["captcha_words_default"] = 'EVO,Access,Better,BitCode,Chunk,Cache,Desc,Design,Excell,Enjoy,URLs,TechView,Gerald,Griff,Humphrey,Holiday,Intel,Integration,Joystick,Join(),Oscope,Genetic,Light,Likeness,Marit,Maaike,Niche,Netherlands,Ordinance,Oscillo,Parser,Phusion,Query,Question,Regalia,Righteous,Snippet,Sentinel,Template,Thespian,Unity,Enterprise,Verily,Tattoo,Veri,Website,WideWeb,Yap,Yellow,Zebra,Zygote';
 $_lang["captcha_words_message"] = 'Wenn Captcha aktiviert ist können Sie eine Liste von Captcha-Wörter eingeben. Trennen Sie die Wörter mit Kommas. Das Eingabe Feld ist mit 255 Zeichen beschränkt.';
 $_lang["captcha_words_title"] = 'Captcha-Wörter';
+$_lang["category"] = 'Category';
+$_lang["categories"] = 'Categories';
 $_lang["category_heading"] = 'Kategorie';
 $_lang["category_manager"] = 'Kategorie-Manager';
 $_lang["category_management"] = 'Kategorie-Verwaltung';
-$_lang["manage_categories"] = 'Kategorien verwalten';
 $_lang["category_msg"] = 'Hier können Sie alle Elemente nach Kategorien sortiert bearbeiten.';
 $_lang["cfg_base_path"] = 'MODX_BASE_PATH';
 $_lang["cfg_base_url"] = 'MODX_BASE_URL';
@@ -322,6 +324,7 @@ $_lang["file_unzip"] = 'Unzip erfolgreich!';
 $_lang["file_unzip_fail"] = 'Unzip fehlgeschlagen!';
 $_lang["filemanager_path_message"] = 'Unter Umständen erkennt Ihr Webserver (z.B. IIS) den Pfad nicht immer korrekt, überprüfen Sie deshalb diese Einstellung. Wenn Sie Probleme mit der Benutzung des Datei-Managers haben, stellen Sie sicher dass dieser Pfad auf das Root-Verzeichnis Ihrer EVO-Installation verweist.';
 $_lang["filemanager_path_title"] = 'Datei-Manager-Pfad';
+$_lang["files"] = 'Files';
 $_lang["files_access_denied"] = 'Zugriff verweigert!';
 $_lang["files_data"] = 'Größe';
 $_lang["files_dir_listing"] = 'Verzeichnisübersicht für:';
@@ -476,16 +479,17 @@ $_lang["login_message"] = 'Bitte melden Sie sich an, um mit EVO zu arbeiten.<br 
 $_lang["logo_slogan"] = 'EVO Content Manager - \nCreate and do more with less';
 $_lang["logout"] = 'Abmelden';
 $_lang["long_title"] = 'Langer Titel';
+$_lang["manage_categories"] = 'Kategorien verwalten';
 $_lang["manage_depends"] = 'Abhängigkeiten verwalten';
 $_lang["manage_files"] = 'Datei-Manager';
-$_lang["manage_htmlsnippets"] = 'Chunks';
+$_lang["manage_htmlsnippets"] = 'Manage Chunks';
 $_lang["manage_metatags"] = 'Meta-Angaben verwalten';
 $_lang["manage_modules"] = 'Module verwalten';
-$_lang["manage_plugins"] = 'Plugins';
-$_lang["manage_snippets"] = 'Snippets';
-$_lang["manage_templates"] = 'Template-Verwaltung';
-$_lang["manage_documents"] = 'Dokumente';
-$_lang["manage_permission"] = 'Berechtigungen';
+$_lang["manage_plugins"] = 'Manage Plugins';
+$_lang["manage_snippets"] = 'Manage Snippets';
+$_lang["manage_templates"] = 'Manage Templates';
+$_lang["manage_documents"] = 'Manage Documents';
+$_lang["manage_permission"] = 'Manage Permissions';
 $_lang["manager"] = 'Manager';
 $_lang["manager_lockout_message"] = '<p>Sie sind gegenwärtig im EVO-Manager angemeldet. Wenn Sie Ihre Login-Session beenden wollen, klicken Sie bitte den Abmelden-Button.</p><p>Um zu Ihrer Start- oder Home-Seite zu gelangen, klicken Sie bitte den Start-Button.';
 $_lang["manager_permissions"] = 'Manager-Zugriffberechtigungen';
@@ -1031,6 +1035,7 @@ $_lang["template_assignedtv_tab"] = 'Zugewiesene Template-Variablen';
 $_lang["template_code"] = 'Template-Code (HTML)';
 $_lang["template_desc"] = 'Template-Beschreibung';
 $_lang["template_edit_tab"] = 'Template bearbeiten';
+$_lang["template_inuse"] = 'This template is in use. Please set the documents using the template to another template. Documents using this template:';
 $_lang["template_management_msg"] = 'Hier können Sie ein Template wählen, das Sie bearbeiten möchten.';
 $_lang["template_msg"] = 'Hier können Sie Templates anlegen und bearbeiten. Die Änderung an einem Template wird nur angezeigt, wenn Sie den Cache leeren. Sie können aber die Vorschaufunktion nutzen, um das Template anzuzeigen.';
 $_lang["template_name"] = 'Template-Name';
@@ -1283,10 +1288,6 @@ $_lang["setting_resource_tree_node_name_desc_add"] = 'Hinweis: Ab EVO 1.1 ist es
 $_lang["resource_opt_alvisibled"] = 'Alias im Alias-Pfad ausgeben';
 $_lang["resource_opt_alvisibled_help"] = 'Den Alias dieser Ressource im generierten Alias-Pfad ausgeben.';
 $_lang['resource_opt_is_published'] = 'Veröffentlicht';
-$_lang["docid_incrmnt_method_title"] = 'Vergabe einer neuen Ressourcen-ID ';
-$_lang["docid_incrmnt_method_0"] = 'DB Autoinkrement';
-$_lang["docid_incrmnt_method_1"] = 'Kleinste fehlende ID';
-$_lang["docid_incrmnt_method_2"] = 'Größte ID + 1';
 
 $_lang["enable_cache_title"] = 'Ressourcen-Cache';
 $_lang["disable_chunk_cache_title"] = 'Chunk-Caching deaktivieren';
