@@ -37,6 +37,7 @@ $_lang["access_permissions_off"] = '<span class="warning">Access Permissions are
 $_lang["access_permissions_resource_groups"] = 'Resource Groups';
 $_lang["access_permissions_resources_in_group"] = '<b>Resources in group:</b> ';
 $_lang["access_permissions_resources_tab"] = 'See which Resource Groups have been created. Also create new groups, rename groups, delete groups and see which Resources are in the different groups (hover over the id of the Resource to see it\'s name). To add a Resource to a group or remove a Resource from a group, edit the Resource directly.';
+$_lang["access_permissions_user_toggle"] = 'Toggle access permissions';
 $_lang["access_permissions_user_groups"] = 'User groups';
 $_lang["access_permissions_user_message"] = 'Select which User Groups this user belongs to:';
 $_lang["access_permissions_users_in_group"] = 'Users in group:';
@@ -84,10 +85,11 @@ $_lang["captcha_title"] = 'Use CAPTCHA codes';
 $_lang["captcha_words_default"] = 'EVO,Access,Better,BitCode,Chunk,Cache,Desc,Design,Excell,Enjoy,URLs,TechView,Gerald,Griff,Humphrey,Holiday,Intel,Integration,Joystick,Join(),Oscope,Genetic,Light,Likeness,Marit,Maaike,Niche,Netherlands,Ordinance,Oscillo,Parser,Phusion,Query,Question,Regalia,Righteous,Snippet,Sentinel,Template,Thespian,Unity,Enterprise,Verily,Tattoo,Veri,Website,WideWeb,Yap,Yellow,Zebra,Zygote';
 $_lang["captcha_words_message"] = 'Enter a list of CAPTCHA words to use if CAPTCHA is enabled. Separate the words with commas. This input field is limited to 255 characters.';
 $_lang["captcha_words_title"] = 'CAPTCHA Words';
+$_lang["category"] = 'Category';
+$_lang["categories"] = 'Categories';
 $_lang["category_heading"] = 'Category';
 $_lang["category_manager"] = 'Category Manager';
 $_lang["category_management"] = 'Category management';
-$_lang["manage_categories"] = 'Manage Categories';
 $_lang["category_msg"] = 'View and edit all Elements grouped by category.';
 $_lang["cfg_base_path"] = 'MODX_BASE_PATH';
 $_lang["cfg_base_url"] = 'MODX_BASE_URL';
@@ -322,6 +324,7 @@ $_lang["file_unzip"] = 'Unzip was successful!';
 $_lang["file_unzip_fail"] = 'Unzip Failed!';
 $_lang["filemanager_path_message"] = 'IIS often does not populate the document_root setting properly, which is used by the file manager to determine what you can see. If you are having problems using the file manager, make sure this path points to the root of your Evolution CMS installation.';
 $_lang["filemanager_path_title"] = 'File Manager path';
+$_lang["files"] = 'Files';
 $_lang["files_access_denied"] = 'Access denied!';
 $_lang["files_data"] = 'Data';
 $_lang["files_dir_listing"] = 'Directory listing for:';
@@ -476,16 +479,17 @@ $_lang["login_message"] = 'Please enter your login credentials to start your Man
 $_lang["logo_slogan"] = 'Evolution CMS - \nCreate and do more with less';
 $_lang["logout"] = 'Logout';
 $_lang["long_title"] = 'Long title';
+$_lang["manage_categories"] = 'Manage Categories';
 $_lang["manage_depends"] = 'Manage Dependencies';
 $_lang["manage_files"] = 'Manage Files';
-$_lang["manage_htmlsnippets"] = 'Chunks';
+$_lang["manage_htmlsnippets"] = 'Manage Chunks';
 $_lang["manage_metatags"] = 'Manage META tags and Keywords';
 $_lang["manage_modules"] = 'Manage Modules';
-$_lang["manage_plugins"] = 'Plugins';
-$_lang["manage_snippets"] = 'Snippets';
-$_lang["manage_templates"] = 'Templates';
-$_lang["manage_documents"] = 'Documents';
-$_lang["manage_permission"] = 'Permissions';
+$_lang["manage_plugins"] = 'Manage Plugins';
+$_lang["manage_snippets"] = 'Manage Snippets';
+$_lang["manage_templates"] = 'Manage Templates';
+$_lang["manage_documents"] = 'Manage Documents';
+$_lang["manage_permission"] = 'Manage Permissions';
 $_lang["manager"] = 'Manager';
 $_lang["manager_lockout_message"] = 'You are currently logged into the Content Manager. If you would like to close your login session please click the "Logout" button. <p />To go to your startup or home page click the "Home" button.';
 $_lang["manager_permissions"] = 'Manager Permissions';
@@ -1031,6 +1035,7 @@ $_lang["template_assignedtv_tab"] = 'Assigned Template Variables';
 $_lang["template_code"] = 'Template code (html)';
 $_lang["template_desc"] = 'Description';
 $_lang["template_edit_tab"] = 'Edit Template';
+$_lang["template_inuse"] = 'This template is in use. Please set the documents using the template to another template. Documents using this template:';
 $_lang["template_management_msg"] = 'Choose which Template you wish to edit.';
 $_lang["template_msg"] = 'Create and edit Templates. Changed or new Templates won\'t be visible in your site\'s cached pages until the cache is emptied, however, you can use the preview function on a page to see the Template in action.';
 $_lang["template_name"] = 'Template name';
@@ -1283,10 +1288,6 @@ $_lang["setting_resource_tree_node_name_desc_add"] = 'Note: Since Evolution CMS 
 $_lang["resource_opt_alvisibled"] = 'Use current alias in alias path';
 $_lang["resource_opt_alvisibled_help"] = 'The alias of this Resource is inserted in Friendly URL alias path';
 $_lang['resource_opt_is_published'] = 'Published';
-$_lang["docid_incrmnt_method_title"] = 'Increment Resource ID method';
-$_lang["docid_incrmnt_method_0"] = 'DB auto increment';
-$_lang["docid_incrmnt_method_1"] = 'Minimum missed ID';
-$_lang["docid_incrmnt_method_2"] = 'Maximal ID + 1';
 
 $_lang["enable_cache_title"] = 'Document caching';
 $_lang["disable_chunk_cache_title"] = 'Disable chunk caching';
