@@ -42,6 +42,7 @@ $_lang["access_permissions_resource_groups"] = 'Grupos de recursos';
 $_lang["access_permissions_resources_in_group"] = '<b>Los recursos en el grupo:</b> ';
 $_lang["access_permissions_resources_tab"] = 'Vea que grupos de recursos se han creado. También cree los grupos nuevos, cambie los nombres de los grupos, elimine los grupos y vea que recursos están en los diferentes grupos (pase el cursor sobre la ID del recurso para ver su nombre).
 Para añadir un recurso a un grupo o removerlo edite el recurso directamente.';
+$_lang["access_permissions_user_toggle"] = 'Toggle access permissions';
 $_lang["access_permissions_user_groups"] = 'Grupos de usuarios';
 $_lang["access_permissions_user_message"] = 'Seleccione a que grupos de usuarios pertenece este usuario:';
 $_lang["access_permissions_users_in_group"] = 'Usuarios en el grupo:';
@@ -70,8 +71,8 @@ $_lang["alias"] = 'Alias URL';
 $_lang["all_doc_groups"] = 'Todos los grupos de recursos (públicos)';
 $_lang["all_events"] = 'Todos los eventos';
 $_lang["all_usr_groups"] = 'Todos los grupos de usuarios (públicos)';
-$_lang["allow_mgr_access"] = 'Acceso al interfase de administrador';
-$_lang["allow_mgr_access_message"] = 'Seleccione esta opción para habilitar o deshabilitar el acceso a la interfase de administrador.
+$_lang["allow_mgr_access"] = 'Acceso a la interfaz de administrador';
+$_lang["allow_mgr_access_message"] = 'Seleccione esta opción para habilitar o deshabilitar el acceso a la interfaz de administrador.
 <b>Nota:</b> Si esta opción está configurada como NO el usuario será redirigido a la página de ingreso de administrador o al Inicio de sitio.';
 $_lang["already_deleted"] = 'ha sido borrado ya.';
 $_lang["attachment"] = 'Adjunto';
@@ -91,10 +92,11 @@ $_lang["captcha_title"] = 'Usar los códigos CAPTCHA:';
 $_lang["captcha_words_default"] = 'EVO,Aceso,Mejor,BitCode,Chunk,Cache,Desc,Diseno,Excel,Disfruta,URLs,TechView,Gerald,Griff,Humphrey,Vacacion,Intel,Integracion,Joystick,Join(),Oscope,Genetico,Luz,Parecido,Marit,Maaike,Nicho,Holanda,Ordinancia,Oscillo,Parser,Fusion,Query,Pregunta,Regalia,Correcto,Snippet,Sentinela,Template,Thespian,Unidad,Enterprise,Verily,Tatuaje,Veri,Website,WideWeb,Yap,Amarillo,Zebra,Zigoto';
 $_lang["captcha_words_message"] = 'Ingrese usted una lista de palabras CAPTCHA para usarlas si el CAPTCHA está activado. Las palabras tienen que estar separadas con las comas. Este campo de entrada está limitado a los 255 caracteres.';
 $_lang["captcha_words_title"] = 'Palabras de CAPTCHA';
+$_lang["category"] = 'Category';
+$_lang["categories"] = 'Categories';
 $_lang["category_heading"] = 'Categoría';
 $_lang["category_manager"] = 'Administrador de categorías';
 $_lang["category_management"] = 'Gestión para las categorías';
-$_lang["manage_categories"] = 'Gestionar las categorías';
 $_lang["category_msg"] = 'Aquí puede ver y editar todos los elementos agrupados por la categoría.';
 $_lang["cfg_base_path"] = 'MODX_BASE_PATH';
 $_lang["cfg_base_url"] = 'MODX_BASE_URL';
@@ -286,8 +288,8 @@ $_lang["elements"] = 'Elementos';
 $_lang["email"] = 'Correo electrónico';
 $_lang["email_sent"] = 'Correo electrónico fue enviado';
 $_lang["email_unique"] = 'Correo eléctronico está en uso ya';
-$_lang["emailsender_message"] = 'La dirección de correo electrónico del administrador de sitio. Por ejemplo, esta dirección de correo electrónico se utilizará como destino del correo electrónico de notificación del sistema, etc.';
-$_lang["emailsender_title"] = 'Dirección de correo electrónico:';
+$_lang["emailsender_message"] = 'El correo electrónico de administrador de sitio. Por ejemplo, este correo electrónico se utilizará como destino del correo electrónico de notificación de sistema, etc.';
+$_lang["emailsender_title"] = 'Correo electrónico:';
 $_lang["emailsubject_default"] = 'Sus detalles de entrada';
 $_lang["emailsubject_message"] = 'Aquí puede especificar el valor del asunto para el correo electrónico de bienvenida.';
 $_lang["emailsubject_title"] = 'Asunto del correo electrónico:';
@@ -331,6 +333,7 @@ $_lang["file_unzip"] = '¡La descompresión fue exitosa!';
 $_lang["file_unzip_fail"] = '¡La descompresión falló!';
 $_lang["filemanager_path_message"] = 'IIS a menudo no actualiza la configuración de document_root correctamente, que es usado por el administrador de archivos para determinar qué puedes ver. Si está teniendo algunas problemas usando el administrador de archivos, asegúrese que esta dirección apunta a la raíz de la instalación Evolution CMS.';
 $_lang["filemanager_path_title"] = 'Dirección del administrador de archivos:';
+$_lang["files"] = 'Files';
 $_lang["files_access_denied"] = '¡El acceso fue denegado!';
 $_lang["files_data"] = 'Datos';
 $_lang["files_dir_listing"] = 'Lista de carpeta para:';
@@ -492,16 +495,17 @@ $_lang["login_message"] = 'Por favor ingrese sus credenciales de entrada para co
 $_lang["logo_slogan"] = 'El administrador de contenido de Evolution CMS - cree y haga más con menos';
 $_lang["logout"] = 'Salir';
 $_lang["long_title"] = 'Título largo';
+$_lang["manage_categories"] = 'Gestionar las categorías';
 $_lang["manage_depends"] = 'Administrar las dependencias';
 $_lang["manage_files"] = 'Administrar los archivos';
-$_lang["manage_htmlsnippets"] = 'Trozos (chunks)';
+$_lang["manage_htmlsnippets"] = 'Manage Chunks';
 $_lang["manage_metatags"] = 'Administrar las etiquetas META y las palabras claves';
 $_lang["manage_modules"] = 'Administrar los módulos';
-$_lang["manage_plugins"] = 'Extensiónes (plugins)';
-$_lang["manage_snippets"] = 'Fragmentos (snippets)';
-$_lang["manage_templates"] = 'Plantillas (templates)';
-$_lang["manage_documents"] = 'Recursos';
-$_lang["manage_permission"] = 'Permisos';
+$_lang["manage_plugins"] = 'Manage Plugins';
+$_lang["manage_snippets"] = 'Manage Snippets';
+$_lang["manage_templates"] = 'Manage Templates';
+$_lang["manage_documents"] = 'Manage Documents';
+$_lang["manage_permission"] = 'Manage Permissions';
 $_lang["manager"] = 'El administrador';
 $_lang["manager_lockout_message"] = 'Está actualmente dentro del administrador de contenido. Si desea cerrar su sesión haga clic en el botón \'Salir\'. Para ir a su página de inicio o la página principal haga clic en el botón \'Inicio\'.';
 $_lang["manager_permissions"] = 'Permisos de administrador';
@@ -614,7 +618,7 @@ $_lang["offline"] = 'Fuera de línea';
 $_lang["online"] = 'En línea';
 $_lang["onlineusers_action"] = 'Acción';
 $_lang["onlineusers_actionid"] = 'ID-de-acción';
-$_lang["onlineusers_ipaddress"] = 'Dirección IP de usuario';
+$_lang["onlineusers_ipaddress"] = 'Dirección IP';
 $_lang["onlineusers_lasthit"] = 'Última visita';
 $_lang["onlineusers_message"] = 'Esta lista muestra todos los usuarios activos durante los últimos 20 minutos (la hora actual es ';
 $_lang["onlineusers_title"] = 'Usuarios en línea';
@@ -946,7 +950,7 @@ $_lang["settings_strip_image_paths_message"] = 'Si se establece en \'NO\', Evolu
 $_lang["settings_strip_image_paths_title"] = '¿Re-escribir las direcciones del navegador?';
 $_lang["settings_templvars"] = 'Variables de plantilla (TV)';
 $_lang["settings_title"] = 'Configuración de sistema';
-$_lang["settings_ui"] = 'El interfaz y las características';
+$_lang["settings_ui"] = 'Interfaz y Características';
 $_lang["settings_users"] = 'Usuario';
 $_lang["settings_email_templates"] = 'Correo electrónico y plantillas';
 $_lang["show_meta"] = 'Mostrar la pestaña de palabras clave META';
@@ -1053,6 +1057,7 @@ $_lang["template_assignedtv_tab"] = 'Variables de plantilla (TV) asignadas';
 $_lang["template_code"] = 'Código de plantilla (html)';
 $_lang["template_desc"] = 'Descripción';
 $_lang["template_edit_tab"] = 'Editar la plantilla (template)';
+$_lang["template_inuse"] = 'This template is in use. Please set the documents using the template to another template. Documents using this template:';
 $_lang["template_management_msg"] = 'Eliga usted que plantilla desea usted editar.';
 $_lang["template_msg"] = 'Crea y edita las plantillas. Las plantillas nuevas o modificadas no serán visibles en las páginas cacheadas de tu sitio hasta que el caché sea vaciado, sin embargo, puede usted usar la función de previsualización en una página para ver el template en acción.';
 $_lang["template_name"] = 'Nombre de plantilla (template)';
@@ -1154,7 +1159,7 @@ $_lang["user_country"] = 'País';
 $_lang["user_dob"] = 'La fecha de nacimiento';
 $_lang["user_doesnt_exist"] = 'Usuario no existe';
 $_lang["user_edit_self_msg"] = '<b>Es posible que deba cerrar sesión y volver a iniciar sesión después de guardar para actualizar completamente su información.</b> Las contraseñas nuevas se enviarán a su dirección de correo electrónico o se mostrarán en la pantalla.';
-$_lang["user_email"] = 'Dirección de correo electrónico';
+$_lang["user_email"] = 'Correo electrónico';
 $_lang["user_failedlogincount"] = 'Ingresos fallidos';
 $_lang["user_fax"] = 'Fax';
 $_lang["user_female"] = 'Femenino';
@@ -1325,10 +1330,6 @@ $_lang["setting_resource_tree_node_name_desc_add"] = 'Nota: desde la version 1.1
 $_lang["resource_opt_alvisibled"] = 'Utilizar el alias actual en la ruta de aliases';
 $_lang["resource_opt_alvisibled_help"] = 'El alias de este recurso se insertará en la ruta de alias de URL amigable';
 $_lang['resource_opt_is_published'] = 'Oublicado';
-$_lang["docid_incrmnt_method_title"] = 'Método de incremento del ID de recursos';
-$_lang["docid_incrmnt_method_0"] = 'DB auto incremental';
-$_lang["docid_incrmnt_method_1"] = 'Mínimo ID perdido';
-$_lang["docid_incrmnt_method_2"] = 'Máximo ID + 1';
 
 $_lang["enable_cache_title"] = 'Caché de recursos';
 $_lang["disable_chunk_cache_title"] = 'Deshabilitar el caché de trozos (chunks)';
