@@ -37,6 +37,7 @@ $_lang["access_permissions_off"] = '<span class="warning">سطوح دسترسی 
 $_lang["access_permissions_resource_groups"] = 'گروههای پرونده';
 $_lang["access_permissions_resources_in_group"] = '<b>پرونده های این گروه:</b> ';
 $_lang["access_permissions_resources_tab"] = 'در اینجا شما میتوانید مشاهده کنید که چه گروههای پرونده ای تهیه شده است. همچنین شما قادر خواهید بود گروههای جدیدی را تعریف و ایجاد کنید, نام گروه ها را تغییر دهید, گروه ها را حذف کنید و ببنید که چه پرونده ای در چه چه گروهی قرار دارد (نشانگر موس را روی ردیف یا شناسه ی پرونده برانید تا اسم آنرا مشاهده کنید). برای افزودن پرونده به گروه یا حذف پرونده از یک گروه, پوشه ی پرونده را ویرایش کنید.';
+$_lang["access_permissions_user_toggle"] = 'Toggle access permissions';
 $_lang["access_permissions_user_groups"] = 'گروههای کاربرات';
 $_lang["access_permissions_user_message"] = 'در اینجا میتوانید مشخص کنید که این کاربر به چه گروههایی از کاربری متعلق است:';
 $_lang["access_permissions_users_in_group"] = '<b>کاربران در گروه:</b> ';
@@ -84,10 +85,11 @@ $_lang["captcha_title"] = 'استفاده از گزاره های حفاظتی:';
 $_lang["captcha_words_default"] = 'MODX,Access,Better,BitCode,Chunk,Cache,Desc,Design,Excell,Enjoy,URLs,TechView,Gerald,Griff,Humphrey,Holiday,Intel,Integration,Joystick,Join(),Oscope,Genetic,Light,Likeness,Marit,Maaike,Niche,Netherlands,Ordinance,Oscillo,Parser,Phusion,Query,Question,Regalia,Righteous,Snippet,Sentinel,Template,Thespian,Unity,Enterprise,Verily,Tattoo,Veri,Website,WideWeb,Yap,Yellow,Zebra,Zygote';
 $_lang["captcha_words_message"] = 'در اینجا شما میتوانید فهرستی مرتب از گزاره های حفاظتی را برای استفاده از آنها در صورتیکه این امکان فعال شده باشد وارد کنید. گزاره ها را با کاما از هم جدا کنید. این قسمت سقف محدود بلندا تا میزان 255 گزاره را دارد.';
 $_lang["captcha_words_title"] = 'کلمات حفاظتی';
+$_lang["category"] = 'Category';
+$_lang["categories"] = 'Categories';
 $_lang["category_heading"] = 'شاخه یا کتگوری';
 $_lang["category_manager"] = 'Category Manager';
 $_lang["category_management"] = 'Category management';
-$_lang["manage_categories"] = 'Manage Categories';
 $_lang["category_msg"] = 'در اینجا شما میتوانید کلیه ی منابع جمعبندی شده بوسیله ی شاخه یا کتگوری را مرور کنید.';
 $_lang["cfg_base_path"] = 'MODX_BASE_PATH';
 $_lang["cfg_base_url"] = 'MODX_BASE_URL';
@@ -139,7 +141,7 @@ $_lang["configcheck_notok"] = 'یک یا چند مورد از تنظیمات ب�
 $_lang["configcheck_ok"] = 'آزمایش موفقیت آمیز بود - هیچ خطایی برای گزارش یافت نشد.';
 $_lang["configcheck_php_gdzip"] = 'GD and/or Zip PHP extensions not found';
 $_lang["configcheck_php_gdzip_msg"] = 'Evolution CMS needs the GD and Zip extension enabled for PHP. While Evolution CMS will work without them, you will not be able to take full advantage of the built-in File Manager, Image Editor or Captcha for logins.';
-$_lang["configcheck_rb_base_dir"] = 'The currently set <a href="index.php?a=17&tab=5">File Browser path</a> seems incorrect.';
+$_lang["configcheck_rb_base_dir"] = 'The currently set <a href="index.php?a=17&tab=5">File base path</a> seems incorrect.';
 $_lang["configcheck_rb_base_dir_msg"] = 'This can happen for example by moving your installation to a different directory or server. Please check and update your Evolution CMS configuration.';
 $_lang["configcheck_register_globals"] = 'رجیستر_گلوبال در تنظیمات فایل اصلی PHP سرور میزبان روشن می باشد';
 $_lang["configcheck_register_globals_msg"] = ' این تنظیمات وبگاه شما را در مقابل حملات کراس سایت اسکریپتها آسیب پذیر میکند. شما بایستی برای رفع این مورد با پشتیبانی سرور میزبان خود تماس بگیرید.';
@@ -322,6 +324,7 @@ $_lang["file_unzip"] = 'بازکردن فایل موفقیت آمیز بود!';
 $_lang["file_unzip_fail"] = 'بازکردن فایل ناموفق بود!';
 $_lang["filemanager_path_message"] = 'آی آی اس اغلب مواقع تنظیمات مخصوص داکیومنت_روت را به درستی پر نمی کند, که این تنظیمات برای بخش مدیریت فایلها استفاده میشود تا مشخص کند شما چه چیزی را میتوانید مشاهده کنید. در صورتیکه در استفاده ی بخش مدیریت فایلها مشکلی داشتید, از این موضوع اطمینان حاصل کنید که این پت یا محل تعیین شده به درستی روی روت نصّاب مادایکس تنظیم شده است.';
 $_lang["filemanager_path_title"] = 'محل مدیریت فایلها :';
+$_lang["files"] = 'Files';
 $_lang["files_access_denied"] = 'دسترسی شما محدود است';
 $_lang["files_data"] = 'داده';
 $_lang["files_dir_listing"] = 'تهیه ی فهرست از پوشه یا دایره به منظور :';
@@ -476,16 +479,17 @@ $_lang["login_message"] = 'لطفا برای ورود به بخش مدیریت �
 $_lang["logo_slogan"] = 'مدیریت محتوای مادایکس - \nبا حداقل امکان بسازید بلکه برتر از آن';
 $_lang["logout"] = 'خروج از سیستم';
 $_lang["long_title"] = 'عنوان بلند';
+$_lang["manage_categories"] = 'Manage Categories';
 $_lang["manage_depends"] = 'مدیریت متعلقات';
 $_lang["manage_files"] = 'مدیریت فایلها';
-$_lang["manage_htmlsnippets"] = 'چانک ها';
+$_lang["manage_htmlsnippets"] = 'Manage Chunks';
 $_lang["manage_metatags"] = 'مدیریت کلمات کلیدی و کلمات شاخص';
 $_lang["manage_modules"] = 'مدیریت ماژولها';
-$_lang["manage_plugins"] = 'پلاگینها';
-$_lang["manage_snippets"] = 'اسنیپتها';
-$_lang["manage_templates"] = 'قالبها یا پوسته ها';
-$_lang["manage_documents"] = 'Documents';
-$_lang["manage_permission"] = 'Permissions';
+$_lang["manage_plugins"] = 'Manage Plugins';
+$_lang["manage_snippets"] = 'Manage Snippets';
+$_lang["manage_templates"] = 'Manage Templates';
+$_lang["manage_documents"] = 'Manage Documents';
+$_lang["manage_permission"] = 'Manage Permissions';
 $_lang["manager"] = 'مدیر';
 $_lang["manager_lockout_message"] = 'همکنون شما وارد بخش مدیریت محتوا شده اید. هر زمان که تمایل به خروج از این بخش دارید لطفا روی دکمه ی "خروج" کلیک کنید. <p />برای ارجاع به صفحه ی اصلی خود روی دکمه ی "منزل" کلیک کنید.';
 $_lang["manager_permissions"] = 'سطوح دسترسی مدیر';
@@ -1028,6 +1032,7 @@ $_lang["template_assignedtv_tab"] = 'متغیرهای قالب';
 $_lang["template_code"] = 'کدهای HTML قالب';
 $_lang["template_desc"] = 'توضیحات';
 $_lang["template_edit_tab"] = 'ویرایش پوسته یا قالب';
+$_lang["template_inuse"] = 'This template is in use. Please set the documents using the template to another template. Documents using this template:';
 $_lang["template_management_msg"] = 'در اینجا میتوانید پوسته یا قالب مورد نظر خود را برای ویرایش انتخاب کنید';
 $_lang["template_msg"] = 'پوسته ها و قالبها را بسازید و ویرایش کنید. پوسته ها یا قالبهای دستکاری شده یا جدید در قسمت ذخیره یا کش وبگاه قابل مشاهده نخواهند بود تا زمانیکه ذخیره یا کش تخلیه شده باشد.به هر حال, شما میتوانید از قابلیت پیش نمایش در صفحه استفاده کنید تا پوسته و قالب دستکاری شده یا جدید خود را بصورت زنده و پویا مشاهده کنید';
 $_lang["template_name"] = 'اسم پوسته یا قالب';
@@ -1280,10 +1285,6 @@ $_lang["setting_resource_tree_node_name_desc_add"] = 'Note: Since Evolution CMS 
 $_lang["resource_opt_alvisibled"] = 'Use current alias in alias path';
 $_lang["resource_opt_alvisibled_help"] = 'The alias of this Resource is inserted in Friendly URL alias path';
 $_lang['resource_opt_is_published'] = 'انتشار یابد؟';
-$_lang["docid_incrmnt_method_title"] = 'Increment Resource ID method';
-$_lang["docid_incrmnt_method_0"] = 'DB auto increment';
-$_lang["docid_incrmnt_method_1"] = 'Minimum missed ID';
-$_lang["docid_incrmnt_method_2"] = 'Maximal ID + 1';
 
 $_lang["enable_cache_title"] = 'Document caching';
 $_lang["disable_chunk_cache_title"] = 'Disable chunk caching';
