@@ -37,6 +37,7 @@ $_lang["access_permissions_off"] = '<span class="warning">Käyttöoikeuksia ei o
 $_lang["access_permissions_resource_groups"] = 'Sivuryhmät';
 $_lang["access_permissions_resources_in_group"] = '<strong>Sivut ryhmässä:</strong> ';
 $_lang["access_permissions_resources_tab"] = 'Muokkaa sivuryhmiä. Lisää yksittäinen sivu ryhmään tai poista yksittäinen sivu ryhmästä muokkaamalla suoraan kyseistä sivua.';
+$_lang["access_permissions_user_toggle"] = 'Toggle access permissions';
 $_lang["access_permissions_user_groups"] = 'Ylläpidon käyttäjäryhmät';
 $_lang["access_permissions_user_message"] = 'Valitse käyttäjän käyttäjäryhmät:';
 $_lang["access_permissions_users_in_group"] = '<strong>Käyttäjiä ryhmässä:</strong> ';
@@ -84,10 +85,11 @@ $_lang["captcha_title"] = 'Käytä varmistuskoodeja:';
 $_lang["captcha_words_default"] = 'MODX,Todellisuus,Oikeus,Turvallisuus,Koodi,Palanen,Parempi,Kuvaus,Voitto,Edistynyt,Nauti,Osoite,Ohjain,Geneettinen,Valo,Kevyt,Mukavuus,Suomi,Internet,Kysely,Sivupohja,Usein,Nettisivut,Tatuointi,Keltainen,Punainen,Sininen,Ruskea,Musta,Valkoinen,Oranssi,Nauru,Mukana,Laskuvarjo,Uinti,Fantasia,Palindromi,Rehellisyys,Nopea,Ihminen,Tietokone,Palapeli,Monitori,Ikkuna,Valinta,Painike';
 $_lang["captcha_words_message"] = 'Lista varmistuskoodeina (CAPTCHA) käytetävistä sanoista. Erottele sanat pilkuilla. Kenttään mahtuu enintään 255 merkkiä.';
 $_lang["captcha_words_title"] = 'Varmistuskoodi -sanat (CAPTCHA):';
+$_lang["category"] = 'Category';
+$_lang["categories"] = 'Categories';
 $_lang["category_heading"] = 'Kategoria';
 $_lang["category_manager"] = 'Category Manager';
 $_lang["category_management"] = 'Category management';
-$_lang["manage_categories"] = 'Manage Categories';
 $_lang["category_msg"] = 'Muokkaa elementtejä.';
 $_lang["cfg_base_path"] = 'MODX_BASE_PATH';
 $_lang["cfg_base_url"] = 'MODX_BASE_URL';
@@ -139,7 +141,7 @@ $_lang["configcheck_notok"] = 'Varoitukset: ';
 $_lang["configcheck_ok"] = 'Tarkistus onnistui - ei varoituksia.';
 $_lang["configcheck_php_gdzip"] = 'GD ja/tai Zip PHP-laajennuksia ei löytynyt';
 $_lang["configcheck_php_gdzip_msg"] = 'EVO vaatii GD ja Zip PHP-laajennukset toimiakseen täydellisesti. EVO toimii ilmankin kyseisiä laajennuksia, mutta tiedostoselaimen, kuvaeditorin ja Captcha -varmistuskoodien kaikki ominaisuudet eivät ole käytettävissä.';
-$_lang["configcheck_rb_base_dir"] = 'The currently set <a href="index.php?a=17&tab=5">File Browser path</a> seems incorrect.';
+$_lang["configcheck_rb_base_dir"] = 'The currently set <a href="index.php?a=17&tab=5">File base path</a> seems incorrect.';
 $_lang["configcheck_rb_base_dir_msg"] = 'This can happen for example by moving your installation to a different directory or server. Please check and update your Evolution CMS configuration.';
 $_lang["configcheck_register_globals"] = 'Palvelimen php.ini-tiedossa register_globals on asetettu päälle ';
 $_lang["configcheck_register_globals_msg"] = 'Tämä asetus tekee sivuston alttiiksi Cross Site Scripting- eli XSS-hyökkäyksille. Kysy palveluntarjoajaltasi, miten asia voitaisiin ratkaista.';
@@ -322,6 +324,7 @@ $_lang["file_unzip"] = 'Purku onnistui!';
 $_lang["file_unzip_fail"] = 'Purku epäonnistui!';
 $_lang["filemanager_path_message"] = 'IIS ei useinkaan muokkaa tiedostonhallintajärjestelmän käyttämää document_root asetusta oikein. Tätä asetusta käytetään määrittämään, mitä tiedostoja käyttäjä voi katsella. Jos tiedostonhallintajärjestelmän käytössä on vaikeuksia, huolehdi siitä, että polku osoittaa EVO-asennuksen juureen.';
 $_lang["filemanager_path_title"] = 'Ylläpidon tiedostoselaimen polku:';
+$_lang["files"] = 'Files';
 $_lang["files_access_denied"] = 'Pääsy estetty!';
 $_lang["files_data"] = 'Ominaisuudet';
 $_lang["files_dir_listing"] = 'Kansio:';
@@ -476,16 +479,17 @@ $_lang["login_message"] = 'Kirjaudu ylläpitoon. Isoilla ja pienillä kirjaimill
 $_lang["logo_slogan"] = 'EVO Sisällönhallintajärjestelmä - \nLuo ja tee enemmän vähemmällä';
 $_lang["logout"] = 'Kirjaudu ulos';
 $_lang["long_title"] = 'Pitkä otsikko';
+$_lang["manage_categories"] = 'Manage Categories';
 $_lang["manage_depends"] = 'Muokkaa riippuvuuksia';
 $_lang["manage_files"] = 'Tiedostot';
-$_lang["manage_htmlsnippets"] = 'HTML-palaset';
+$_lang["manage_htmlsnippets"] = 'Manage Chunks';
 $_lang["manage_metatags"] = 'Metatiedot ja avainsanat';
 $_lang["manage_modules"] = 'Moduulit';
-$_lang["manage_plugins"] = 'Liitännäiset';
-$_lang["manage_snippets"] = 'PHP-palaset';
-$_lang["manage_templates"] = 'Sivupohjat';
-$_lang["manage_documents"] = 'Documents';
-$_lang["manage_permission"] = 'Permissions';
+$_lang["manage_plugins"] = 'Manage Plugins';
+$_lang["manage_snippets"] = 'Manage Snippets';
+$_lang["manage_templates"] = 'Manage Templates';
+$_lang["manage_documents"] = 'Manage Documents';
+$_lang["manage_permission"] = 'Manage Permissions';
 $_lang["manager"] = 'Ylläpito';
 $_lang["manager_lockout_message"] = 'Olet kirjautuneena ylläpitoon. Jos haluat kirjautua ulos, napsauta "Kirjaudu ulos" -painiketta.<p />';
 $_lang["manager_permissions"] = 'Ylläpidon käyttöoikeudet';
@@ -1027,6 +1031,7 @@ $_lang["template_assignedtv_tab"] = 'Määritetyt sivupohjan muuttujat';
 $_lang["template_code"] = 'Lähdekoodi (html)';
 $_lang["template_desc"] = 'Kuvaus';
 $_lang["template_edit_tab"] = 'Muokkaa sivupohjaa';
+$_lang["template_inuse"] = 'This template is in use. Please set the documents using the template to another template. Documents using this template:';
 $_lang["template_management_msg"] = 'Muokkaa sivupohjia (template).';
 $_lang["template_msg"] = 'Muokkaa sivupohjaa (template). Muokatut sivupohjat otetaan käyttöön vasta, kun välimuisti on tyhjennetty. Tätä ennen voit tarkastella muokattua sivupohjaa käytössä sivun esikatselutoiminnon avulla.';
 $_lang["template_name"] = 'Nimi';
@@ -1279,10 +1284,6 @@ $_lang["setting_resource_tree_node_name_desc_add"] = 'Note: Since Evolution CMS 
 $_lang["resource_opt_alvisibled"] = 'Käytä nykyistä aliasta aliaspolussa';
 $_lang["resource_opt_alvisibled_help"] = 'Sivun alias tulee osaksi osoitteen selkokielistä aliaspolkua';
 $_lang['resource_opt_is_published'] = 'Julkaistu';
-$_lang["docid_incrmnt_method_title"] = 'Tapa jonka mukaan uuden sivun ID numero määräytyy';
-$_lang["docid_incrmnt_method_0"] = 'Tietokantajärjestelmän automaattinen ID numeron kasvatus (auto increment)';
-$_lang["docid_incrmnt_method_1"] = 'Pienin vapaana oleva luku';
-$_lang["docid_incrmnt_method_2"] = 'Suurin käytössä oleva ID numero + 1';
 
 $_lang["enable_cache_title"] = 'Document caching';
 $_lang["disable_chunk_cache_title"] = 'Disable chunk caching';
