@@ -37,6 +37,7 @@ $_lang["access_permissions_off"] = '<span class="warning">Les autorisations d\'a
 $_lang["access_permissions_resource_groups"] = 'Groupes de Ressources';
 $_lang["access_permissions_resources_in_group"] = '<b>Ressources dans le groupe:</b> ';
 $_lang["access_permissions_resources_tab"] = 'Les Groupes de Ressources actuels sont affichés ci-dessous. Vous pouvez en créer de nouveaux, les renommer, les supprimer et voir quelles Ressources appartiennent aux différents groupes (le survol par le pointeur de l\'ID de la Ressource en affichera le nom). Pour ajouter une Ressource à un groupe ou pour l\'en retirer, modifiez directement la Ressource.';
+$_lang["access_permissions_user_toggle"] = 'Toggle access permissions';
 $_lang["access_permissions_user_groups"] = 'Groupes d\'Utilisateurs';
 $_lang["access_permissions_user_message"] = 'Vous pouvez choisir ici les Groupes d\'Utilisateurs auxquels cet utilisateur appartient';
 $_lang["access_permissions_users_in_group"] = 'Utilisateurs dans le groupe:';
@@ -84,10 +85,11 @@ $_lang["captcha_title"] = 'Utiliser les codes CAPTCHA:';
 $_lang["captcha_words_default"] = 'MODX,Access,Better,BitCode,Chunk,Cache,Desc,Design,Excell,Enjoy,URLs,TechView,Gerald,Griff,Humphrey,Holiday,Intel,Integration,Joystick,Join(),Oscope,Genetic,Light,Likeness,Marit,Maaike,Niche,Netherlands,Ordinance,Oscillo,Parser,Phusion,Query,Question,Regalia,Righteous,Snippet,Sentinel,Template,Thespian,Unity,Enterprise,Verily,Tattoo,Veri,Website,WideWeb,Yap,Yellow,Zebra,Zygote';
 $_lang["captcha_words_message"] = 'Veuillez saisir une liste de mots à utiliser si CAPTCHA est activé. Séparez les mots par des virgules. Ce champ est limité à 255 caractères.';
 $_lang["captcha_words_title"] = 'Mots CAPTCHA';
+$_lang["category"] = 'Category';
+$_lang["categories"] = 'Categories';
 $_lang["category_heading"] = 'Catégorie';
 $_lang["category_manager"] = 'Gestionnaire de catégories';
 $_lang["category_management"] = 'Gestion par catégorie';
-$_lang["manage_categories"] = 'Gérer les catégories';
 $_lang["category_msg"] = 'Vous pouvez consulter et éditer ici tous les Éléments classés par catégorie.';
 $_lang["cfg_base_path"] = 'MODX_BASE_PATH';
 $_lang["cfg_base_url"] = 'MODX_BASE_URL';
@@ -322,6 +324,7 @@ $_lang["file_unzip"] = 'Les fichiers ont été décompressés!';
 $_lang["file_unzip_fail"] = 'Les fichiers n\'ont pas pu être décompressés!';
 $_lang["filemanager_path_message"] = 'Souvent, IIS ne remplit pas correctement le réglage document_root, qui est utilisé par le Gestionnaire de Fichiers pour déterminer ce que vous pouvez voir. Si vous rencontrez des problèmes avec le Gestionnaire de Fichiers, assurez-vous que ce chemin pointe bien vers la racine de votre installation EVO.';
 $_lang["filemanager_path_title"] = 'Chemin du Gestionnaire de Fichiers:';
+$_lang["files"] = 'Files';
 $_lang["files_access_denied"] = 'Accès interdit!';
 $_lang["files_data"] = 'Données';
 $_lang["files_dir_listing"] = 'Contenu du répertoire pour:';
@@ -476,16 +479,17 @@ $_lang["login_message"] = 'Veuillez vous identifier afin d\'accéder au Gestionn
 $_lang["logo_slogan"] = 'Faire plus avec moins - Gestionnaire de Contenu EVO';
 $_lang["logout"] = 'Déconnexion';
 $_lang["long_title"] = 'Titre long';
+$_lang["manage_categories"] = 'Gérer les catégories';
 $_lang["manage_depends"] = 'Gestion des Dépendances';
 $_lang["manage_files"] = 'Gestion des Fichiers';
-$_lang["manage_htmlsnippets"] = 'Chunks';
+$_lang["manage_htmlsnippets"] = 'Manage Chunks';
 $_lang["manage_metatags"] = 'Gestion des balises META et des mots-clés';
 $_lang["manage_modules"] = 'Gestion des Modules';
-$_lang["manage_plugins"] = 'Plugins';
-$_lang["manage_snippets"] = 'Snippets';
-$_lang["manage_templates"] = 'Modèles';
-$_lang["manage_documents"] = 'Documents';
-$_lang["manage_permission"] = 'Autorisations';
+$_lang["manage_plugins"] = 'Manage Plugins';
+$_lang["manage_snippets"] = 'Manage Snippets';
+$_lang["manage_templates"] = 'Manage Templates';
+$_lang["manage_documents"] = 'Manage Documents';
+$_lang["manage_permission"] = 'Manage Permissions';
 $_lang["manager"] = 'Gestionnaire';
 $_lang["manager_lockout_message"] = 'Vous êtes déjà connecté au Gestionnaire de Contenu. Si vous souhaitez fermer votre session, merci de bien vouloir utiliser le bouton «Déconnexion». <br />Pour aller à la page d\'accueil, cliquez sur le bouton «Accueil».';
 $_lang["manager_permissions"] = 'Autorisations du Gestionnaire';
@@ -995,6 +999,7 @@ $_lang["template_assignedtv_tab"] = 'Variables de Modèle associées';
 $_lang["template_code"] = 'Code source du Modèle (HTML)';
 $_lang["template_desc"] = 'Description';
 $_lang["template_edit_tab"] = 'Modifier le Modèle';
+$_lang["template_inuse"] = 'This template is in use. Please set the documents using the template to another template. Documents using this template:';
 $_lang["template_management_msg"] = 'Choisissez le Modèle que vous souhaitez modifier.';
 $_lang["template_msg"] = 'Créer/modifier des Modèles. Les Modèles créés ou modifiés ne seront pas actualisés dans les pages en cache de votre site jusqu\'à ce que le cache soit vidé. Cependant, vous pouvez utiliser la fonction de prévisualisation d\'une page pour voir le nouveau Modèle en action.';
 $_lang["template_name"] = 'Nom du Modèle';
@@ -1247,10 +1252,6 @@ $_lang["setting_resource_tree_node_name_desc_add"] = 'Remarque : depuis Evolutio
 $_lang["resource_opt_alvisibled"] = 'Utiliser l\'alias du document dans chemin d\'alias';
 $_lang["resource_opt_alvisibled_help"] = 'L\'alias de ce document sera utilisé dans le chemin d\'alias';
 $_lang['resource_opt_is_published'] = 'Publié';
-$_lang["docid_incrmnt_method_title"] = 'Mode de gestion de l\'incrément des ID des documents';
-$_lang["docid_incrmnt_method_0"] = 'Incrément automatique de la base de données';
-$_lang["docid_incrmnt_method_1"] = 'Plus petit ID disponible';
-$_lang["docid_incrmnt_method_2"] = 'Plus grand ID + 1';
 
 $_lang["enable_cache_title"] = 'Mise en cache des documents';
 $_lang["disable_chunk_cache_title"] = 'Désactiver la mise en cache des chunks';
