@@ -37,6 +37,7 @@ $_lang["access_permissions_off"] = '<span class="warning">הרשאות גישה 
 $_lang["access_permissions_resource_groups"] = 'קבוצות מסמכים';
 $_lang["access_permissions_resources_in_group"] = '<b>מסמכים בקבוצה:</b> ';
 $_lang["access_permissions_resources_tab"] = 'כאן תוכלו לראות איזה קבוצות מסמכים הוגדרו. ניתן ליצור קבוצות חדשות, לשנות שם לקבוצות, למחוק קבוצות ולראות אילו מסמכים קיימים בקבוצות שונות (ניתן לראות את שם המסמך במעבר עכבר מעל מזהה המסמך). כדי להוסיף מסמך לקבוצה או להסיר מסמך מהקבוצה, יש לערוך את המסמך ישירות.';
+$_lang["access_permissions_user_toggle"] = 'Toggle access permissions';
 $_lang["access_permissions_user_groups"] = 'קבוצות משתמשים';
 $_lang["access_permissions_user_message"] = 'כאן אתה יכול לבחור לאילו קבוצות משתמשים משתמש זה שייך:';
 $_lang["access_permissions_users_in_group"] = 'משתמשים בקבוצות:';
@@ -84,10 +85,11 @@ $_lang["captcha_title"] = 'שימוש בקוד קאפצה:';
 $_lang["captcha_words_default"] = 'MODX,Access,Better,BitCode,Chunk,Cache,Desc,Design,Excell,Enjoy,URLs,TechView,Gerald,Griff,Humphrey,Holiday,Intel,Integration,Joystick,Join(),Oscope,Genetic,Light,Likeness,Marit,Maaike,Niche,Netherlands,Ordinance,Oscillo,Parser,Phusion,Query,Question,Regalia,Righteous,Snippet,Sentinel,Template,Thespian,Unity,Enterprise,Verily,Tattoo,Veri,Website,WideWeb,Yap,Yellow,Zebra,Zygote';
 $_lang["captcha_words_message"] = 'Enter a list of CAPTCHA words to use if CAPTCHA is enabled. Separate the words with commas. This input field is limited to 255 characters.';
 $_lang["captcha_words_title"] = 'מילות קאפצה';
+$_lang["category"] = 'Category';
+$_lang["categories"] = 'Categories';
 $_lang["category_heading"] = 'קטגוריה';
 $_lang["category_manager"] = 'Category Manager';
 $_lang["category_management"] = 'Category management';
-$_lang["manage_categories"] = 'Manage Categories';
 $_lang["category_msg"] = 'כאן תוכלו לצפות ולערוך את כל רכיבי הקבוצות לפי קטגוריות.';
 $_lang["cfg_base_path"] = 'MODX_BASE_PATH';
 $_lang["cfg_base_url"] = 'MODX_BASE_URL';
@@ -139,7 +141,7 @@ $_lang["configcheck_notok"] = 'One or more configuration details didn\'t check o
 $_lang["configcheck_ok"] = 'Check passed OK - no warnings to report.';
 $_lang["configcheck_php_gdzip"] = 'GD and/or Zip PHP extensions not found';
 $_lang["configcheck_php_gdzip_msg"] = 'Evolution CMS needs the GD and Zip extension enabled for PHP. While Evolution CMS will work without them, you will not be able to take full advantage of the built-in File Manager, Image Editor or Captcha for logins.';
-$_lang["configcheck_rb_base_dir"] = 'The currently set <a href="index.php?a=17&tab=5">File Browser path</a> seems incorrect.';
+$_lang["configcheck_rb_base_dir"] = 'The currently set <a href="index.php?a=17&tab=5">File base path</a> seems incorrect.';
 $_lang["configcheck_rb_base_dir_msg"] = 'This can happen for example by moving your installation to a different directory or server. Please check and update your Evolution CMS configuration.';
 $_lang["configcheck_register_globals"] = 'register_globals is set to ON in your php.ini configuration file';
 $_lang["configcheck_register_globals_msg"] = 'This configuration makes your site much more susceptible to Cross Site Scripting (XSS) attacks. You should speak to your host about what you can do to disable this setting.';
@@ -322,6 +324,7 @@ $_lang["file_unzip"] = 'פתיחת הקובץ הצליחה!';
 $_lang["file_unzip_fail"] = 'פתיחת הקובץ נכשלה!';
 $_lang["filemanager_path_message"] = 'IIS often does not populate the document_root setting properly, which is used by the file manager to determine what you can see. If you are having problems using the file manager, make sure this path points to the root of your Evolution CMS installation.';
 $_lang["filemanager_path_title"] = 'File Manager path';
+$_lang["files"] = 'Files';
 $_lang["files_access_denied"] = 'הגישה נדחתה!';
 $_lang["files_data"] = 'נתונים';
 $_lang["files_dir_listing"] = 'רשימת תיקיות ל:';
@@ -476,16 +479,17 @@ $_lang["login_message"] = 'אנא הזן את פרטי ההתחברות שלך �
 $_lang["logo_slogan"] = 'Evolution CMS - \nCreate and do more with less';
 $_lang["logout"] = 'התנתק';
 $_lang["long_title"] = 'כותרת ארוכה';
+$_lang["manage_categories"] = 'Manage Categories';
 $_lang["manage_depends"] = 'ניהול תלויות';
 $_lang["manage_files"] = 'ניהול קבצים';
-$_lang["manage_htmlsnippets"] = ' (Chunks) מיני תבניות';
+$_lang["manage_htmlsnippets"] = 'Manage Chunks';
 $_lang["manage_metatags"] = 'ניהול מילות מפתח ותגיות מטה';
 $_lang["manage_modules"] = 'ניהול מודולים';
-$_lang["manage_plugins"] = '(Plugins) התקנים';
-$_lang["manage_snippets"] = '(Snippets) קטעי קוד';
-$_lang["manage_templates"] = 'תבניות עיצוב';
-$_lang["manage_documents"] = 'Documents';
-$_lang["manage_permission"] = 'Permissions';
+$_lang["manage_plugins"] = 'Manage Plugins';
+$_lang["manage_snippets"] = 'Manage Snippets';
+$_lang["manage_templates"] = 'Manage Templates';
+$_lang["manage_documents"] = 'Manage Documents';
+$_lang["manage_permission"] = 'Manage Permissions';
 $_lang["manager"] = 'מנהל';
 $_lang["manager_lockout_message"] = 'You are currently logged into the Content Manager. If you would like to close your login session please click the "Logout" button. <p />To go to your startup or home page click the "Home" button.';
 $_lang["manager_permissions"] = 'הרשאות מנהל';
@@ -1031,6 +1035,7 @@ $_lang["template_assignedtv_tab"] = 'משתני תבניות משוייכים';
 $_lang["template_code"] = 'קוד תבנית (html)';
 $_lang["template_desc"] = 'תיאור';
 $_lang["template_edit_tab"] = 'ערוך תבנית עיצוב';
+$_lang["template_inuse"] = 'This template is in use. Please set the documents using the template to another template. Documents using this template:';
 $_lang["template_management_msg"] = 'Choose which Template you wish to edit.';
 $_lang["template_msg"] = 'Create and edit Templates. Changed or new Templates won\'t be visible in your site\'s cached pages until the cache is emptied, however, you can use the preview function on a page to see the Template in action.';
 $_lang["template_name"] = 'שם תבנית עיצוב';
@@ -1283,10 +1288,6 @@ $_lang["setting_resource_tree_node_name_desc_add"] = 'Note: Since Evolution CMS 
 $_lang["resource_opt_alvisibled"] = 'Use current alias in alias path';
 $_lang["resource_opt_alvisibled_help"] = 'The alias of this Resource is inserted in Friendly URL alias path';
 $_lang['resource_opt_is_published'] = 'Published';
-$_lang["docid_incrmnt_method_title"] = 'שיטת קידום ערך הקוד של מסמך חדש';
-$_lang["docid_incrmnt_method_0"] = 'DB auto increment';
-$_lang["docid_incrmnt_method_1"] = 'Minimum missed ID';
-$_lang["docid_incrmnt_method_2"] = 'Maximal ID + 1';
 
 $_lang["enable_cache_title"] = 'Document caching';
 $_lang["disable_chunk_cache_title"] = 'Disable chunk caching';
