@@ -16,7 +16,7 @@
 $modx_lang_attribute = 'it'; // Manager HTML/XML Language Attribute see http://en.wikipedia.org/wiki/ISO_639-1
 $modx_manager_charset = 'UTF-8';
 
-$_lang["about_msg"] = 'Evolution CMS is a <a href="https://evo-cms.com/" target="_blank">PHP Application Framework and Content Management System</a> licensed under the <a href="https://www.gnu.org/licenses/gpl-3.0.html">GNU GPL</a>.';
+$_lang["about_msg"] = 'EVO è un <a href="https://evo-cms.com/" target="_blank">Framework PHP e Sistema di Gestione dei Contenuti (CMS)</a> distribuito con licenza <a href="https://www.gnu.org/licenses/gpl-3.0.html">GNU GPL</a>.';
 $_lang["about_title"] = 'Cos\'è EVO';
 $_lang["access_permission_denied"] = 'Non avete i permessi per accedere alla Risorsa.';
 $_lang["access_permission_parent_denied"] = 'Non avete il permesso di creare o spostare qui una Risorsa! Scegliete un\'altra destinazione.';
@@ -27,7 +27,7 @@ $_lang["access_permissions_docs_message"] = 'Qui potete selezionare a quali grup
 $_lang["access_permissions_group_link"] = 'Crea un nuovo gruppo';
 $_lang["access_permissions_introtext"] = 'Qui potete gestire i Gruppi di Utenti e i Gruppi di Risorse. Per aggiungere un utente a un gruppo, modificate l\'utente e selezionate il gruppo di cui deve fare parte. Per assegnare una Risorsa ad un Gruppo di Utenti, modificate la Risorsa e selezionate i gruppi a cui deve appartenere.';
 $_lang["access_permissions_link_to_group"] = 'al Gruppo Risorse';
-$_lang["access_permissions_context"] = 'in context';
+$_lang["access_permissions_context"] = 'nel contesto';
 $_lang["access_permissions_link_user_group"] = 'Link a Gruppo Utenti';
 $_lang["access_permissions_links"] = 'Collegamenti Gruppi di Utenti/Risorse';
 $_lang["access_permissions_links_tab"] = 'Qui potete specificare a quali Gruppi di Utenti è permesso l\'accesso ai Gruppi di Risorse (ad es.: possono modificare o creare Risorse). Per collegare un Gruppo di Risorse a un Gruppo di Utenti, selezionatelo dalla lista a tendina, e fate clic su \'Aggiungi gruppo\'. Per rimuovere il collegamento ad un certo gruppo premete \'Rimuovi ->\'. Il collegamento sarà immediatamente eliminato.';
@@ -37,6 +37,7 @@ $_lang["access_permissions_off"] = '<span class=\'warning\'>I permessi di access
 $_lang["access_permissions_resource_groups"] = 'Gruppi di Risorse';
 $_lang["access_permissions_resources_in_group"] = '<b>Risorse nel gruppo:</b> ';
 $_lang["access_permissions_resources_tab"] = 'Qui potete vedere i Gruppi di Risorse che sono stati impostati. Potete anche crearne di nuovi, rinominarli, eliminarli e vedere quali Risorse sono presenti nei differenti gruppi (posizionate il mouse sull\'ID della Risorsa per vederne il nome). Per aggiungere una Risorsa ad un gruppo o per rimuoverla da un gruppo, modificate direttamente la Risorsa.';
+$_lang["access_permissions_user_toggle"] = 'Toggle access permissions';
 $_lang["access_permissions_user_groups"] = 'Gruppi di Utenti';
 $_lang["access_permissions_user_message"] = 'Qui potete selezionare a quali Gruppi di Utenti appartiene l\'utente:';
 $_lang["access_permissions_users_in_group"] = '<b>Utenti del gruppo:</b> ';
@@ -84,10 +85,11 @@ $_lang["captcha_title"] = 'Usa codici CAPTCHA:';
 $_lang["captcha_words_default"] = 'EVO,Access,Better,BitCode,Chunk,Cache,Desc,Design,Excell,Enjoy,URLs,TechView,Gerald,Griff,Humphrey,Holiday,Intel,Integration,Joystick,Join(),Oscope,Genetic,Light,Likeness,Marit,Maaike,Niche,Netherlands,Ordinance,Oscillo,Parser,Phusion,Query,Question,Regalia,Righteous,Snippet,Sentinel,Template,Thespian,Unity,Enterprise,Verily,Tattoo,Veri,Website,WideWeb,Yap,Yellow,Zebra,Zygote';
 $_lang["captcha_words_message"] = 'Qui potete inserire una lista di parole CAPTCHA da usare quando CAPTCHA è abilitato. Separate le parole con virgole. Questo campo è limitato a 255 caratteri.';
 $_lang["captcha_words_title"] = 'Parole CAPTCHA:';
+$_lang["category"] = 'Category';
+$_lang["categories"] = 'Categories';
 $_lang["category_heading"] = 'Categoria';
 $_lang["category_manager"] = 'Gestione Categorie';
 $_lang["category_management"] = 'Gestione Categorie';
-$_lang["manage_categories"] = 'Gestione Categorie';
 $_lang["category_msg"] = 'Qui potete vedere tutte le Risorse per categoria.';
 $_lang["cfg_base_path"] = 'MODX_BASE_PATH';
 $_lang["cfg_base_url"] = 'MODX_BASE_URL';
@@ -174,7 +176,7 @@ $_lang["confirm_delete_tags"] = 'Siete sicuri di voler cancellare questo META ta
 $_lang["confirm_delete_template"] = 'Siete sicuri di voler eliminare il Template?';
 $_lang["confirm_delete_tmplvars"] = 'Siete sicuri di voler rimuovere la Variabile di Template e tutti i suoi valori memorizzati?';
 $_lang["confirm_delete_user"] = 'Siete sicuri di voler eliminare l\'utente?';
-$_lang["delete_yourself"] = 'You can\'t delete yourself';
+$_lang["delete_yourself"] = 'Non potete eliminare voi stessi!';
 $_lang["delete_last_admin"] = 'You can\'t delete last admin user';
 $_lang["confirm_delete_permission"] = 'Are you sure you want to delete this Permission?';
 $_lang["confirm_duplicate_record"] = 'Siete sicuri di voler duplicare questo record?';
@@ -277,7 +279,7 @@ $_lang["element_selector_title"] = 'Selettore Elementi';
 $_lang["elements"] = 'Elementi';
 $_lang["email"] = 'Email';
 $_lang["email_sent"] = 'Email inviata';
-$_lang["email_unique"] = 'Email is already in use!';
+$_lang["email_unique"] = 'Email già in uso!';
 $_lang["emailsender_message"] = 'L\'indirizzo e-mail dell\'amministratore del sito. Ad esempio, questo indirizzo e-mail verrà utilizzato come destinazione di notifica del sistema di e-mail, ecc';
 $_lang["emailsender_title"] = 'Indirizzo e-mail:';
 $_lang["emailsubject_default"] = 'Le vostre credenziali per il login';
@@ -322,6 +324,7 @@ $_lang["file_unzip"] = 'Decompressione cartella effettuata con successo!';
 $_lang["file_unzip_fail"] = 'Decompressione cartella fallita!';
 $_lang["filemanager_path_message"] = 'IIS spesso non riesce a determinare correttamente la cartella principale usata dal File Manager per determinare ciò che potete vedere. Se avete problemi nell\'uso del File Manager, assicuratevi che il percorso punti alla root dell\'installazione di EVO.';
 $_lang["filemanager_path_title"] = 'Percorso del File Manager:';
+$_lang["files"] = 'Files';
 $_lang["files_access_denied"] = 'Accesso negato!';
 $_lang["files_data"] = 'Dati';
 $_lang["files_dir_listing"] = 'Elenco cartelle per:';
@@ -476,16 +479,17 @@ $_lang["login_message"] = 'Inserisci le tue credenziali per effettuare l\'access
 $_lang["logo_slogan"] = 'Sistema di gestione dei contenuti EVO - \nCreate e fate di più con meno';
 $_lang["logout"] = 'Uscita';
 $_lang["long_title"] = 'Titolo esteso';
+$_lang["manage_categories"] = 'Gestione Categorie';
 $_lang["manage_depends"] = 'Configurazione Dipendenze';
 $_lang["manage_files"] = 'Gestione File';
-$_lang["manage_htmlsnippets"] = 'Chunks';
+$_lang["manage_htmlsnippets"] = 'Manage Chunks';
 $_lang["manage_metatags"] = 'Gestione META tag e parole chiave';
 $_lang["manage_modules"] = 'Gestione moduli';
-$_lang["manage_plugins"] = 'Plugins';
-$_lang["manage_snippets"] = 'Snippets';
-$_lang["manage_templates"] = 'Templates';
-$_lang["manage_documents"] = 'Documenti';
-$_lang["manage_permission"] = 'Permissions';
+$_lang["manage_plugins"] = 'Manage Plugins';
+$_lang["manage_snippets"] = 'Manage Snippets';
+$_lang["manage_templates"] = 'Manage Templates';
+$_lang["manage_documents"] = 'Manage Documents';
+$_lang["manage_permission"] = 'Manage Permissions';
 $_lang["manager"] = 'Manager';
 $_lang["manager_lockout_message"] = '<p>Siete entrati nel Manager EVO. Se desiderate uscire dalla sessione, fate clic sul pulsante "Uscita".</p>Per andare alla pagina inziale fate clic sul pulsante "Home".';
 $_lang["manager_permissions"] = 'Gestione Permessi';
@@ -561,7 +565,7 @@ $_lang["new_file_permissions_message"] = 'Durante il caricamento di un nuovo fil
 $_lang["new_file_permissions_title"] = 'Permessi dei nuovi files:';
 $_lang["new_folder_permissions_message"] = 'Durante la creazione di una nuova cartella nel sistema di gestione file, il sistema cercherà di cambiare i permessi della cartella secondo questa impostazione. Questo però potrebbe non funzionare in alcune configurazioni, come ad esempio IIS, nel qual caso dovrete procedere ad una modifica manuale dei permessi.';
 $_lang["new_folder_permissions_title"] = 'Permessi nuova cartella:';
-$_lang["new_permission"] = 'New Permission';
+$_lang["new_permission"] = 'Nuovo Permesso';
 $_lang["new_htmlsnippet"] = 'Nuovo Chunk';
 $_lang["new_keyword"] = 'Aggiungi nuova parola chiave:';
 $_lang["new_module"] = 'Nuovo Modulo';
@@ -1040,6 +1044,7 @@ $_lang["template_assignedtv_tab"] = 'Variabili di Template Assegnate';
 $_lang["template_code"] = 'Codice del Template (HTML)';
 $_lang["template_desc"] = 'Descrizione';
 $_lang["template_edit_tab"] = 'Modifica Template';
+$_lang["template_inuse"] = 'This template is in use. Please set the documents using the template to another template. Documents using this template:';
 $_lang["template_management_msg"] = 'Qui potete scegliere il Template da modificare o creare un nuovo Template.';
 $_lang["template_msg"] = 'Qui potete creare/modificare i Templates. I Template modificati o nuovi non saranno visibili nelle pagine in cache finchè la cache stessa non verrà svuotata. Ad ogni modo, potete usare la funzione di anteprima su una pagina per vedere il Template in azione.';
 $_lang["template_name"] = 'Nome del Template';
@@ -1144,9 +1149,9 @@ $_lang["user_failedlogincount"] = 'Login falliti';
 $_lang["user_fax"] = 'Numero di fax';
 $_lang["user_female"] = 'Femmina';
 $_lang["user_full_name"] = 'Nome completo';
-$_lang["user_first_name"] = 'First name';
-$_lang["user_last_name"] = 'Last Name';
-$_lang["user_middle_name"] = 'Middle Name';
+$_lang["user_first_name"] = 'Nome';
+$_lang["user_last_name"] = 'Cognome';
+$_lang["user_middle_name"] = 'Secondo nome';
 $_lang["user_gender"] = 'Sesso';
 $_lang["user_is_blocked"] = 'L\'utente è bloccato!';
 $_lang["user_logincount"] = 'Contatore login';
@@ -1159,15 +1164,15 @@ $_lang["user_photo"] = 'Foto dell\'Utente';
 $_lang["user_photo_message"] = 'Inserite l\'url dell\'immagine o caricate l\'immagine sul server.';
 $_lang["user_prevlogin"] = 'Ultima autenticazione';
 $_lang["user_role"] = 'Ruolo dell\'utente';
-$_lang["no_user_role"] = 'No user role';
+$_lang["no_user_role"] = 'Nessun ruolo utente';
 $_lang["user_state"] = 'Regione';
 $_lang["user_title"] = 'Crea/modifica utente';
 $_lang["user_upload_message"] = 'Se volete impedire che questo utente possa caricare files di questo tipo, assicuratevi che la casella \'Use Main Configuration Setting\' non sia selezionata e lasciate vuoto il campo.';
 $_lang["user_use_config"] = 'Usare le Impostazioni di Sistema';
-$_lang["user_verification"] = 'User is verified';
+$_lang["user_verification"] = 'Utente verificato';
 $_lang["user_zip"] = 'CAP';
 $_lang["username"] = 'Nome utente';
-$_lang["username_unique"] = 'User name is already in use!';
+$_lang["username_unique"] = 'Il nome utente è già in uso';
 $_lang["users"] = 'Utenti';
 $_lang["valid_hostnames_message"] = 'Inserendo una lista di nomi di host consentiti per questa installazione di EVO è possibile prevenire che con un attacco XSS si possa falsificare l\'impostazione di sistema site_url. Può essere importante farlo per alcuni tipi di server condivisi o altri server direttamente accessibili da un indirizzo IP. Il primo nome di host della lista verà usato quando HTTP_HOST non corrisponde ad alcun nome di host valido.';
 $_lang["valid_hostnames_title"] = 'Nomi di host consentiti:';
@@ -1189,7 +1194,7 @@ $_lang["web_access_permissions_user_groups"] = 'Gruppi di Utenti Web';
 $_lang["web_permissions"] = 'Permessi Web';
 $_lang["web_user_management_msg"] = 'Qui potete selezionare l\'utente web da modificare. Gli utenti web sono gli utenti autorizzati ad accedere al sito.';
 $_lang["web_user_management_title"] = 'Gestione Utenti Web';
-$_lang["web_user_management_select_role"] = 'All roles';
+$_lang["web_user_management_select_role"] = 'Tutti i ruoli';
 $_lang["web_user_title"] = 'Crea/modifica Utenti Web';
 $_lang["web_users"] = 'Utenti web';
 $_lang["weblink"] = 'Link Web';
@@ -1292,15 +1297,11 @@ $_lang["setting_resource_tree_node_name_desc_add"] = 'Nota: Dalla versione EVO 1
 $_lang["resource_opt_alvisibled"] = 'Usa l\'alias attuale nel percorso';
 $_lang["resource_opt_alvisibled_help"] = 'L\'alias di questa risora è inserito nel percorso degli Url Semplici';
 $_lang['resource_opt_is_published'] = 'Published';
-$_lang["docid_incrmnt_method_title"] = 'Metodo di incremento per l\'ID della Risorsa:';
-$_lang["docid_incrmnt_method_0"] = 'Incremento automatico DB';
-$_lang["docid_incrmnt_method_1"] = 'ID libero più basso';
-$_lang["docid_incrmnt_method_2"] = 'ID più grande + 1';
 
 $_lang["enable_cache_title"] = 'Cache delle risorse';
-$_lang["disable_chunk_cache_title"] = 'Disable chunk caching';
-$_lang["disable_snippet_cache_title"] = 'Disable snippet caching';
-$_lang["disable_plugins_cache_title"] = 'Disable plugins caching';
+$_lang["disable_chunk_cache_title"] = 'Disabilita il caching dei Chunk';
+$_lang["disable_snippet_cache_title"] = 'Disabilita il caching degli Snippet';
+$_lang["disable_plugins_cache_title"] = 'Disabilita il caching dei Plugins';
 $_lang["disabled_at_login"] = 'Disabilitato all\'accesso';
 
 $_lang["cache_type_title"] = 'Tipologia della cache dei documenti:';
@@ -1309,11 +1310,11 @@ $_lang["cache_type_2"] = 'La cache si basa sull\'ID della risorsa e sui parametr
 $_lang["seostrict_title"] = 'Usa URL univoci:';
 $_lang["seostrict_message"] = 'Forza l\'uso di URL SEO univoche per prevenire contenuti duplicati (se necessario)';
 
-$_lang["alias_listing_title"] = 'Use AliasListing cache';
+$_lang["alias_listing_title"] = 'Usa la cache AliasListing';
 $_lang["alias_listing_message"] = 'Caching page aliases, have to be disabled if a site have huge amount of resources. "Disabled" reduces memory consumption when site have large number of resources.';
-$_lang["alias_listing_disabled"] = 'Disabled';
+$_lang["alias_listing_disabled"] = 'Disabilitato';
 $_lang["alias_listing_folders"] = 'Only for folders';
-$_lang["alias_listing_enabled"] = 'Enabled';
+$_lang["alias_listing_enabled"] = 'Abilitato';
 
 $_lang["settings_friendlyurls_alert"] = 'E\' necessario rinominare il file ht.access presente nella root dell\'installazione di EVO in .htaccess per poter utilizzare gli URL Semplici.';
 $_lang["settings_friendlyurls_alert2"] = 'Dasto che EVO è installato in una sottodirectory è necessario modificare di conseguenza il file .htaccess.';
@@ -1507,7 +1508,7 @@ $_lang["role_tv_msg"] = 'The Variables assigned to this Role are listed below.';
 $_lang["tmplvar_roles_access_msg"] = 'Select the Roles that are allowed to access/process this Template Variable';
 
 $_lang["setting_from_file"] = '<strong class="text-danger">Parameter value is defined in core/custom/confic/cms/settings</strong>';
-$_lang['disable'] = 'Disable';
-$_lang['enable'] = 'Enable';
+$_lang['disable'] = 'Disabilita';
+$_lang['enable'] = 'Abilita';
 
 return $_lang;
