@@ -37,6 +37,7 @@ $_lang["access_permissions_off"] = '<span class="warning">グループ管理機�
 $_lang["access_permissions_resource_groups"] = 'リソースグループ';
 $_lang["access_permissions_resources_in_group"] = '<strong>グループ内リソース:</strong>';
 $_lang["access_permissions_resources_tab"] = 'IDの上にマウスポインタを重ねるとリソース名を確認できます。リソースをグループに加えるには、リソース編集画面を個別に開いて「所属グループ」タブで設定してください。管理画面とウェブページを区別して扱わないため、両方のグループを作る場合はグループ名を工夫するとよいでしょう。';
+$_lang["access_permissions_user_toggle"] = 'Toggle access permissions';
 $_lang["access_permissions_user_groups"] = 'ユーザーグループ';
 $_lang["access_permissions_user_message"] = 'このユーザーを所属させたいユーザーグループを選択してください。ユーザは所属しているグループに割り当てられたリソースのみを編集できます。';
 $_lang["access_permissions_users_in_group"] = 'グループ内ユーザー: ';
@@ -84,10 +85,11 @@ $_lang["captcha_title"] = 'CAPTCHAコードを使用';
 $_lang["captcha_words_default"] = 'maguro,toro,tako,ika,hotate,awabi,kazunoko,ebi,kani,uni,iwashi,aji,saba,tamago,negitoro,tekka,hamachi,sanma,sake,tai,buri,hirame,unagi,anago,amaebi,ikura,kanpachi,syako';
 $_lang["captcha_words_message"] = '「CAPTCHAコードの使用」を「はい」にした場合に使用されるCAPTCHAワードリストを入力してください。CAPTCHAワードはカンマで区切って入力します。このフィールドには、255文字まで入力することができます。';
 $_lang["captcha_words_title"] = 'CAPTCHA用ワード';
+$_lang["category"] = 'Category';
+$_lang["categories"] = 'Categories';
 $_lang["category_heading"] = 'カテゴリー';
 $_lang["category_manager"] = 'カテゴリーマネージャー';
 $_lang["category_management"] = 'カテゴリーの管理';
-$_lang["manage_categories"] = 'カテゴリー管理';
 $_lang["category_msg"] = '<h3 style="font-weight:bold;">全エレメントの一覧</h3>カテゴリーでグループ化したエレメントを一覧・編集できます。不要なカテゴリーの削除もこちらで。削除されるのはカテゴリー名だけで、エレメントは「カテゴリーなし」に移動します。';
 $_lang["cfg_base_path"] = 'MODX_BASE_PATH';
 $_lang["cfg_base_url"] = 'MODX_BASE_URL';
@@ -126,7 +128,7 @@ $_lang["configcheck_errorpage_unavailable"] = '設定した「エラーページ
 $_lang["configcheck_errorpage_unavailable_msg"] = '設定した「エラーページ」が一般的なエンドユーザーからアクセスできないページ（private）かあるいは存在しないことを意味しています。この場合、システムがループ状態に陥り、多くのエラーメッセージをエラーログに記録してしまいます。「エラーページ」には存在するページでかつ、公開ページ、かつ「public」なページを指定してください。';
 $_lang["configcheck_errorpage_unpublished"] = 'グローバル設定で設定されたエラーページは公開されていません。';
 $_lang["configcheck_errorpage_unpublished_msg"] = '設定したエラーページが誰でも見られるようにはなっていないことを意味します。エラーページは全ての人(Public)に公開するようにしてください。';
-$_lang["configcheck_filemanager_path"] = 'The currently set <a href="index.php?a=17&tab=4">File Manager path</a> seems incorrect.';
+$_lang["configcheck_filemanager_path"] = '現在設定されている <a href="/[+MGR_DIR+]/?a=17&amp;tab=4">ファイルマネージャーのパス</a>は不完全なようです。';
 $_lang["configcheck_filemanager_path_msg"] = '例えば、この問題は異なるディレクトリやサーバーに移動されたことで生じています。Evoのグローバル設定を確認し、更新してください。';
 $_lang["configcheck_hide_warning"] = '<a href="javascript:hideConfigCheckWarning(\'%s\');"><em>次からこの警告を表示しない</em></a>';
 $_lang["configcheck_images"] = 'images/ディレクトリに書き込みできません';
@@ -139,7 +141,7 @@ $_lang["configcheck_notok"] = '動作環境に問題があります。';
 $_lang["configcheck_ok"] = '問題なし - 警告はありません。';
 $_lang["configcheck_php_gdzip"] = 'GDライブラリが読み込まれてません。';
 $_lang["configcheck_php_gdzip_msg"] = 'このサーバでは<a href="http://www.google.com/search?q=gd+%22php.ini%22&lr=lang_ja" target="_blank">GDライブラリ</a>が読み込まれてません。<a href="http://www.google.com/cse?cx=007286147079563201032%3Aigbcdgg0jyo&ie=UTF-8&q=CAPTCHA&sa.x=26&sa.y=8&siteurl=http:%2F%2Fforum.modx.jp%2F" target="_blank">ログイン時の画像認証機能(CAPTCHA)</a>など画像まわりの処理に制限があります。';
-$_lang["configcheck_rb_base_dir"] = 'The currently set <a href="index.php?a=17&tab=5">File Browser path</a> seems incorrect.';
+$_lang["configcheck_rb_base_dir"] = '現在の設定している <a href="/[+MGR_DIR+]/#?a=17&amp;tab=5">ファイルのベースパス</a> は不完全のようです。';
 $_lang["configcheck_rb_base_dir_msg"] = '例えば、この問題は異なるディレクトリやサーバーに移動されたことで生じています。グローバル設定を確認し、更新してください。';
 $_lang["configcheck_register_globals"] = '「register_globals」が「ON」に設定されており、セキュリティ上問題があります。';
 $_lang["configcheck_register_globals_msg"] = 'register_globals がonになっているため、この環境は非常に脆弱です。.htaccessまたはphp.iniによって off にすることを強くおすすめします。';
@@ -322,6 +324,7 @@ $_lang["file_unzip"] = '解凍しました。';
 $_lang["file_unzip_fail"] = '解凍できません。';
 $_lang["filemanager_path_message"] = '<a href="index.php?a=31">「ファイル管理」</a>で扱える範囲を決定します。コンテンツ編集に利用するファイルブラウザと違い、同じサーバ内で権限を持つ領域であればどのディレクトリでも管理できます。ドキュメントルート外の指定も可能です。';
 $_lang["filemanager_path_title"] = 'ファイル管理機能の<br />管理対象ディレクトリ';
+$_lang["files"] = 'Files';
 $_lang["files_access_denied"] = 'このアカウントは休止中です。';
 $_lang["files_data"] = 'データ';
 $_lang["files_dir_listing"] = 'ディレクトリ一覧';
@@ -481,16 +484,17 @@ $_lang["login_message"] = 'ログイン名とパスワードは大文字・小�
 $_lang["logo_slogan"] = 'EVO Content Manager - \nCreate and do more with less';
 $_lang["logout"] = 'ログアウト';
 $_lang["long_title"] = 'タイトル';
+$_lang["manage_categories"] = 'カテゴリー管理';
 $_lang["manage_depends"] = '関連付けの管理';
 $_lang["manage_files"] = 'ファイル管理';
-$_lang["manage_htmlsnippets"] = 'チャンク';
+$_lang["manage_htmlsnippets"] = 'Manage Chunks';
 $_lang["manage_metatags"] = 'METAタグとキーワードの管理';
 $_lang["manage_modules"] = 'モジュールの管理';
-$_lang["manage_plugins"] = 'プラグイン';
-$_lang["manage_snippets"] = 'スニペット';
-$_lang["manage_templates"] = 'テンプレート';
-$_lang["manage_documents"] = 'ドキュメント';
-$_lang["manage_permission"] = 'パーミッション';
+$_lang["manage_plugins"] = 'Manage Plugins';
+$_lang["manage_snippets"] = 'Manage Snippets';
+$_lang["manage_templates"] = 'Manage Templates';
+$_lang["manage_documents"] = 'Manage Documents';
+$_lang["manage_permission"] = 'Manage Permissions';
 $_lang["manager"] = '管理画面';
 $_lang["manager_lockout_message"] = '管理画面にログインしています。セッションを閉じたいときは「ログアウト」ボタンをクリックしてください。<br />管理画面のホームあるいはスタート画面に行きたいときは「ホーム」ボタンをクリックします。';
 $_lang["manager_permissions"] = 'グループ管理';
@@ -1037,6 +1041,7 @@ $_lang["template_assignedtv_tab"] = '定義されているテンプレート変�
 $_lang["template_code"] = 'テンプレート コード (html)';
 $_lang["template_desc"] = '説明';
 $_lang["template_edit_tab"] = 'テンプレートの編集';
+$_lang["template_inuse"] = 'This template is in use. Please set the documents using the template to another template. Documents using this template:';
 $_lang["template_management_msg"] = '<h3 style="font-weight:bold;">テンプレートデザインの管理</h3><p>テンプレートはいくつでも作成でき、各ページ(リソース)ごとに自由に割り当てることができます。カッコ内の数字はテンプレートIDです。</p>';
 $_lang["template_msg"] = 'テンプレートを作成/編集します。[*リソース変数*] [*テンプレート変数*] [(コンフィグ変数)] [~リンクタグ~] [[スニペット]] {{チャンク}} [^ベンチマークタグ^] などを記述できます。';
 $_lang["template_name"] = 'テンプレート名';
@@ -1289,10 +1294,6 @@ $_lang["setting_resource_tree_node_name_desc_add"] = '注：EVO 1.1以降、こ�
 $_lang["resource_opt_alvisibled"] = 'エイリアスパス内で現在のエイリアスを使う';
 $_lang["resource_opt_alvisibled_help"] = 'このリソースのエイリアスはフレンドリーURKのエイリアスパスで挿入されます';
 $_lang['resource_opt_is_published'] = '発行完了';
-$_lang["docid_incrmnt_method_title"] = 'リソースIDの自動増分';
-$_lang["docid_incrmnt_method_0"] = 'データベース側で処理する(オートインクリメント)';
-$_lang["docid_incrmnt_method_1"] = '欠番の最小値を埋める';
-$_lang["docid_incrmnt_method_2"] = 'リソースIDの最大値として追加する';
 
 $_lang["enable_cache_title"] = 'リソースのキャッシュ';
 $_lang["disable_chunk_cache_title"] = 'チャンクのキャッシュを無効';
