@@ -37,6 +37,7 @@ $_lang["access_permissions_off"] = '<span class="warning">Toegangsbeheer is niet
 $_lang["access_permissions_resource_groups"] = 'Paginagroepen';
 $_lang["access_permissions_resources_in_group"] = '<b>Pagina\'s in deze groep:</b> ';
 $_lang["access_permissions_resources_tab"] = 'Bekijk hier de Paginagroepen. U kunt ook nieuwe groepen maken, bestaande bewerken of verwijderen of nagaan welke Pagina deel uitmaakt van een groep. Door met de muis over de ID van de Pagina te bewegen, ziet u de naam ervan. Door de Pagina te bewerken kunt u de groepsrechten wijzigen.';
+$_lang["access_permissions_user_toggle"] = 'Toggle access permissions';
 $_lang["access_permissions_user_groups"] = 'Gebruikersgroepen';
 $_lang["access_permissions_user_message"] = 'Geef aan tot welke groepen deze Gebruiker behoort:';
 $_lang["access_permissions_users_in_group"] = 'Gebruikers in deze groep:';
@@ -84,10 +85,11 @@ $_lang["captcha_title"] = 'CAPTCHA codes gebruiken';
 $_lang["captcha_words_default"] = 'EVO,Access,Better,BitCode,Chunk,Cache,Desc,Design,Excell,Enjoy,URLs,TechView,Gerald,Griff,Humphrey,Holiday,Intel,Integration,Joystick,Join,Oscope,Genetic,Light,Likeness,Marit,Maaike,Niche,Netherlands,Ordinance,Oscillo,Parser,Phusion,Query,Question,Regalia,Righteous,Snippet,Sentinel,Template,Thespian,Unity,Enterprise,Verily,Tattoo,Veri,Website,WideWeb,Yap,Yellow,Zebra,Zygote';
 $_lang["captcha_words_message"] = 'Vul hier een lijst van woorden in, van elkaar gescheiden door komma\'s, die gebruikt zal worden indien CAPTCHA is ingeschakeld. Het veld kan maximaal 255 tekens bevatten.';
 $_lang["captcha_words_title"] = 'CAPTCHA woorden';
+$_lang["category"] = 'Category';
+$_lang["categories"] = 'Categories';
 $_lang["category_heading"] = 'Categorie';
 $_lang["category_manager"] = 'Categorie Manager';
 $_lang["category_management"] = 'Categorie beheren';
-$_lang["manage_categories"] = 'Beheer categorieën';
 $_lang["category_msg"] = 'Bekijk en wijzig hier alle Elementen gegroepeerd per categorie.';
 $_lang["cfg_base_path"] = 'MODX_BASE_PATH';
 $_lang["cfg_base_url"] = 'MODX_BASE_URL';
@@ -322,6 +324,7 @@ $_lang["file_unzip"] = 'Uitpakken is gelukt!';
 $_lang["file_unzip_fail"] = 'Uitpakken is mislukt!';
 $_lang["filemanager_path_message"] = 'In IIS wordt de instelling document root vaak niet correct ingevuld. Deze wordt door de bestandsbeheerder gebruikt om te bepalen wat u kunt zien. Als u problemen ondervindt bij het gebruik van Bestandsbeheer, zorg er dan voor dat dit pad naar de hoofdmap van uw EVO installatie wijst.';
 $_lang["filemanager_path_title"] = 'Pad naar de bestanden';
+$_lang["files"] = 'Files';
 $_lang["files_access_denied"] = 'Toegang geweigerd!';
 $_lang["files_data"] = 'Gegevens';
 $_lang["files_dir_listing"] = 'Bestandenlijst voor:';
@@ -476,16 +479,17 @@ $_lang["login_message"] = 'Om Evolution te gebruiken dient u zich eerst aan te m
 $_lang["logo_slogan"] = 'EVO Content Manager - \nCreate and do more with less';
 $_lang["logout"] = 'Afmelden';
 $_lang["long_title"] = 'Uitgebreide titel';
+$_lang["manage_categories"] = 'Beheer categorieën';
 $_lang["manage_depends"] = 'Beheer afhankelijkheden';
 $_lang["manage_files"] = 'Bestandsbeheer';
-$_lang["manage_htmlsnippets"] = 'Chunks';
+$_lang["manage_htmlsnippets"] = 'Manage Chunks';
 $_lang["manage_metatags"] = 'Beheer META tags en keywords';
 $_lang["manage_modules"] = 'Modulebeheer';
-$_lang["manage_plugins"] = 'Plug-ins';
-$_lang["manage_snippets"] = 'Snippets';
-$_lang["manage_templates"] = 'Templates';
-$_lang["manage_documents"] = 'Documenten';
-$_lang["manage_permission"] = 'Toegangsbeheer';
+$_lang["manage_plugins"] = 'Manage Plugins';
+$_lang["manage_snippets"] = 'Manage Snippets';
+$_lang["manage_templates"] = 'Manage Templates';
+$_lang["manage_documents"] = 'Manage Documents';
+$_lang["manage_permission"] = 'Manage Permissions';
 $_lang["manager"] = 'Content Management Systeem';
 $_lang["manager_lockout_message"] = 'U bent momenteel aangemeld bij het Content Management Systeem. Als u uw sessie wilt sluiten, klik dan a.u.b. op de knop "Afmelden". <p />Wilt u naar uw start- of homepage klik dan op de knop "Start".';
 $_lang["manager_permissions"] = 'Toegangsbeheer';
@@ -1031,6 +1035,7 @@ $_lang["template_assignedtv_tab"] = 'Toegekende Template Variabelen';
 $_lang["template_code"] = 'Template code (html)';
 $_lang["template_desc"] = 'Beschrijving';
 $_lang["template_edit_tab"] = 'Wijzig Template';
+$_lang["template_inuse"] = 'This template is in use. Please set the documents using the template to another template. Documents using this template:';
 $_lang["template_management_msg"] = 'Hier kunt u Templates toevoegen en bewerken.';
 $_lang["template_msg"] = 'Template toevoegen of bewerken. Gewijzigde/nieuwe Templates zullen niet zichtbaar zijn in de gecachte Pagina\'s van uw website totdat de cache is leeggemaakt. U kunt echter de voorbeeldfunctie op een Pagina gebruiken om de Template te bekijken.';
 $_lang["template_name"] = 'Naam Template';
@@ -1283,10 +1288,6 @@ $_lang["setting_resource_tree_node_name_desc_add"] = 'Opmerking: U kunt deze wee
 $_lang["resource_opt_alvisibled"] = 'Gebruik huidige alias als alias pad';
 $_lang["resource_opt_alvisibled_help"] = 'The alias of this Resource is inserted in Friendly URL alias path';
 $_lang['resource_opt_is_published'] = 'Gepubliceerd';
-$_lang["docid_incrmnt_method_title"] = 'Optel methode pagina ID';
-$_lang["docid_incrmnt_method_0"] = 'DB auto increment';
-$_lang["docid_incrmnt_method_1"] = 'Minimaal aantal gemiste ID';
-$_lang["docid_incrmnt_method_2"] = 'Maximaal ID+1';
 
 $_lang["enable_cache_title"] = 'Document caching';
 $_lang["disable_chunk_cache_title"] = 'Chunk caching uitzetten';
