@@ -37,6 +37,7 @@ $_lang["access_permissions_off"] = '<span class="warning">Tilgangsinnstillinger 
 $_lang["access_permissions_resource_groups"] = 'Dokumentgrupper';
 $_lang["access_permissions_resources_in_group"] = '<b>Dokument i gruppen:</b> ';
 $_lang["access_permissions_resources_tab"] = 'Her kan du se hvilke dokumentgrupper som har blitt konfigurert. Du kan til og med lage nye og bytte navn på grupper, samt se hvilke dokumenter som tilhører en spesifikk gruppe (hold musepekeren over dokumentets ID for å se navnet på dokumentet). For å legge til eller fjerne et dokument fra en gruppe, rediger dokumentet direkte.';
+$_lang["access_permissions_user_toggle"] = 'Toggle access permissions';
 $_lang["access_permissions_user_groups"] = 'Brukergrupper';
 $_lang["access_permissions_user_message"] = 'Her kan du velge hvilke brukergrupper denne brukeren skal tilhøre:';
 $_lang["access_permissions_users_in_group"] = '<b>Brukere i gruppen:</b> ';
@@ -84,10 +85,11 @@ $_lang["captcha_title"] = 'Bruk CAPTCHA-koder';
 $_lang["captcha_words_default"] = 'MODX,Access,Better,BitCode,Chunk,Cache,Desc,Design,Excell,Enjoy,URLs,TechView,Gerald,Griff,Humphrey,Holiday,Intel,Integration,Joystick,Join(),Oscope,Genetic,Light,Likeness,Marit,Maaike,Niche,Netherlands,Ordinance,Oscillo,Parser,Phusion,Query,Question,Regalia,Righteous,Snippet,Sentinel,Template,Thespian,Unity,Enterprise,Verily,Tattoo,Veri,Website,WideWeb,Yap,Yellow,Zebra,Zygote';
 $_lang["captcha_words_message"] = 'Her kan du skrive en liste av CAPTCHA-ord som brukes hvis CAPTCHA er slått på. Separer ordene med et kommategn. Dette tekstfeltet er begrenset til 255 tegn.';
 $_lang["captcha_words_title"] = 'CAPTCHA-ord';
+$_lang["category"] = 'Category';
+$_lang["categories"] = 'Categories';
 $_lang["category_heading"] = 'Kategori';
 $_lang["category_manager"] = 'Category Manager';
 $_lang["category_management"] = 'Category management';
-$_lang["manage_categories"] = 'Manage Categories';
 $_lang["category_msg"] = 'Her kan du vise og redigere alle ressurser etter kategori.';
 $_lang["cfg_base_path"] = 'MODX_BASE_PATH';
 $_lang["cfg_base_url"] = 'MODX_BASE_URL';
@@ -139,7 +141,7 @@ $_lang["configcheck_notok"] = 'En eller flere konfigurasjonsdetaljer er ikke OK:
 $_lang["configcheck_ok"] = 'Kontroll utført OK - Ingen advarsler å rapportere.';
 $_lang["configcheck_php_gdzip"] = 'GD and/or Zip PHP extensions not found';
 $_lang["configcheck_php_gdzip_msg"] = 'Evolution CMS needs the GD and Zip extension enabled for PHP. While Evolution CMS will work without them, you will not be able to take full advantage of the built-in File Manager, Image Editor or Captcha for logins.';
-$_lang["configcheck_rb_base_dir"] = 'The currently set <a href="index.php?a=17&tab=5">File Browser path</a> seems incorrect.';
+$_lang["configcheck_rb_base_dir"] = 'The currently set <a href="index.php?a=17&tab=5">File base path</a> seems incorrect.';
 $_lang["configcheck_rb_base_dir_msg"] = 'This can happen for example by moving your installation to a different directory or server. Please check and update your Evolution CMS configuration.';
 $_lang["configcheck_register_globals"] = 'register_globals is set to ON in your php.ini configuration file';
 $_lang["configcheck_register_globals_msg"] = 'This configuration makes your site much more susceptible to Cross Site Scripting (XSS) attacks. You should speak to your host about what you can do to disable this setting.';
@@ -322,6 +324,7 @@ $_lang["file_unzip"] = 'Oppakking ferdig!';
 $_lang["file_unzip_fail"] = 'Oppakkingen mislyktes!';
 $_lang["filemanager_path_message"] = 'IIS fyller oftest ikke inn innstillingen document_root ordentlig. Denne brukes av filhåndtereren for å bestemme hva du får se. Hvis du har problemer med filhåndtereren, pass på at denne innstillingen peker til roten av din innstallasjon av EVO.';
 $_lang["filemanager_path_title"] = 'Søkesti til filhåndtereren';
+$_lang["files"] = 'Files';
 $_lang["files_access_denied"] = 'Tilgang nektet!';
 $_lang["files_data"] = 'Data';
 $_lang["files_dir_listing"] = 'Mappeliste for:';
@@ -476,16 +479,17 @@ $_lang["login_message"] = 'Fyll inn innloggningsinformasjon for å starte din se
 $_lang["logo_slogan"] = 'Lag og gjør mer med mindre - EVO innholdshåndterer';
 $_lang["logout"] = 'Logg ut';
 $_lang["long_title"] = 'Lang tittel';
+$_lang["manage_categories"] = 'Manage Categories';
 $_lang["manage_depends"] = 'Håndtere avhengigheter';
 $_lang["manage_files"] = 'Håndtere filer';
-$_lang["manage_htmlsnippets"] = 'Htmlstykker';
+$_lang["manage_htmlsnippets"] = 'Manage Chunks';
 $_lang["manage_metatags"] = 'Håndtere META-tagger og nøkkelord';
 $_lang["manage_modules"] = 'Håndtere moduler';
-$_lang["manage_plugins"] = 'Plugins';
-$_lang["manage_snippets"] = 'Kodesnutter';
-$_lang["manage_templates"] = 'Maler';
-$_lang["manage_documents"] = 'Documents';
-$_lang["manage_permission"] = 'Permissions';
+$_lang["manage_plugins"] = 'Manage Plugins';
+$_lang["manage_snippets"] = 'Manage Snippets';
+$_lang["manage_templates"] = 'Manage Templates';
+$_lang["manage_documents"] = 'Manage Documents';
+$_lang["manage_permission"] = 'Manage Permissions';
 $_lang["manager"] = 'Håndterere';
 $_lang["manager_lockout_message"] = 'You are currently logged into the Content Manager. If you would like to close your login session please click the "Logout" button. <p />To go to your startup or home page click the "Home" button.';
 $_lang["manager_permissions"] = 'Rettigheter for håndtereren';
@@ -995,6 +999,7 @@ $_lang["template_assignedtv_tab"] = 'Assigned Template Variables';
 $_lang["template_code"] = 'Mal-kode (html)';
 $_lang["template_desc"] = 'Beskrivelse';
 $_lang["template_edit_tab"] = 'Edit Template';
+$_lang["template_inuse"] = 'This template is in use. Please set the documents using the template to another template. Documents using this template:';
 $_lang["template_management_msg"] = 'Her kan du opprette en ny mal eller velge en allerede eksisterende mal for redigering.';
 $_lang["template_msg"] = 'Her kan du opprette og redigere maler. Endrede og nye maler kommer ikke på cachede sider før cachen er tømt. Du kan derimot bruke forhåndsvisningen for å se hvordan den oppdaterte malen kommer til å se ut.';
 $_lang["template_name"] = 'Navn på mal';
@@ -1247,10 +1252,6 @@ $_lang["setting_resource_tree_node_name_desc_add"] = 'Note: Since Evolution CMS 
 $_lang["resource_opt_alvisibled"] = 'Use current alias in alias path';
 $_lang["resource_opt_alvisibled_help"] = 'The alias of this Resource is inserted in Friendly URL alias path';
 $_lang['resource_opt_is_published'] = 'Published';
-$_lang["docid_incrmnt_method_title"] = 'Increment Resource ID method';
-$_lang["docid_incrmnt_method_0"] = 'DB auto increment';
-$_lang["docid_incrmnt_method_1"] = 'Minimum missed ID';
-$_lang["docid_incrmnt_method_2"] = 'Maximal ID + 1';
 
 $_lang["enable_cache_title"] = 'Document caching';
 $_lang["disable_chunk_cache_title"] = 'Disable chunk caching';
