@@ -37,6 +37,7 @@ $_lang["access_permissions_off"] = '<span class=\'warning\'>Uprawnienia Dostępu
 $_lang["access_permissions_resource_groups"] = 'Grupy dokumentów';
 $_lang["access_permissions_resources_in_group"] = '<b>Dokumenty w grupie:</b> ';
 $_lang["access_permissions_resources_tab"] = 'Tutaj możesz zobaczyć jakie grupy dokumentów zostały stworzone. Możesz także stworzyć nowe grupy, zmieniać nazwy, usuwać i sprawdzać które dokumenty należą do danej grupy (najedź kursorem na ID dokumentu, aby zobaczyć jego nazwę). Aby dodać dokument do grupy albo usunąć go z grupy, edytuj dokument bezpośrednio.';
+$_lang["access_permissions_user_toggle"] = 'Toggle access permissions';
 $_lang["access_permissions_user_groups"] = 'Grupy użytkowników';
 $_lang["access_permissions_user_message"] = 'Wybierz do których grup należy ten użytkownik:';
 $_lang["access_permissions_users_in_group"] = '<b>Użytkownicy w grupie:</b> ';
@@ -84,10 +85,11 @@ $_lang["captcha_title"] = 'Używaj kodów CAPTCHA';
 $_lang["captcha_words_default"] = 'EVO,Access,Better,BitCode,Chunk,Cache,Desc,Design,Excell,Enjoy,URLs,TechView,Gerald,Griff,Humphrey,Holiday,Intel,Integration,Joystick,Join(),Oscope,Genetic,Light,Likeness,Marit,Maaike,Niche,Netherlands,Ordinance,Oscillo,Parser,Phusion,Query,Question,Regalia,Righteous,Snippet,Sentinel,Template,Thespian,Unity,Enterprise,Verily,Tattoo,Veri,Website,WideWeb,Yap,Yellow,Zebra,Zygote';
 $_lang["captcha_words_message"] = 'Tutaj możesz wprowadzić listę słów CAPTCHA do użycia jeśli system CAPTCHA jest włączony. Oddziel słowa przecinkiem. Możesz wprowadzić maksymalnie 255 znaków.';
 $_lang["captcha_words_title"] = 'Słowa CAPTCHA';
+$_lang["category"] = 'Category';
+$_lang["categories"] = 'Categories';
 $_lang["category_heading"] = 'Kategoria';
 $_lang["category_manager"] = 'Menadżer kategorii';
 $_lang["category_management"] = 'Zarządzanie kategoriami';
-$_lang["manage_categories"] = 'Zarządzanie kategoriami';
 $_lang["category_msg"] = 'Przeglądaj i edytuj wszystkie elementy pogrupowane w kategorie.';
 $_lang["cfg_base_path"] = 'MODX_BASE_PATH';
 $_lang["cfg_base_url"] = 'MODX_BASE_URL';
@@ -322,6 +324,7 @@ $_lang["file_unzip"] = 'Rozpakowanie zakończone sukcesem!';
 $_lang["file_unzip_fail"] = 'Błąd podczas rozpakowywania!';
 $_lang["filemanager_path_message"] = 'IIS często błędnie określa ustawienie document_root, które jest używane przez menedżer plików do określenia co może być widoczne dla ciebie. Jeśli masz problemy z menadżerem plików, upewnij się, że ta ścieżka wskazuje na katalog główny instalacji EVO.';
 $_lang["filemanager_path_title"] = 'Ścieżka menadżera plików';
+$_lang["files"] = 'Files';
 $_lang["files_access_denied"] = 'Brak dostępu!';
 $_lang["files_data"] = 'Dane';
 $_lang["files_dir_listing"] = 'Lista katalogu dla:';
@@ -476,16 +479,17 @@ $_lang["login_message"] = 'Proszę podać swoje dane uwierzytelniające, aby si�
 $_lang["logo_slogan"] = 'EVO System Zarządzania Treścią - \nTwórz więcej robiąc mniej';
 $_lang["logout"] = 'Wyloguj się';
 $_lang["long_title"] = 'Długi tytuł';
+$_lang["manage_categories"] = 'Zarządzanie kategoriami';
 $_lang["manage_depends"] = 'Menedżer zależności';
 $_lang["manage_files"] = 'Zarządzanie plikami';
-$_lang["manage_htmlsnippets"] = 'Chunki';
+$_lang["manage_htmlsnippets"] = 'Manage Chunks';
 $_lang["manage_metatags"] = 'Zarządzanie META tagami i słowami kluczowymi';
 $_lang["manage_modules"] = 'Zarządzanie modułami';
-$_lang["manage_plugins"] = 'Pluginy';
-$_lang["manage_snippets"] = 'Snippety';
-$_lang["manage_templates"] = 'Szablony';
-$_lang["manage_documents"] = 'Dokumenty';
-$_lang["manage_permission"] = 'Permissions';
+$_lang["manage_plugins"] = 'Manage Plugins';
+$_lang["manage_snippets"] = 'Manage Snippets';
+$_lang["manage_templates"] = 'Manage Templates';
+$_lang["manage_documents"] = 'Manage Documents';
+$_lang["manage_permission"] = 'Manage Permissions';
 $_lang["manager"] = 'Menedżer';
 $_lang["manager_lockout_message"] = 'Jesteś obecnie zalogowany do Menadżera treści. W celu zakończenia sesji kliknij przycisk "Wyloguj". Aby przejść do strony głównej, kliknij przycisk "Start".';
 $_lang["manager_permissions"] = 'Uprawnienia Menedżera';
@@ -1031,6 +1035,7 @@ $_lang["template_assignedtv_tab"] = 'Zmienne szablonu';
 $_lang["template_code"] = 'Kod szablonu (html)';
 $_lang["template_desc"] = 'Opis';
 $_lang["template_edit_tab"] = 'Edytuj szablon';
+$_lang["template_inuse"] = 'This template is in use. Please set the documents using the template to another template. Documents using this template:';
 $_lang["template_management_msg"] = 'Tutaj możesz wybrać, które szablony chcesz edytować.';
 $_lang["template_msg"] = 'Tutaj możesz tworzyć/ edytować szablony. Zmienione/ nowe szablony nie będą widoczne na zcacheowanych stronach Twojej witryny dopóki cache nie zostanie wyczyszczony, jednakże, możesz użyć funkcji podglądu na stronie, aby zobaczyć szablon w działaniu.';
 $_lang["template_name"] = 'Nazwa szablonu';
@@ -1283,10 +1288,6 @@ $_lang["setting_resource_tree_node_name_desc_add"] = 'Informacja: Od wydania EVO
 $_lang["resource_opt_alvisibled"] = 'Uwzględnij alias w URL';
 $_lang["resource_opt_alvisibled_help"] = 'Alias tego zasobu zostanie dołączony do ścieżki przyjaznego URL';
 $_lang['resource_opt_is_published'] = 'Published';
-$_lang["docid_incrmnt_method_title"] = 'Metoda przyrostu ID zasobu';
-$_lang["docid_incrmnt_method_0"] = 'DB auto increment';
-$_lang["docid_incrmnt_method_1"] = 'Minimum missed ID';
-$_lang["docid_incrmnt_method_2"] = 'Maximal ID+1';
 
 $_lang["enable_cache_title"] = 'Cache dokumentu';
 $_lang["disable_chunk_cache_title"] = 'Disable chunk caching';
