@@ -37,6 +37,7 @@ $_lang["access_permissions_off"] = '<span class=\'warning\'>As permissões de ac
 $_lang["access_permissions_resource_groups"] = 'Grupos de documentos';
 $_lang["access_permissions_resources_in_group"] = '<b>Documentos no grupo:</b> ';
 $_lang["access_permissions_resources_tab"] = 'Aqui pode ver que grupos de documentos estão definidos. Pode criar novos grupos, alterar o nome dos grupos, apagar grupos e ver que documentos pertencem aos diferentes grupos. Para adicionar ou remover um documento de um grupo, edite-o directamente.';
+$_lang["access_permissions_user_toggle"] = 'Toggle access permissions';
 $_lang["access_permissions_user_groups"] = 'Grupos de utilizadores';
 $_lang["access_permissions_user_message"] = 'Aqui pode seleccionar a que grupos este utilizador pertence.';
 $_lang["access_permissions_users_in_group"] = '<b>Utilizadores no grupo:</b> ';
@@ -84,10 +85,11 @@ $_lang["captcha_title"] = 'Usar códigos de segurança (CAPTCHA):';
 $_lang["captcha_words_default"] = 'MODX,Access,Better,BitCode,Chunk,Cache,Desc,Design,Excell,Enjoy,URLs,TechView,Gerald,Griff,Humphrey,Holiday,Intel,Integration,Joystick,Join(),Oscope,Genetic,Light,Likeness,Marit,Maaike,Niche,Netherlands,Ordinance,Oscillo,Parser,Phusion,Query,Question,Regalia,Righteous,Snippet,Sentinel,Template,Thespian,Unity,Enterprise,Verily,Tattoo,Veri,Website,WideWeb,Yap,Yellow,Zebra,Zygote';
 $_lang["captcha_words_message"] = 'Aqui pode introduzir uma lista de palavras a criptografar em imagens (CAPTCHA) quando este recurso estiver activado. Separe-as com vírgulas. Este campo está limitado a 255 caracteres.';
 $_lang["captcha_words_title"] = 'Palavras encriptadas (CAPTCHA)';
+$_lang["category"] = 'Category';
+$_lang["categories"] = 'Categories';
 $_lang["category_heading"] = 'Categoria';
 $_lang["category_manager"] = 'Category Manager';
 $_lang["category_management"] = 'Category management';
-$_lang["manage_categories"] = 'Manage Categories';
 $_lang["category_msg"] = 'Aqui pode ver e editar todos os recursos agrupados por categoria.';
 $_lang["cfg_base_path"] = 'MODX_BASE_PATH';
 $_lang["cfg_base_url"] = 'MODX_BASE_URL';
@@ -139,7 +141,7 @@ $_lang["configcheck_notok"] = 'Um ou mais pontos de configuração não foram va
 $_lang["configcheck_ok"] = 'Aprovados com sucesso (OK) - nada a assinalar.';
 $_lang["configcheck_php_gdzip"] = 'GD and/or Zip PHP extensions not found';
 $_lang["configcheck_php_gdzip_msg"] = 'EVO needs the GD and Zip extension enabled for PHP. While EVO will work without them, you will not be able to take full advantage of the built-in File Manager, Image Editor or Captcha for logins.';
-$_lang["configcheck_rb_base_dir"] = 'The currently set <a href="index.php?a=17&tab=5">File Browser path</a> seems incorrect.';
+$_lang["configcheck_rb_base_dir"] = 'The currently set <a href="index.php?a=17&tab=5">File base path</a> seems incorrect.';
 $_lang["configcheck_rb_base_dir_msg"] = 'This can happen for example by moving your installation to a different directory or server. Please check and update your Evolution CMS configuration.';
 $_lang["configcheck_register_globals"] = 'register_globals está definido como ON no ficheiro de configuração php.ini';
 $_lang["configcheck_register_globals_msg"] = 'Esta configuração torna o site muito mais suspceptível a ataques Cross Site Scripting (XSS). Aconselha-se que contacte o seu provedor de hosting para saber o que pode ser feito para desactivar esta opção.';
@@ -322,6 +324,7 @@ $_lang["file_unzip"] = 'Descompactação realizada com sucesso!';
 $_lang["file_unzip_fail"] = 'A descompactação falhou!';
 $_lang["filemanager_path_message"] = 'Frequentemente, o IIS não interpreta correctamente a opção document_root utilizada pelo Gestor de Ficheiros para determinar o que o utilizador pode ver. Se tiver problemas ao usar o Gestor de Ficheiros, certifique-se de que este caminho aponta para a raiz da instalação do sistema EVO.';
 $_lang["filemanager_path_title"] = 'Caminho do Gestor de Ficheiros:';
+$_lang["files"] = 'Files';
 $_lang["files_access_denied"] = 'Acesso negado!';
 $_lang["files_data"] = 'Dados';
 $_lang["files_dir_listing"] = 'Listagem do directório:';
@@ -476,16 +479,17 @@ $_lang["login_message"] = 'Por favor digite o seu username e senha para aceder a
 $_lang["logo_slogan"] = 'Gestor de Conteúdos EVO - \nCrie e faça mais com menos';
 $_lang["logout"] = 'Sair';
 $_lang["long_title"] = 'Título longo';
+$_lang["manage_categories"] = 'Manage Categories';
 $_lang["manage_depends"] = 'Gerir dependências';
 $_lang["manage_files"] = 'Gerir Ficheiros';
-$_lang["manage_htmlsnippets"] = 'Chunks';
+$_lang["manage_htmlsnippets"] = 'Manage Chunks';
 $_lang["manage_metatags"] = 'Gerir Etiquetas (tags) META e Palavras-chave';
 $_lang["manage_modules"] = 'Gerir Módulos';
-$_lang["manage_plugins"] = 'Plugins';
-$_lang["manage_snippets"] = 'Snippets';
-$_lang["manage_templates"] = 'Templates';
-$_lang["manage_documents"] = 'Documents';
-$_lang["manage_permission"] = 'Permissions';
+$_lang["manage_plugins"] = 'Manage Plugins';
+$_lang["manage_snippets"] = 'Manage Snippets';
+$_lang["manage_templates"] = 'Manage Templates';
+$_lang["manage_documents"] = 'Manage Documents';
+$_lang["manage_permission"] = 'Manage Permissions';
 $_lang["manager"] = 'Administrador';
 $_lang["manager_lockout_message"] = 'Está actualmente ligado no Gestor de Conteúdos. Se deseja fechar a sua sessão, por favor clique no botão \'Sair\'. <br />Para ir para a sua página inicial clique no botão \'Início\'';
 $_lang["manager_permissions"] = 'Permissões (no Interface de Gestão)';
@@ -1028,6 +1032,7 @@ $_lang["template_assignedtv_tab"] = 'Variáveis de Template atribuídas';
 $_lang["template_code"] = 'Código HTML do Template';
 $_lang["template_desc"] = 'Descrição';
 $_lang["template_edit_tab"] = 'Editar Template';
+$_lang["template_inuse"] = 'This template is in use. Please set the documents using the template to another template. Documents using this template:';
 $_lang["template_management_msg"] = 'Aqui pode escolher que Template deseja editar.';
 $_lang["template_msg"] = 'Criar e editar Templates. Templates novos ou modificados não serão visíveis nas páginas em cache até que esta seja esvaziada, no entanto pode usar a função de pré-visualização na página para ver o Template em acção.';
 $_lang["template_name"] = 'Nome do Template';
@@ -1280,10 +1285,6 @@ $_lang["setting_resource_tree_node_name_desc_add"] = 'Note: Since Evolution CMS 
 $_lang["resource_opt_alvisibled"] = 'Use current alias in alias path';
 $_lang["resource_opt_alvisibled_help"] = 'The alias of this Resource is inserted in Friendly URL alias path';
 $_lang['resource_opt_is_published'] = 'Publicado';
-$_lang["docid_incrmnt_method_title"] = 'Increment Resource ID method';
-$_lang["docid_incrmnt_method_0"] = 'DB auto incremento';
-$_lang["docid_incrmnt_method_1"] = 'Minima perda de ID';
-$_lang["docid_incrmnt_method_2"] = 'Maximal ID + 1';
 
 $_lang["enable_cache_title"] = 'Document caching';
 $_lang["disable_chunk_cache_title"] = 'Disable chunk caching';
