@@ -37,6 +37,7 @@ $_lang["access_permissions_off"] = '<span class="warning">Права досту�
 $_lang["access_permissions_resource_groups"] = 'Группы ресурсов';
 $_lang["access_permissions_resources_in_group"] = '<b>ресурсов в группе:</b> ';
 $_lang["access_permissions_resources_tab"] = 'Здесь вы можете увидеть все группы ресурсов. Также, здесь вы можете создать, переименовать, удалить группу, просмотреть ресурсы в группе. Для добавления или удаления ресурса из группы редактируйте ресурс непосредственно.';
+$_lang["access_permissions_user_toggle"] = 'Изменить все разрешения';
 $_lang["access_permissions_user_groups"] = 'Группы пользователей';
 $_lang["access_permissions_user_message"] = 'Здесь вы можете выбрать, к какой пользовательской группе принадлежит этот пользователь:';
 $_lang["access_permissions_users_in_group"] = 'Пользователей в группе:';
@@ -84,10 +85,11 @@ $_lang["captcha_title"] = 'Использовать код CAPTCHA:';
 $_lang["captcha_words_default"] = 'EVO,Access,Better,BitCode,Chunk,Cache,Desc,Design,Excell,Enjoy,URLs,TechView,Gerald,Griff,Humphrey,Holiday,Intel,Integration,Joystick,Join(),Oscope,Genetic,Light,Likeness,Marit,Maaike,Niche,Netherlands,Ordinance,Oscillo,Parser,Phusion,Query,Question,Regalia,Righteous,Snippet,Sentinel,Template,Thespian,Unity,Enterprise,Verily,Tattoo,Veri,Website,WideWeb,Yap,Yellow,Zebra,Zygote';
 $_lang["captcha_words_message"] = 'Здесь вы можете задать список слов для генерирования кода CAPTCHA. Слова в списке должны разделяться запятыми.';
 $_lang["captcha_words_title"] = 'Слова для генерации CAPTCHA-кодов:';
+$_lang["category"] = 'Категория';
+$_lang["categories"] = 'Категории';
 $_lang["category_heading"] = 'Категория';
 $_lang["category_manager"] = 'Менеджер категорий';
 $_lang["category_management"] = 'Категории';
-$_lang["manage_categories"] = 'Категории';
 $_lang["category_msg"] = 'Здесь вы можете просматривать и редактировать все элементы, сгруппированные по категориям.';
 $_lang["cfg_base_path"] = 'MODX_BASE_PATH';
 $_lang["cfg_base_url"] = 'MODX_BASE_URL';
@@ -322,6 +324,7 @@ $_lang["file_unzip"] = 'Архив успешно распакован';
 $_lang["file_unzip_fail"] = 'Архив не удалось распаковать';
 $_lang["filemanager_path_message"] = 'Часто IIS неправильно выводит параметр document_root, который используется файл-менеджером. Установите правильный путь во избежание проблем с файл-менеджером.';
 $_lang["filemanager_path_title"] = 'Путь для файл-менеджера:';
+$_lang["files"] = 'Файлы';
 $_lang["files_access_denied"] = 'Доступ запрещен';
 $_lang["files_data"] = 'Суммарный объём';
 $_lang["files_dir_listing"] = 'Список файлов в папке: ';
@@ -476,16 +479,17 @@ $_lang["login_message"] = 'Введите ваше имя пользовател
 $_lang["logo_slogan"] = 'Создавайте больше с меньшими усилиями - \nСистема управления сайтом Evolution CMS';
 $_lang["logout"] = 'Выйти';
 $_lang["long_title"] = 'Расширенный заголовок';
+$_lang["manage_categories"] = 'Категории';
 $_lang["manage_depends"] = 'Зависимости';
 $_lang["manage_files"] = 'Все файлы';
-$_lang["manage_htmlsnippets"] = 'Чанки';
+$_lang["manage_htmlsnippets"] = 'Управление чанками';
 $_lang["manage_metatags"] = 'META-теги и ключевые слова';
 $_lang["manage_modules"] = 'Модули';
-$_lang["manage_plugins"] = 'Плагины';
-$_lang["manage_snippets"] = 'Сниппеты';
-$_lang["manage_templates"] = 'Шаблоны';
-$_lang["manage_documents"] = 'Документы';
-$_lang["manage_permission"] = 'Права доступа';
+$_lang["manage_plugins"] = 'Управление плагинами';
+$_lang["manage_snippets"] = 'Управление сниппетами';
+$_lang["manage_templates"] = 'Управление шаблонами';
+$_lang["manage_documents"] = 'Управление документами';
+$_lang["manage_permission"] = 'Управление разрешениями';
 $_lang["manager"] = 'Менеджер';
 $_lang["manager_lockout_message"] = 'В данный момент вы находитесь в режиме управления сайтом. Чтобы закончить сеанс администрирования, нажмите кнопку \'Выйти\'. <p />Чтобы перейти на главную или стартовую страницу, нажмите кнопку \'Домой\'.';
 $_lang["manager_permissions"] = 'Права менеджеров';
@@ -1023,6 +1027,7 @@ $_lang["template_assignedtv_tab"] = 'Назначенные параметры (
 $_lang["template_code"] = 'Код шаблона (HTML)';
 $_lang["template_desc"] = 'Описание';
 $_lang["template_edit_tab"] = 'Редактировать шаблон';
+$_lang["template_inuse"] = 'Этот шаблон используется. Пожалуйста, выберите другой шаблон для следующих документов:';
 $_lang["template_management_msg"] = '<b>Шаблоны</b> - это (X)HTML-разметка страниц сайта. Шаблон описывает структуру и дизайн отображения элементов страницы (статичных и генерируемых). Шаблон может содержать вызовы <i>сниппетов</i>, <i>чанков</i>, <i>параметров (TV)</i>, а также ссылки на CSS-файлы и другие объекты, определяющие визуальное отображение страницы и её элементов.<p><b>Использование:</b> назначьте шаблон какому-либо ресурсу.</p><p><br />Выберите шаблон для редактирования или создайте новый.';
 $_lang["template_msg"] = 'Здесь вы можете создать / отредактировать шаблон.';
 $_lang["template_name"] = 'Имя шаблона';
@@ -1275,10 +1280,6 @@ $_lang["setting_resource_tree_node_name_desc_add"] = 'Внимание: Начи
 $_lang["resource_opt_alvisibled"] = 'Участвует в URL';
 $_lang["resource_opt_alvisibled_help"] = 'Для участия данного документа в адресной строке поставьте галку, и наоборот - снимите, если псевдоним этого документа нужно убрать из URL.';
 $_lang['resource_opt_is_published'] = 'Опубликован';
-$_lang["docid_incrmnt_method_title"] = 'Способ добавления Resource ID ';
-$_lang["docid_incrmnt_method_0"] = 'DB автоинкримент';
-$_lang["docid_incrmnt_method_1"] = 'Минимальный пропущенный ID';
-$_lang["docid_incrmnt_method_2"] = 'Максимальный ID+1';
 
 $_lang["enable_cache_title"] = 'Способ кэширования';
 $_lang["disable_chunk_cache_title"] = 'Отключить кэширование чанков';
