@@ -37,6 +37,7 @@ $_lang["access_permissions_off"] = '<span class="warning">Åtkomstinställningar
 $_lang["access_permissions_resource_groups"] = 'Resursgrupper';
 $_lang["access_permissions_resources_in_group"] = '<b>Resurser i gruppen:</b> ';
 $_lang["access_permissions_resources_tab"] = 'Se vilka resursgrupper som har skapats. Det går också att skapa nya eller byta namn på grupper samt se vilka resurser som tillhör de olika grupperna (håll musen över resursens ID för att se dess namn). För att lägga till eller ta bort en resurs från en grupp - redigera resursen direkt.';
+$_lang["access_permissions_user_toggle"] = 'Toggle access permissions';
 $_lang["access_permissions_user_groups"] = 'Användargrupper';
 $_lang["access_permissions_user_message"] = 'Ange vilka användargrupper denna användare ska tillhöra:';
 $_lang["access_permissions_users_in_group"] = 'Användare i gruppen:';
@@ -84,10 +85,11 @@ $_lang["captcha_title"] = 'Använd CAPTCHA-koder';
 $_lang["captcha_words_default"] = 'MODX,Access,Better,BitCode,Chunk,Cache,Desc,Design,Excell,Enjoy,URLs,TechView,Gerald,Griff,Humphrey,Holiday,Intel,Integration,Joystick,Join(),Oscope,Genetic,Light,Likeness,Marit,Maaike,Niche,Netherlands,Ordinance,Oscillo,Parser,Phusion,Query,Question,Regalia,Righteous,Snippet,Sentinel,Template,Thespian,Unity,Enterprise,Verily,Tattoo,Veri,Website,WideWeb,Yap,Yellow,Zebra,Zygote';
 $_lang["captcha_words_message"] = 'Skriv en lista med CAPTCHA-ord som används om CAPTCHA är påslaget. Separera orden med kommatecken. Detta textfält är begränsat till 255 tecken.';
 $_lang["captcha_words_title"] = 'CAPTCHA-ord';
+$_lang["category"] = 'Category';
+$_lang["categories"] = 'Categories';
 $_lang["category_heading"] = 'Kategori';
 $_lang["category_manager"] = 'Category Manager';
 $_lang["category_management"] = 'Category management';
-$_lang["manage_categories"] = 'Manage Categories';
 $_lang["category_msg"] = 'Visa och redigera alla element grupperade i kategorier.';
 $_lang["cfg_base_path"] = 'MODX_BASE_PATH';
 $_lang["cfg_base_url"] = 'MODX_BASE_URL';
@@ -139,7 +141,7 @@ $_lang["configcheck_notok"] = 'En eller flera konfigurationsdetaljer är inte ko
 $_lang["configcheck_ok"] = 'Kontrollen utförd OK - inga varningar att rapportera.';
 $_lang["configcheck_php_gdzip"] = 'GD och/eller Zip PHP-tillägg kunde inte hittas';
 $_lang["configcheck_php_gdzip_msg"] = 'EVO behöver tilläggen GD och Zip aktiverade för PHP. Även om EVO kommer att fungera utan dom, så kommer du inte att kunna utnyttja alla möjligheterna i den inbyggda filhanteraren, bildredigeraren eller captcha för inloggningar.';
-$_lang["configcheck_rb_base_dir"] = 'The currently set <a href="index.php?a=17&tab=5">File Browser path</a> seems incorrect.';
+$_lang["configcheck_rb_base_dir"] = 'The currently set <a href="index.php?a=17&tab=5">File base path</a> seems incorrect.';
 $_lang["configcheck_rb_base_dir_msg"] = 'This can happen for example by moving your installation to a different directory or server. Please check and update your Evolution CMS configuration.';
 $_lang["configcheck_register_globals"] = 'register_globals är satt till ON i din konfigurationsfil för php - php.ini';
 $_lang["configcheck_register_globals_msg"] = 'Denna konfiguration gör din webbplats betydligt mer sårbar för så kallade serveröverskridande scriptattacker (Cross Site Scripting eller XSS). Du bör ta kontakt med din webbhost och ta reda på vad du kan göra för att stänga av den här inställningen.';
@@ -322,6 +324,7 @@ $_lang["file_unzip"] = 'Uppzippningen klar!';
 $_lang["file_unzip_fail"] = 'Uppzippningen misslyckades!';
 $_lang["filemanager_path_message"] = 'IIS fyller oftast inte i inställningarna för document_root ordentligt, vilket används av filhanteraren för att bestämma vad du kan se. Om du har problem med filhanteraren, se till så att denna sökväg pekar till roten på din EVO-installation.';
 $_lang["filemanager_path_title"] = 'Sökväg till filhanteraren';
+$_lang["files"] = 'Files';
 $_lang["files_access_denied"] = 'Åtkomst nekad!';
 $_lang["files_data"] = 'Data';
 $_lang["files_dir_listing"] = 'Kataloglista för:';
@@ -476,16 +479,17 @@ $_lang["login_message"] = 'Fyll i dina inloggningsuppgifter för att starta din 
 $_lang["logo_slogan"] = 'Skapa och gör mer med mindre - \nEVO innehållshanterare';
 $_lang["logout"] = 'Logga ut';
 $_lang["long_title"] = 'Lång titel';
+$_lang["manage_categories"] = 'Manage Categories';
 $_lang["manage_depends"] = 'Hantera beroenden';
 $_lang["manage_files"] = 'Hantera filer';
-$_lang["manage_htmlsnippets"] = 'Chunks';
+$_lang["manage_htmlsnippets"] = 'Manage Chunks';
 $_lang["manage_metatags"] = 'Hantera META-taggar och nyckelord';
 $_lang["manage_modules"] = 'Hantera moduler';
-$_lang["manage_plugins"] = 'Plugins';
-$_lang["manage_snippets"] = 'Snippets';
-$_lang["manage_templates"] = 'Mallar';
-$_lang["manage_documents"] = 'Documents';
-$_lang["manage_permission"] = 'Permissions';
+$_lang["manage_plugins"] = 'Manage Plugins';
+$_lang["manage_snippets"] = 'Manage Snippets';
+$_lang["manage_templates"] = 'Manage Templates';
+$_lang["manage_documents"] = 'Manage Documents';
+$_lang["manage_permission"] = 'Manage Permissions';
 $_lang["manager"] = 'Hanterare';
 $_lang["manager_lockout_message"] = 'Du är för närvarande inloggad i innehållshanteraren. Klicka på "Logga ut" om du vill stänga din session. <p />Klicka på "Hem"-knappen om du vill gå till din start- eller hemsida.';
 $_lang["manager_permissions"] = 'Rättigheter för hanteraren';
@@ -1028,6 +1032,7 @@ $_lang["template_assignedtv_tab"] = 'Tilldelade mallvariabler';
 $_lang["template_code"] = 'Mall-kod (html)';
 $_lang["template_desc"] = 'Beskrivning';
 $_lang["template_edit_tab"] = 'Redigera mall';
+$_lang["template_inuse"] = 'This template is in use. Please set the documents using the template to another template. Documents using this template:';
 $_lang["template_management_msg"] = 'Skapa en ny mall eller välja en redan befintlig för redigering.';
 $_lang["template_msg"] = 'Skapa och redigera mallar här. Ändrade eller nya mallar kommer inte att synas på din webbplats cachade sidor, förrän cachen har tömts. Du kan däremot använda förhandsgranskningen för att se hur den uppdaterade mallen kommer att se ut.';
 $_lang["template_name"] = 'Mallnamn';
@@ -1280,10 +1285,6 @@ $_lang["setting_resource_tree_node_name_desc_add"] = 'Note: Since Evolution CMS 
 $_lang["resource_opt_alvisibled"] = 'Använd aktuellt alias i aliassökväg';
 $_lang["resource_opt_alvisibled_help"] = 'Denna resurs alias infogas i vänliga URL-aliassökvägar';
 $_lang['resource_opt_is_published'] = 'Publicerad';
-$_lang["docid_incrmnt_method_title"] = 'Uppräkningsmetod för resurs-ID';
-$_lang["docid_incrmnt_method_0"] = 'Databasens autouppräkning';
-$_lang["docid_incrmnt_method_1"] = 'Minsta lediga ID';
-$_lang["docid_incrmnt_method_2"] = 'Högsta ID + 1';
 
 $_lang["enable_cache_title"] = 'Document caching';
 $_lang["disable_chunk_cache_title"] = 'Disable chunk caching';
