@@ -37,6 +37,7 @@ $_lang["access_permissions_off"] = '<span class="warning">Права досту�
 $_lang["access_permissions_resource_groups"] = 'Групи ресурсів';
 $_lang["access_permissions_resources_in_group"] = '<b>ресурсів в групі:</b> ';
 $_lang["access_permissions_resources_tab"] = 'Тут ви можете побачити всі групи ресурсів. Також, тут ви можете створити, перейменувати, видалити групу, переглянути ресурси в групі. Для додавання або видалення ресурсу з групи перейдіть в редагування ресурсу.';
+$_lang["access_permissions_user_toggle"] = 'Toggle access permissions';
 $_lang["access_permissions_user_groups"] = 'Групи користувачів';
 $_lang["access_permissions_user_message"] = 'Тут ви можете вибрати, до якої групи користувачів належить цей користувач:';
 $_lang["access_permissions_users_in_group"] = 'Користувачів в групі:';
@@ -84,10 +85,11 @@ $_lang["captcha_title"] = 'Використовувати код CAPTCHA:';
 $_lang["captcha_words_default"] = 'EVO,Access,Better,BitCode,Chunk,Cache,Desc,Design,Excell,Enjoy,URLs,TechView,Gerald,Griff,Humphrey,Holiday,Intel,Integration,Joystick,Join(),Oscope,Genetic,Light,Likeness,Marit,Maaike,Niche,Netherlands,Ordinance,Oscillo,Parser,Phusion,Query,Question,Regalia,Righteous,Snippet,Sentinel,Template,Thespian,Unity,Enterprise,Verily,Tattoo,Veri,Website,WideWeb,Yap,Yellow,Zebra,Zygote';
 $_lang["captcha_words_message"] = 'Тут ви можете задати список слів для генерування коду CAPTCHA. Слова в списку повинні бути розділені комами.';
 $_lang["captcha_words_title"] = 'Слова для генерації CAPTCHA-кодів:';
+$_lang["category"] = 'Category';
+$_lang["categories"] = 'Categories';
 $_lang["category_heading"] = 'Категорія';
 $_lang["category_manager"] = 'Менеджер категорій';
 $_lang["category_management"] = 'Управління категоріями';
-$_lang["manage_categories"] = 'Управління категоріями';
 $_lang["category_msg"] = 'Тут ви можете переглядати і редагувати всі елементи, згруповані за категоріями.';
 $_lang["cfg_base_path"] = 'MODX_BASE_PATH';
 $_lang["cfg_base_url"] = 'MODX_BASE_URL';
@@ -322,6 +324,7 @@ $_lang["file_unzip"] = 'Архів успішно розпакований';
 $_lang["file_unzip_fail"] = 'Архів не вдалося розпакувати';
 $_lang["filemanager_path_message"] = 'Часто IIS неправильно виводить параметр document_root, який використовується файл-менеджером. Встановіть правильний шлях, щоб уникнути проблем з файл-менеджером.';
 $_lang["filemanager_path_title"] = 'Шлях для файл-менеджера:';
+$_lang["files"] = 'Files';
 $_lang["files_access_denied"] = 'Доступ заборонено';
 $_lang["files_data"] = 'Сумарний обсяг';
 $_lang["files_dir_listing"] = 'Список файлів в папці: ';
@@ -476,16 +479,17 @@ $_lang["login_message"] = 'Введіть ваше ім*я користувач�
 $_lang["logo_slogan"] = 'Створюйте більше з меншими зусиллями - \n Система управління сайтом EVO';
 $_lang["logout"] = 'Вийти';
 $_lang["long_title"] = 'Розширений заголовок';
+$_lang["manage_categories"] = 'Управління категоріями';
 $_lang["manage_depends"] = 'Керування залежностями';
 $_lang["manage_files"] = 'Керування файлами';
-$_lang["manage_htmlsnippets"] = 'Чанки';
+$_lang["manage_htmlsnippets"] = 'Manage Chunks';
 $_lang["manage_metatags"] = 'Керування META-тегами и ключовими словами';
 $_lang["manage_modules"] = 'Керування модулями';
-$_lang["manage_plugins"] = 'Плагіни';
-$_lang["manage_snippets"] = 'Сніппети';
-$_lang["manage_templates"] = 'Шаблони';
-$_lang["manage_documents"] = 'Документи';
-$_lang["manage_permission"] = 'Права доступу';
+$_lang["manage_plugins"] = 'Manage Plugins';
+$_lang["manage_snippets"] = 'Manage Snippets';
+$_lang["manage_templates"] = 'Manage Templates';
+$_lang["manage_documents"] = 'Manage Documents';
+$_lang["manage_permission"] = 'Manage Permissions';
 $_lang["manager"] = 'Менеджер';
 $_lang["manager_lockout_message"] = 'В даний момент ви знаходитесь в режимі управління сайтом. Щоб закінчити сеанс адміністрування, натисніть кнопку \'Вийти\'. <p />Щоб перейти на головну або стартову сторінку, натисніть кнопку \'Додому\'.';
 $_lang["manager_permissions"] = 'Права менеджерів';
@@ -1023,6 +1027,7 @@ $_lang["template_assignedtv_tab"] = 'Призначені параметри (TV
 $_lang["template_code"] = 'Код шаблона (HTML)';
 $_lang["template_desc"] = 'Опис';
 $_lang["template_edit_tab"] = 'Редагувати шаблон';
+$_lang["template_inuse"] = 'This template is in use. Please set the documents using the template to another template. Documents using this template:';
 $_lang["template_management_msg"] = '<b>Шаблони</b> - це (X)HTML-розмітка сторінок сайту. Шаблон визначає структуру і дизайн відображення елементів сторінки (статичних і генеруються). Шаблон може містити виклики <i>сніпетів</i>, <i>чанків</i>, <i>параметрів (TV)</i>,а також посилання на CSS-файли та інші об*єкти, що визначають візуальне відображення сторінки і її елементів.<p><b>Використання:</b> призначте шаблон будь-якого ресурсу.</p><p><br />Виберіть шаблон для редагування або створіть новий.';
 $_lang["template_msg"] = 'Тут ви можете створити / відредагувати шаблон.';
 $_lang["template_name"] = 'Ім*я шаблону';
@@ -1276,10 +1281,6 @@ $_lang["setting_resource_tree_node_name_desc_add"] = 'Увага: Починаю
 $_lang["resource_opt_alvisibled"] = 'Відображається в URL';
 $_lang["resource_opt_alvisibled_help"] = 'Для участі даного документа в адресному рядку треба поставити пташку, і навпаки - зніміть, якщо псевдонім цього документа потрібно прибрати з URL.';
 $_lang['resource_opt_is_published'] = 'Опублікований';
-$_lang["docid_incrmnt_method_title"] = 'Спосіб додавання Resource ID ';
-$_lang["docid_incrmnt_method_0"] = 'DB автоінкрімент';
-$_lang["docid_incrmnt_method_1"] = 'Мінімальний пропущений ID';
-$_lang["docid_incrmnt_method_2"] = 'Максимальний ID+1';
 
 $_lang["enable_cache_title"] = 'Спосіб кешування';
 $_lang["disable_chunk_cache_title"] = 'Заборонити кешування чанків';
