@@ -37,6 +37,7 @@ $_lang["access_permissions_off"] = '<span class=\'warning\'>访问权控制没�
 $_lang["access_permissions_resource_groups"] = '文档组';
 $_lang["access_permissions_resources_in_group"] = '<b>组里的文档：</b> ';
 $_lang["access_permissions_resources_tab"] = '你可以在这里查看建立了哪些文档组。你也可以建立/重命名/删除组。查看哪些文档归属于哪个文档组，(把鼠标放在文件的ID上就可以看见它的名字). 要在阻力添加/移除一个用户，就直接编辑文档。';
+$_lang["access_permissions_user_toggle"] = 'Toggle access permissions';
 $_lang["access_permissions_user_groups"] = '用户组';
 $_lang["access_permissions_user_message"] = 'Select which User Groups this user belongs to:';
 $_lang["access_permissions_users_in_group"] = '<b>在本组的用户:</b> ';
@@ -84,10 +85,11 @@ $_lang["captcha_title"] = '使用安全码：';
 $_lang["captcha_words_default"] = 'EVO,Access,Better,BitCode,Chunk,Cache,Desc,Design,Excell,Enjoy,URLs,TechView,Gerald,Griff,Humphrey,Holiday,Intel,Integration,Joystick,Join(),Oscope,Genetic,Light,Likeness,Marit,Maaike,Niche,Netherlands,Ordinance,Oscillo,Parser,Phusion,Query,Question,Regalia,Righteous,Snippet,Sentinel,Template,Thespian,Unity,Enterprise,Verily,Tattoo,Veri,Website,WideWeb,Yap,Yellow,Zebra,Zygote';
 $_lang["captcha_words_message"] = '你可以在这里输入一串机器码以让抗机器功能工作，用逗号（注意是英文状态下的逗号）隔开每一字符串，你最多可以输入225个字符。';
 $_lang["captcha_words_title"] = '机器机密码';
+$_lang["category"] = 'Category';
+$_lang["categories"] = 'Categories';
 $_lang["category_heading"] = 'Category';
 $_lang["category_manager"] = 'Category Manager';
 $_lang["category_management"] = 'Category management';
-$_lang["manage_categories"] = 'Manage Categories';
 $_lang["category_msg"] = 'View and edit all Elements grouped by category.';
 $_lang["cfg_base_path"] = 'MODX_BASE_PATH';
 $_lang["cfg_base_url"] = 'MODX_BASE_URL';
@@ -139,7 +141,7 @@ $_lang["configcheck_notok"] = '一个或多个配置未通过设置: ';
 $_lang["configcheck_ok"] = '检查通过，没有错误报告';
 $_lang["configcheck_php_gdzip"] = 'GD and/or Zip PHP extensions not found';
 $_lang["configcheck_php_gdzip_msg"] = 'Evolution CMS needs the GD and Zip extension enabled for PHP. While Evolution CMS will work without them, you will not be able to take full advantage of the built-in File Manager, Image Editor or Captcha for logins.';
-$_lang["configcheck_rb_base_dir"] = 'The currently set <a href="index.php?a=17&tab=5">File Browser path</a> seems incorrect.';
+$_lang["configcheck_rb_base_dir"] = 'The currently set <a href="index.php?a=17&tab=5">File base path</a> seems incorrect.';
 $_lang["configcheck_rb_base_dir_msg"] = 'This can happen for example by moving your installation to a different directory or server. Please check and update your Evolution CMS configuration.';
 $_lang["configcheck_register_globals"] = 'register_globals is set to ON in your php.ini configuration file';
 $_lang["configcheck_register_globals_msg"] = 'This configuration makes your site much more susceptible to Cross Site Scripting (XSS) attacks. You should speak to your host about what you can do to disable this setting.';
@@ -322,6 +324,7 @@ $_lang["file_unzip"] = '解压成功！';
 $_lang["file_unzip_fail"] = '解压失败！';
 $_lang["filemanager_path_message"] = 'IIS often does not populate the document_root setting properly, which is used by the file manager to determine what you can see. If you are having problems using the file manager, make sure this path points to the root of your Evolution CMS installation.';
 $_lang["filemanager_path_title"] = '文件管理路径:';
+$_lang["files"] = 'Files';
 $_lang["files_access_denied"] = '拒绝访问！';
 $_lang["files_data"] = '数据';
 $_lang["files_dir_listing"] = '目录列表:';
@@ -476,16 +479,17 @@ $_lang["login_message"] = '请输入你的管理帐号，注意，大小写';
 $_lang["logo_slogan"] = '简易而不简单 - \nEVO内容管理';
 $_lang["logout"] = '退出';
 $_lang["long_title"] = '长标题';
+$_lang["manage_categories"] = 'Manage Categories';
 $_lang["manage_depends"] = '依赖性管理';
 $_lang["manage_files"] = '文件管理';
-$_lang["manage_htmlsnippets"] = '块管理';
+$_lang["manage_htmlsnippets"] = 'Manage Chunks';
 $_lang["manage_metatags"] = '管理META标签和关键词';
 $_lang["manage_modules"] = '管理模块';
-$_lang["manage_plugins"] = '插件';
-$_lang["manage_snippets"] = '切片';
-$_lang["manage_templates"] = '模板';
-$_lang["manage_documents"] = 'Documents';
-$_lang["manage_permission"] = 'Permissions';
+$_lang["manage_plugins"] = 'Manage Plugins';
+$_lang["manage_snippets"] = 'Manage Snippets';
+$_lang["manage_templates"] = 'Manage Templates';
+$_lang["manage_documents"] = 'Manage Documents';
+$_lang["manage_permission"] = 'Manage Permissions';
 $_lang["manager"] = '管理员';
 $_lang["manager_lockout_message"] = 'You are currently logged into the Content Manager. If you would like to close your login session please click the "Logout" button. <p />To go to your startup or home page click the "Home" button.';
 $_lang["manager_permissions"] = '管理者许可';
@@ -1028,6 +1032,7 @@ $_lang["template_assignedtv_tab"] = 'Assigned Template Variables';
 $_lang["template_code"] = '模板代码 (html)';
 $_lang["template_desc"] = '描述';
 $_lang["template_edit_tab"] = 'Edit Template';
+$_lang["template_inuse"] = 'This template is in use. Please set the documents using the template to another template. Documents using this template:';
 $_lang["template_management_msg"] = '选择你需要编辑的模板。';
 $_lang["template_msg"] = '你可以在这里添加编辑/模板，当你站点的CACHE没清空的时候，你是不可以看见COOKIE里的文件采取了这个新的模板，当然你可以用预览功能来看到他。';
 $_lang["template_name"] = '模板名';
@@ -1280,10 +1285,6 @@ $_lang["setting_resource_tree_node_name_desc_add"] = 'Note: Since Evolution CMS 
 $_lang["resource_opt_alvisibled"] = 'Use current alias in alias path';
 $_lang["resource_opt_alvisibled_help"] = 'The alias of this Resource is inserted in Friendly URL alias path';
 $_lang['resource_opt_is_published'] = 'Published';
-$_lang["docid_incrmnt_method_title"] = 'Increment Resource ID method';
-$_lang["docid_incrmnt_method_0"] = 'DB auto increment';
-$_lang["docid_incrmnt_method_1"] = 'Minimum missed ID';
-$_lang["docid_incrmnt_method_2"] = 'Maximal ID + 1';
 
 $_lang["enable_cache_title"] = 'Document caching';
 $_lang["disable_chunk_cache_title"] = 'Disable chunk caching';
