@@ -48,7 +48,7 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
         $this->managerTheme->getCore()->invokeEvent(
             'OnManagerPreFrameLoader',
             [
-                'action' => $this->managerTheme->getActionId()
+                'action' => $this->managerTheme->getActionId(),
             ]
         );
 
@@ -94,7 +94,7 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
             'type5' => $this->managerTheme->getLexicon('lock_element_type_5'),
             'type6' => $this->managerTheme->getLexicon('lock_element_type_6'),
             'type7' => $this->managerTheme->getLexicon('lock_element_type_7'),
-            'type8' => $this->managerTheme->getLexicon('lock_element_type_8')
+            'type8' => $this->managerTheme->getLexicon('lock_element_type_8'),
         ];
 
         foreach ($unlockTranslations as $key => $value) {
@@ -115,15 +115,15 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
         $this->registerCss();
 
         $flag = $user['role'] == 1 || $this->managerTheme->getCore()->hasAnyPermissions([
-                'edit_template',
-                'edit_chunk',
-                'edit_snippet',
-                'edit_plugin'
-            ]);
+            'edit_template',
+            'edit_chunk',
+            'edit_snippet',
+            'edit_plugin',
+        ]);
 
         $this->managerTheme->getCore()->setConfig(
             'global_tabs',
-            (int)($this->managerTheme->getCore()->getConfig('global_tabs') && $flag)
+            (int) ($this->managerTheme->getCore()->getConfig('global_tabs') && $flag)
         );
 
         $this->makeMenu();
@@ -171,7 +171,6 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
                         $minifier->addFile($item);
                     }
                     file_put_contents($themeDir . 'css/styles.min.css', $minifier->minify());
-
                 }
             }
 
@@ -196,8 +195,8 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
             ->menuCategories()
             ->menuNewModule()
             ->menuRunModules()
-            ->menuWebUserManagment()
-            ->menuRoleManagment()
+            ->menuUserManagement()
+            ->menuRoleManagement()
             ->menuWebPermissions()
             ->menuRefreshSite()
             ->menuSearch()
@@ -230,7 +229,7 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
                     'parentClass' => 'dropdown',
                     'parentLinkClass' => 'dropdown-toggle',
                     'parentLinkAttr' => '',
-                    'parentLinkIn' => ''
+                    'parentLinkIn' => '',
                 ],
                 false
             );
@@ -249,7 +248,7 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
             '',
             0,
             10,
-            ''
+            '',
         ];
 
         return $this;
@@ -268,7 +267,7 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
             'main',
             0,
             10,
-            'active'
+            'active',
         ];
 
         return $this;
@@ -295,7 +294,7 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
             '',
             0,
             20,
-            ''
+            '',
         ];
 
         return $this;
@@ -318,7 +317,7 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
             '',
             0,
             30,
-            ''
+            '',
         ];
 
         return $this;
@@ -326,14 +325,11 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
 
     protected function menuUsers()
     {
-        if (!$this->managerTheme->getCore()
-            ->hasAnyPermissions([
-                    'edit_user',
-                    'edit_role',
-                    'manage_groups'
-                ]
-            )
-        ) {
+        $flag = $this->managerTheme->getCore()->hasAnyPermissions(
+            ['edit_user', 'edit_role', 'manage_groups']
+        );
+
+        if (!$flag) {
             return $this;
         }
 
@@ -348,7 +344,7 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
             '',
             0,
             40,
-            ''
+            '',
         ];
 
         return $this;
@@ -359,6 +355,7 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
         $flag = $this->managerTheme->getCore()->hasAnyPermissions(
             ['empty_cache', 'bk_manager', 'remove_locks', 'import_static', 'export_static']
         );
+
         if (!$flag) {
             return $this;
         }
@@ -374,7 +371,7 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
             '',
             0,
             50,
-            ''
+            '',
         ];
 
         return $this;
@@ -401,15 +398,15 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
         $this->sitemenu['element_templates'] = [
             'element_templates',
             'elements',
-            '<i class="' . $this->managerTheme->getStyle('icon_template') . '"></i>' . $this->managerTheme->getLexicon('manage_templates') . '<i class="' . $this->managerTheme->getStyle('icon_angle_right') . ' toggle"></i>',
+            '<i class="' . $this->managerTheme->getStyle('icon_template') . '"></i>' . $this->managerTheme->getLexicon('templates') . '<i class="' . $this->managerTheme->getStyle('icon_angle_right') . ' toggle"></i>',
             'index.php?a=76&tab=0',
-            $this->managerTheme->getLexicon('manage_templates'),
+            $this->managerTheme->getLexicon('templates'),
             '',
             'new_template,edit_template',
             'main',
             0,
             10,
-            'dropdown-toggle'
+            'dropdown-toggle',
         ];
     }
 
@@ -430,7 +427,7 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
             'main',
             0,
             20,
-            'dropdown-toggle'
+            'dropdown-toggle',
         ];
     }
 
@@ -443,15 +440,15 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
         $this->sitemenu['element_htmlsnippets'] = [
             'element_htmlsnippets',
             'elements',
-            '<i class="' . $this->managerTheme->getStyle('icon_chunk') . '"></i>' . $this->managerTheme->getLexicon('manage_htmlsnippets') . '<i class="' . $this->managerTheme->getStyle('icon_angle_right') . ' toggle"></i>',
+            '<i class="' . $this->managerTheme->getStyle('icon_chunk') . '"></i>' . $this->managerTheme->getLexicon('htmlsnippets') . '<i class="' . $this->managerTheme->getStyle('icon_angle_right') . ' toggle"></i>',
             'index.php?a=76&tab=2',
-            $this->managerTheme->getLexicon('manage_htmlsnippets'),
+            $this->managerTheme->getLexicon('htmlsnippets'),
             '',
             'new_chunk,edit_chunk',
             'main',
             0,
             30,
-            'dropdown-toggle'
+            'dropdown-toggle',
         ];
     }
 
@@ -464,15 +461,15 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
         $this->sitemenu['element_snippets'] = [
             'element_snippets',
             'elements',
-            '<i class="' . $this->managerTheme->getStyle('icon_code') . '"></i>' . $this->managerTheme->getLexicon('manage_snippets') . '<i class="' . $this->managerTheme->getStyle('icon_angle_right') . ' toggle"></i>',
+            '<i class="' . $this->managerTheme->getStyle('icon_code') . '"></i>' . $this->managerTheme->getLexicon('snippets') . '<i class="' . $this->managerTheme->getStyle('icon_angle_right') . ' toggle"></i>',
             'index.php?a=76&tab=3',
-            $this->managerTheme->getLexicon('manage_snippets'),
+            $this->managerTheme->getLexicon('snippets'),
             '',
             'new_snippet,edit_snippet',
             'main',
             0,
             40,
-            'dropdown-toggle'
+            'dropdown-toggle',
         ];
     }
 
@@ -485,15 +482,15 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
         $this->sitemenu['element_plugins'] = [
             'element_plugins',
             'elements',
-            '<i class="' . $this->managerTheme->getStyle('icon_plugin') . '"></i>' . $this->managerTheme->getLexicon('manage_plugins') . '<i class="' . $this->managerTheme->getStyle('icon_angle_right') . ' toggle"></i>',
+            '<i class="' . $this->managerTheme->getStyle('icon_plugin') . '"></i>' . $this->managerTheme->getLexicon('plugins') . '<i class="' . $this->managerTheme->getStyle('icon_angle_right') . ' toggle"></i>',
             'index.php?a=76&tab=4',
-            $this->managerTheme->getLexicon('manage_plugins'),
+            $this->managerTheme->getLexicon('plugins'),
             '',
             'new_plugin,edit_plugin',
             'main',
             0,
             50,
-            'dropdown-toggle'
+            'dropdown-toggle',
         ];
     }
 
@@ -506,19 +503,19 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
         $this->sitemenu['element_modules'] = [
             'element_modules',
             'elements',
-            '<i class="' . $this->managerTheme->getStyle('icon_module') . '"></i>' . $this->managerTheme->getLexicon('manage_modules') . '<i class="' . $this->managerTheme->getStyle('icon_angle_right') . ' toggle"></i>',
+            '<i class="' . $this->managerTheme->getStyle('icon_module') . '"></i>' . $this->managerTheme->getLexicon('modules') . '<i class="' . $this->managerTheme->getStyle('icon_angle_right') . ' toggle"></i>',
             'index.php?a=76&tab=5',
-            $this->managerTheme->getLexicon('manage_modules'),
+            $this->managerTheme->getLexicon('modules'),
             '',
             'new_module,edit_module',
             'main',
             0,
             60,
-            'dropdown-toggle'
+            'dropdown-toggle',
         ];
     }
 
-    function menuFiles()
+    protected function menuFiles()
     {
         if (!$this->managerTheme->getCore()->hasPermission('file_manager')) {
             return $this;
@@ -527,15 +524,15 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
         $this->sitemenu['manage_files'] = [
             'manage_files',
             'elements',
-            '<i class="' . $this->managerTheme->getStyle('icon_folder_open') . '"></i>' . $this->managerTheme->getLexicon('manage_files'),
+            '<i class="' . $this->managerTheme->getStyle('icon_folder_open') . '"></i>' . $this->managerTheme->getLexicon('files'),
             'index.php?a=31',
-            $this->managerTheme->getLexicon('manage_files'),
+            $this->managerTheme->getLexicon('files'),
             '',
             'file_manager',
             'main',
             0,
             80,
-            ''
+            '',
         ];
 
         return $this;
@@ -550,15 +547,15 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
         $this->sitemenu['manage_categories'] = [
             'manage_categories',
             'elements',
-            '<i class="' . $this->managerTheme->getStyle('icon_category') . '"></i>' . $this->managerTheme->getLexicon('manage_categories'),
+            '<i class="' . $this->managerTheme->getStyle('icon_category') . '"></i>' . $this->managerTheme->getLexicon('categories'),
             'index.php?a=120',
-            $this->managerTheme->getLexicon('manage_categories'),
+            $this->managerTheme->getLexicon('categories'),
             '',
             'category_manager',
             'main',
             0,
             70,
-            ''
+            '',
         ];
 
         return $this;
@@ -580,19 +577,26 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
     protected function menuRunModules()
     {
         if ($this->managerTheme->getCore()->hasPermission('exec_module')) {
+            $items = [];
+
+            // 1. modules from DB
             if ($_SESSION['mgrRole'] != 1 && $this->managerTheme->getCore()->getConfig('use_udperms') === true) {
                 $modules = SiteModule::select('site_modules.id', 'site_modules.name', 'site_modules.icon', 'member_groups.member')
                     ->withoutProtected()
                     ->lockedView()
                     ->where('site_modules.disabled', 0)
-                    ->orderBy('site_modules.name')->get()->toArray();
-
+                    ->orderBy('site_modules.name')
+                    ->get()
+                    ->toArray();
             } else {
-                $modules = SiteModule::where('disabled', '!=', 1)->orderBy('name')->get()->toArray();
+                $modules = SiteModule::where('disabled', '!=', 1)
+                    ->orderBy('name')
+                    ->get()
+                    ->toArray();
             }
             if (count($modules) > 0) {
                 foreach ($modules as $row) {
-                    $this->sitemenu['module' . $row['id']] = [
+                    $items['module' . $row['id']] = [
                         'module' . $row['id'],
                         'modules',
                         ($row['icon'] != '' ? '<i class="' . e($row['icon']) . '"></i>' : '<i class="' . $this->managerTheme->getStyle('icon_module') . '"></i>') . e($row['name']),
@@ -603,16 +607,18 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
                         'main',
                         0,
                         1,
-                        ''
+                        '',
                     ];
                 }
             }
+
+            // 2. modules from files
             foreach ($this->managerTheme->getCore()->modulesFromFile as $module) {
                 if (!empty($module['properties']['hidden'])) {
                     continue;
                 }
 
-                $this->sitemenu['module' . $module['id']] = [
+                $items['module' . $module['id']] = [
                     'module' . $module['id'],
                     'modules',
                     ($module['icon'] != '' ? '<i class="' . $module['icon'] . '"></i>' : '<i class="' . $this->managerTheme->getStyle('icon_module') . '"></i>') . $module['name'],
@@ -623,43 +629,54 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
                     'main',
                     0,
                     1,
-                    ''
+                    '',
                 ];
             }
+
+            usort($items, fn($a, $b) => $a[4] <=> $b[4]);
+
+            foreach ($items as $index => &$item) {
+                $item[0] = 'module' . $index;
+            }
+
+            $this->sitemenu = array_merge($this->sitemenu, $items);
         }
 
         return $this;
     }
 
-
-    protected function menuWebUserManagment()
+    protected function menuUserManagement()
     {
-        if ($this->managerTheme->getCore()->hasPermission('edit_user')) {
-            $this->sitemenu['web_user_management_title'] = [
-                'web_user_management_title',
-                'users',
-                '<i class="' . $this->managerTheme->getStyle('icon_web_user') . '"></i>' . $this->managerTheme->getLexicon('web_user_management_title') . '<i class="' . $this->managerTheme->getStyle('icon_angle_right') . ' toggle"></i>',
-                'index.php?a=99',
-                $this->managerTheme->getLexicon('web_user_management_title'),
-                '',
-                'edit_user',
-                'main',
-                0,
-                20,
-                'dropdown-toggle'
-            ];
+        $flag = $this->managerTheme->getCore()->hasPermission('edit_user');
+
+        if (!$flag) {
+            return $this;
         }
+
+        $this->sitemenu['web_user_management_title'] = [
+            'web_user_management_title',
+            'users',
+            '<i class="' . $this->managerTheme->getStyle('icon_web_user') . '"></i>' . $this->managerTheme->getLexicon('web_user_management_title') . '<i class="' . $this->managerTheme->getStyle('icon_angle_right') . ' toggle"></i>',
+            'index.php?a=99',
+            $this->managerTheme->getLexicon('web_user_management_title'),
+            '',
+            'edit_user',
+            'main',
+            0,
+            20,
+            'dropdown-toggle',
+        ];
 
         return $this;
     }
 
-    protected function menuRoleManagment()
+    protected function menuRoleManagement()
     {
         if ($this->managerTheme->getCore()->hasPermission('edit_role')) {
             $this->sitemenu['role_management_title'] = [
                 'role_management_title',
                 'users',
-                '<i class="' . $this->managerTheme->getStyle('icon_role') . '"></i>' . $this->managerTheme->getLexicon('role_management_title'),
+                '<i class="' . $this->managerTheme->getStyle('icon_role') . '"></i>' . $this->managerTheme->getLexicon('role_management_title') . '<i class="' . $this->managerTheme->getStyle('icon_angle_right') . ' toggle"></i>',
                 'index.php?a=86',
                 $this->managerTheme->getLexicon('role_management_title'),
                 '',
@@ -667,13 +684,12 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
                 'main',
                 0,
                 30,
-                ''
+                'dropdown-toggle',
             ];
         }
 
         return $this;
     }
-
 
     protected function menuWebPermissions()
     {
@@ -692,7 +708,7 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
             'main',
             0,
             50,
-            ''
+            '',
         ];
 
         return $this;
@@ -724,10 +740,10 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
                     // onclick
                     $this->managerTheme->getLexicon('refresh_site'),
                     // title
-                    '<i class="' . $this->managerTheme->getStyle('icon_recycle') . '"></i>'
+                    '<i class="' . $this->managerTheme->getStyle('icon_recycle') . '"></i>',
                     // innerHTML
-                ]
-            ]
+                ],
+            ],
         ];
 
         return $this;
@@ -746,7 +762,7 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
             'main',
             1,
             9,
-            ''
+            '',
         ];
 
         return $this;
@@ -769,7 +785,7 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
             'main',
             0,
             10,
-            ''
+            '',
         ];
 
         return $this;
@@ -792,7 +808,7 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
             '',
             0,
             20,
-            ''
+            '',
         ];
 
         return $this;
@@ -811,7 +827,7 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
             'main',
             0,
             30,
-            ''
+            '',
         ];
 
         return $this;
