@@ -164,8 +164,6 @@ class ManagerTheme implements ManagerThemeInterface
         94,
         /** Update Tree for Closure Table */
         95 => Controllers\UpdateTree::class,
-        /** Help */
-        9 => Controllers\Help::class,
         /** Template Variables - Based on Apodigm's Docvars */
         300 => Controllers\Tmplvar::class,
         301 => Controllers\Tmplvar::class,
@@ -770,7 +768,7 @@ class ManagerTheme implements ManagerThemeInterface
 
     public function isLoadDatePicker(): bool
     {
-        $actions = [85, 27, 4, 72, 13, 11, 12, 87, 88];
+        $actions = [85, 27, 4, 72, 13, 87, 88];
         return \in_array($this->getCore()->getManagerApi()->action, $actions, true);
     }
 
@@ -822,7 +820,6 @@ class ManagerTheme implements ManagerThemeInterface
                 $this->getThemeDir() . $minCssName,
                 $css
             );
-
         }
         if (file_exists($this->getThemeDir() . $minCssName)) {
             $css = $this->getThemeUrl() . $minCssName;
