@@ -1,14 +1,14 @@
 <?php
-if( ! defined('IN_MANAGER_MODE') || IN_MANAGER_MODE !== true) {
+if (!defined('IN_MANAGER_MODE') || IN_MANAGER_MODE !== true) {
     die("<b>INCLUDE_ORDERING_ERROR</b><br /><br />Please use the EVO Content Manager instead of accessing this file directly.");
 }
-if(!$modx->hasPermission('delete_snippet')) {
-	$modx->webAlertAndQuit($_lang["error_no_privileges"]);
+if (!$modx->hasPermission('delete_snippet')) {
+    $modx->webAlertAndQuit($_lang["error_no_privileges"]);
 }
 
-$id = isset($_GET['id'])? (int)$_GET['id'] : 0;
-if($id==0) {
-	$modx->webAlertAndQuit($_lang["error_no_id"]);
+$id = isset($_GET['id']) ? (int) $_GET['id'] : 0;
+if ($id == 0) {
+    $modx->webAlertAndQuit($_lang["error_no_id"]);
 }
 
 // Set the item name for logger
@@ -16,23 +16,27 @@ $name = EvolutionCMS\Models\SiteSnippet::findOrFail($id)->name;
 $_SESSION['itemname'] = $name;
 
 // invoke OnBeforeSnipFormDelete event
-$modx->invokeEvent("OnBeforeSnipFormDelete",
-	array(
-		"id"	=> $id
-	));
+$modx->invokeEvent(
+    "OnBeforeSnipFormDelete",
+    array(
+        "id" => $id,
+    )
+);
 
 // delete the snippet.
 EvolutionCMS\Models\SiteSnippet::destroy($id);
 
 // invoke OnSnipFormDelete event
-$modx->invokeEvent("OnSnipFormDelete",
-	array(
-		"id"	=> $id
-	));
+$modx->invokeEvent(
+    "OnSnipFormDelete",
+    array(
+        "id" => $id,
+    )
+);
 
 // empty cache
 $modx->clearCache('full');
 
 // finished emptying cache - redirect
-$header="Location: index.php?a=76&r=2&tab=3";
+$header = "Location: index.php?a=76&r=2&tab=3";
 header($header);
