@@ -1,12 +1,12 @@
 <?php
-if( ! defined('IN_MANAGER_MODE') || IN_MANAGER_MODE !== true) {
+if (!defined('IN_MANAGER_MODE') || IN_MANAGER_MODE !== true) {
     die("<b>INCLUDE_ORDERING_ERROR</b><br /><br />Please use the EVO Content Manager instead of accessing this file directly.");
 }
 if (!$modx->hasPermission('new_template')) {
     $modx->webAlertAndQuit($_lang["error_no_privileges"]);
 }
 
-$id = isset($_GET['id']) ? (int)$_GET['id'] : 0;
+$id = isset($_GET['id']) ? (int) $_GET['id'] : 0;
 if ($id == 0) {
     $modx->webAlertAndQuit($_lang["error_no_id"]);
 }
@@ -14,8 +14,11 @@ if ($id == 0) {
 // count duplicates
 $name = EvolutionCMS\Models\SiteTemplate::select('templatename')->findOrFail($id)->templatename;
 $count = EvolutionCMS\Models\SiteTemplate::where('templatename', 'LIKE', "{$name} {$_lang['duplicated_el_suffix']}%'")->count();
-if ($count >= 1) $count = ' ' . ($count + 1);
-else $count = '';
+if ($count >= 1) {
+    $count = ' ' . ($count + 1);
+} else {
+    $count = '';
+}
 
 // duplicate template
 $template = EvolutionCMS\Models\SiteTemplate::select("templatename", "description", "content", "category")
@@ -38,5 +41,5 @@ $name = EvolutionCMS\Models\SiteTemplate::select('templatename')->findOrFail($ne
 $_SESSION['itemname'] = $name;
 
 // finish duplicating - redirect to new template
-$header="Location: index.php?r=2&a=16&id=$newid";
+$header = "Location: index.php?r=2&a=16&id=$newid";
 header($header);
