@@ -1,28 +1,61 @@
 <?php
-if( ! defined('IN_MANAGER_MODE') || IN_MANAGER_MODE !== true) {
+
+use EvolutionCMS\Models\SiteTemplate;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
+
+if (!defined('IN_MANAGER_MODE') || IN_MANAGER_MODE !== true) {
     die("<b>INCLUDE_ORDERING_ERROR</b><br /><br />Please use the EVO Content Manager instead of accessing this file directly.");
 }
 if (!$modx->hasPermission('save_template')) {
     $modx->webAlertAndQuit($_lang["error_no_privileges"]);
 }
 
-$id = (int)$_POST['id'];
+if (isset($_GET['selectable'])) {
+    $selectable = (int) $_GET['selectable'];
+    $id = (int) ($_REQUEST['id'] ?? 0);
+
+    try {
+        /** @var SiteTemplate $template */
+        $template = SiteTemplate::query()->findOrFail($id);
+
+        $modx->invokeEvent("OnBeforeTempFormSave", [
+            'mode' => 'upd',
+            'id' => $id,
+        ]);
+
+        $_SESSION['itemname'] = $template->templatename;
+
+        $template->update(['selectable' => $selectable]);
+        $modx->invokeEvent('OnTempFormSave', [
+            'mode' => 'upd',
+            'id' => $id,
+        ]);
+    } catch (ModelNotFoundException $e) {
+        $modx->webAlertAndQuit(__('global.error_no_id'));
+    }
+
+    $header = 'Location: index.php?a=76&tab=0&r=2';
+    header($header);
+    exit;
+}
+
+$id = (int) $_POST['id'];
 $template = $_POST['post'];
 $templatename = trim($_POST['templatename']);
 $templatealias = trim($_POST['templatealias']);
 $description = $_POST['description'];
 $locked = isset($_POST['locked']) && $_POST['locked'] == 'on' ? 1 : 0;
-$selectable = $id == $modx->config['default_template'] ? 1 :    // Force selectable
-    (isset($_POST['selectable']) && $_POST['selectable'] == 'on' ? 1 : 0);
+$selectable = $id == $modx->config['default_template'] ? 1 : // Force selectable
+(isset($_POST['selectable']) && $_POST['selectable'] == 'on' ? 1 : 0);
 $currentdate = time() + $modx->config['server_offset_time'];
 
 //Kyle Jaebker - added category support
 if (empty($_POST['newcategory']) && $_POST['categoryid'] > 0) {
-    $categoryid = (int)$_POST['categoryid'];
+    $categoryid = (int) $_POST['categoryid'];
 } elseif (empty($_POST['newcategory']) && $_POST['categoryid'] <= 0) {
     $categoryid = 0;
 } else {
-    include_once(MODX_MANAGER_PATH . 'includes/categories.inc.php');
+    include_once MODX_MANAGER_PATH . 'includes/categories.inc.php';
     $categoryid = checkCategory($_POST['newcategory']);
     if (!$categoryid) {
         $categoryid = newCategory($_POST['newcategory']);
@@ -57,11 +90,10 @@ function createBladeFile($templatealias)
 
 switch ($_POST['mode']) {
     case '19':
-
         // invoke OnBeforeTempFormSave event
         $modx->invokeEvent("OnBeforeTempFormSave", array(
             "mode" => "new",
-            "id" => $id
+            "id" => $id,
         ));
 
         // disallow duplicate names for new templates
@@ -71,8 +103,10 @@ switch ($_POST['mode']) {
             $modx->webAlertAndQuit(sprintf($_lang['duplicate_name_found_general'], $_lang['template'], $templatename), "index.php?a=19");
         }
 
-        if($templatealias == '')
+        if ($templatealias == '') {
             $templatealias = $templatename;
+        }
+
         $templatealias = strtolower($modx->stripAlias(trim($templatealias)));
 
         $count = \EvolutionCMS\Models\SiteTemplate::where('templatealias', $templatealias)->count();
@@ -91,13 +125,13 @@ switch ($_POST['mode']) {
             'selectable' => $selectable,
             'category' => $categoryid,
             'createdon' => $currentdate,
-            'editedon' => $currentdate
+            'editedon' => $currentdate,
         ));
 
         // invoke OnTempFormSave event
         $modx->invokeEvent("OnTempFormSave", array(
             "mode" => "new",
-            "id" => $newid
+            "id" => $newid,
         ));
         // Set new assigned Tvs
         saveTemplateAccess($newid);
@@ -124,11 +158,10 @@ switch ($_POST['mode']) {
 
         break;
     case '16':
-
         // invoke OnBeforeTempFormSave event
         $modx->invokeEvent("OnBeforeTempFormSave", array(
             "mode" => "upd",
-            "id" => $id
+            "id" => $id,
         ));
 
         // disallow duplicate names for templates
@@ -138,8 +171,10 @@ switch ($_POST['mode']) {
             $modx->webAlertAndQuit(sprintf($_lang['duplicate_name_found_general'], $_lang['template'], $templatename), "index.php?a=16&id={$id}");
         }
 
-        if($templatealias == '')
+        if ($templatealias == '') {
             $templatealias = $templatename;
+        }
+
         $templatealias = strtolower($modx->stripAlias(trim($templatealias)));
 
         $count = \EvolutionCMS\Models\SiteTemplate::where('templatealias', $templatealias)->where('id', '!=', $id)->count();
@@ -157,7 +192,7 @@ switch ($_POST['mode']) {
             'locked' => $locked,
             'selectable' => $selectable,
             'category' => $categoryid,
-            'editedon' => $currentdate
+            'editedon' => $currentdate,
         ));
         // Set new assigned Tvs
         saveTemplateAccess($id);
@@ -169,7 +204,7 @@ switch ($_POST['mode']) {
         // invoke OnTempFormSave event
         $modx->invokeEvent("OnTempFormSave", array(
             "mode" => "upd",
-            "id" => $id
+            "id" => $id,
         ));
 
         // Set the item name for logger
@@ -188,8 +223,6 @@ switch ($_POST['mode']) {
             $header = "Location: index.php?a=76&r=2";
             header($header);
         }
-
-
         break;
     default:
         $modx->webAlertAndQuit("No operation set in request.");
