@@ -1,8 +1,11 @@
 <div class="tab-page {{ $tabPageName }}" id="{{ $tabIndexPageName }}">
     <h2 class="tab">
-        <a href="?a=76&tab={{ $index }}"><i class="{{ $_style['icon_chunk'] }}"></i>{{ ManagerTheme::getLexicon('manage_htmlsnippets') }}</a>
+        <a href="?a=76&tab={{ $index }}"><i
+                class="{{ $_style['icon_chunk'] }}"></i>{{ ManagerTheme::getLexicon('htmlsnippets') }}</a>
     </h2>
-    <script>tpResources.addTabPage(document.getElementById('{{ $tabIndexPageName }}'));</script>
+    <script>
+        tpResources.addTabPage(document.getElementById('{{ $tabIndexPageName }}'));
+    </script>
 
     <div id="{{ $tabIndexPageName }}-info" class="msg-container" style="display:none">
         <div class="element-edit-message-tab">{!! ManagerTheme::getLexicon('htmlsnippet_management_msg') !!}</div>
@@ -12,9 +15,11 @@
     <div id="_actions">
         <form class="btn-group form-group form-inline">
             <div class="input-group input-group-sm">
-                <input class="form-control filterElements-form" type="text" id="{{ $tabIndexPageName }}_search" size="30" placeholder="{{ ManagerTheme::getLexicon('element_filter_msg') }}" />
+                <input class="form-control filterElements-form" type="text" id="{{ $tabIndexPageName }}_search"
+                    size="30" placeholder="{{ ManagerTheme::getLexicon('element_filter_msg') }}" />
                 <div class="input-group-btn">
-                    <a class="btn btn-success" target="main" href="{{ (new EvolutionCMS\Models\SiteHtmlsnippet)->makeUrl('actions.new') }}">
+                    <a class="btn btn-success" target="main"
+                        href="{{ (new EvolutionCMS\Models\SiteHtmlsnippet())->makeUrl('actions.new') }}">
                         <i class="{{ $_style['icon_add'] }}"></i>
                         <span>{{ ManagerTheme::getLexicon('new_htmlsnippet') }}</span>
                     </a>
@@ -22,7 +27,8 @@
                         <i class="{{ $_style['icon_question_circle'] }}"></i>
                         <span>{{ ManagerTheme::getLexicon('help') }}</span>
                     </a>
-                    <a class="btn btn-secondary switchform-btn" href="javascript:;" data-target="switchForm_{{ $tabIndexPageName }}">
+                    <a class="btn btn-secondary switchform-btn" href="javascript:;"
+                        data-target="switchForm_{{ $tabIndexPageName }}">
                         <i class="{{ $_style['icon_bars'] }}"></i>
                         <span>{{ ManagerTheme::getLexicon('btn_view_options') }}</span>
                     </a>
@@ -36,22 +42,34 @@
     <div class="clearfix"></div>
     <div class="panel-group no-transition">
         <div id="{{ $tabIndexPageName }}_content" class="resourceTable panel panel-default">
-            @if(isset($outCategory) && $outCategory->count() > 0)
-                @component('manager::partials.panelCollapse', ['name' => $tabIndexPageName . '_content', 'id' => 0, 'title' => ManagerTheme::getLexicon('no_category')])
+            @if (isset($outCategory) && $outCategory->count() > 0)
+                @component('manager::partials.panelCollapse', [
+                    'name' => $tabIndexPageName . '_content',
+                    'id' => 0,
+                    'title' => ManagerTheme::getLexicon('no_category'),
+                ])
                     <ul class="elements">
-                        @foreach($outCategory as $item)
-                            @include('manager::page.resources.elements.chunk', compact('item', 'tabIndexPageName'))
+                        @foreach ($outCategory as $item)
+                            @include(
+                                'manager::page.resources.elements.chunk',
+                                compact('item', 'tabIndexPageName'))
                         @endforeach
                     </ul>
                 @endcomponent
             @endif
 
-            @if(isset($categories))
-                @foreach($categories as $cat)
-                    @component('manager::partials.panelCollapse', ['name' => $tabIndexPageName . '_content', 'id' => $cat->id, 'title' => $cat->name])
+            @if (isset($categories))
+                @foreach ($categories as $cat)
+                    @component('manager::partials.panelCollapse', [
+                        'name' => $tabIndexPageName . '_content',
+                        'id' => $cat->id,
+                        'title' => $cat->name,
+                    ])
                         <ul class="elements">
-                            @foreach($cat->chunks as $item)
-                                @include('manager::page.resources.elements.chunk', compact('item', 'tabIndexPageName'))
+                            @foreach ($cat->chunks as $item)
+                                @include(
+                                    'manager::page.resources.elements.chunk',
+                                    compact('item', 'tabIndexPageName'))
                             @endforeach
                         </ul>
                     @endcomponent
@@ -65,6 +83,6 @@
 @push('scripts.bot')
     <script>
         initQuicksearch('{{ $tabIndexPageName }}_search', '{{ $tabIndexPageName }}_content');
-      initViews('ch', '{{ $tabIndexPageName }}', '{{ $tabIndexPageName }}_content');
+        initViews('ch', '{{ $tabIndexPageName }}', '{{ $tabIndexPageName }}_content');
     </script>
 @endpush
